Calculate Level Paths

Calculates level paths using the stream-leveling approach of NHD and NHDPlus. In addition to a levelpath identifier, a topological sort and levelpath outlet identifier is provided in output. If arbolate sum is provided in the weight column, this will match the behavior of NHDPlus. Any numeric value can be included in this column and the largest value will be followed when no nameID is available.

calculate_levelpaths(flowline)

Arguments

flowline
data.frame with ID, toID, nameID, and weight columns.

Value

data.frame with ID, outletID, topo_sort, and levelpath collumns. See details for more info.

Details

levelpath provides an identifier for the collection of flowlines that make up the single mainstem flowpath of a total upstream aggregate catchment.

outletID is the catchment ID (COMID in the case of NHDPlus) for the catchment at the outlet of the levelpath the catchment is part of.

topo_sort is similar to Hydroseq in NHDPlus in that large topo_sort values are upstream of small topo_sort values. Note that there are many valid topological sort orders of a directed graph. The sort order output by this function is generated using `igraph::topo_sort`.

Examples

source(system.file("extdata", "walker_data.R", package = "nhdplusTools"))

test_flowline <- prepare_nhdplus(walker_flowline, 0, 0, FALSE)

#> Warning: removing geometry

#> Warning: Removed 0 flowlines that don't apply.
#>  Includes: Coastlines, non-dendritic paths, 
#> and networks with drainage area less than 0 sqkm, and drainage basins smaller than FALSE

test_flowline <- data.frame(
  ID = test_flowline$COMID,
  toID = test_flowline$toCOMID,
  nameID = walker_flowline$GNIS_ID,
  weight = walker_flowline$ArbolateSu,
  stringsAsFactors = FALSE)

calculate_levelpaths(test_flowline)

#>         ID outletID topo_sort levelpath
#> 1  5329435  5329303        62         1
#> 2  5329347  5329347        61        62
#> 3  5329291  5329291        60        61
#> 4  5329359  5329363        59        33
#> 5  5329333  5329333        58        60
#> 6  5329371  5329371        57        59
#> 7  5329377  5329377        56        58
#> 8  5329399  5329375        55        36
#> 9  5329405  5329405        54        57
#> 10 5329427  5329427        53        56
#> 11 5329413  5329413        52        55
#> 12 5329419  5329419        51        54
#> 13 5329407  5329407        50        53
#> 14 5329387  5329387        49        52
#> 15 5329415  5329845        48        43
#> 16 5329355  5329355        47        51
#> 17 5329337  5329337        46        50
#> 18 5329345  5329335        45        41
#> 19 5329341  5329341        44        49
#> 20 5329815  5329313        43        19
#> 21 5329319  5329319        42        48
#> 22 5329309  5329309        41        47
#> 23 5329307  5329307        40        46
#> 24 5329297  5329299        39        39
#> 25 5329295  5329295        38        45
#> 26 5329871  5329391        37        28
#> 27 5329379  5329375        36        36
#> 28 5329845  5329845        35        43
#> 29 5329819  5329363        34        33
#> 30 5329335  5329335        33        41
#> 31 5329841  5329313        32        19
#> 32 5329299  5329299        31        39
#> 33 5329383  5329391        30        28
#> 34 5329375  5329375        29        36
#> 35 5329363  5329363        28        33
#> 36 5329321  5329313        27        19
#> 37 5329849  5329391        26        28
#> 38 5329325  5329313        25        19
#> 39 5329393  5329391        24        28
#> 40 5329327  5329313        23        19
#> 41 5329391  5329391        22        28
#> 42 5329323  5329313        21        19
#> 43 5329389  5329303        20         1
#> 44 5329817  5329313        19        19
#> 45 5329397  5329303        18         1
#> 46 5329311  5329313        17        19
#> 47 5329395  5329303        16         1
#> 48 5329313  5329313        15        19
#> 49 5329821  5329303        14         1
#> 50 5329385  5329303        13         1
#> 51 5329847  5329303        12         1
#> 52 5329843  5329303        11         1
#> 53 5329373  5329303        10         1
#> 54 5329365  5329303         9         1
#> 55 5329357  5329303         8         1
#> 56 5329343  5329303         7         1
#> 57 5329339  5329303         6         1
#> 58 5329315  5329303         5         1
#> 59 5329317  5329303         4         1
#> 60 5329305  5329303         3         1
#> 61 5329293  5329303         2         1
#> 62 5329303  5329303         1         1