Discover Characteristics Metadata

Provides access to metadata for characteristics that are returned by `get_nldi_characteristics()`.

discover_nldi_characteristics(type = "all", tier = "prod")

Arguments

type
character "all", "local", "total", or "divergence_routed".

tier
character optional "prod" or "test"

Examples

chars <- discover_nldi_characteristics()
names(chars)

#> [1] "local"             "total"             "divergence_routed"

head(chars$local, 10)

#>    characteristic_id
#> 1      CAT_NPDES_MAJ
#> 2       CAT_ECOL3_24
#> 3            CAT_BFI
#> 4            CAT_PET
#> 5        CAT_CONTACT
#> 6            CAT_EWT
#> 7           CAT_IEOF
#> 8          CAT_RECHG
#> 9          CAT_SATOF
#> 10           CAT_TWI
#>                                                                                                                                                                                                                                          characteristic_description
#> 1                                                                                                                                                                                                                                       Number of major NPDES sites
#> 2                                                                                                                                                                                                                             level 3 ecoregion: Chihuahuan Desert.
#> 3    Base Flow Index (BFI), The BFI is a ratio of base flow to total streamflow, expressed as a percentage and ranging from 0 to 100. Base flow is the sustained, slowly varying component of streamflow, usually attributed to ground-water discharge to a stream.
#> 4                                                                                                                                                                        Mean-annual potential evapotranspiration (PET), estimated using the Hamon (1961) equation.
#> 5                                               Subsurface flow contact time index. The subsurface contact time index estimates the number of days that infiltrated water resides in the saturated subsurface zone of the basin before discharging into the stream.
#> 6                                                                                                                                                                                                 Average depth to water table relatice to the land surface(meters)
#> 7                                                                                                                                                                                                                   Percentage of Horton overland flow as a percent
#> 8                                                                                                                                                                                                 Mean annual natural ground-water recharge in millimeters per year
#> 9                                                                                                                                                                                                                    Percentage of Dunne overland flow as a percent
#> 10 Topographic wetness index, ln(a/S); where ln is the natural log, a is the upslope area per unit contour length and S is the slope at that point.  See http://ks.water.usgs.gov/Kansas/pubs/reports/wrir.99-4242.html and Wolock and McCabe, 1995 for more detail
#>      units
#> 1    count
#> 2  unknown
#> 3  percent
#> 4  mm/year
#> 5     days
#> 6   meters
#> 7  percent
#> 8    mm/yr
#> 9  percent
#> 10   ln(m)
#>                                                                                                                                                                                                                                                       dataset_label
#> 1                                                                                                                                                                                                                                       Number of major NPDES sites
#> 2                                                                                                                                                                                                        percent of catchment containing dominant level 3 ecoregion
#> 3    Base Flow Index (BFI), The BFI is a ratio of base flow to total streamflow, expressed as a percentage and ranging from 0 to 100. Base flow is the sustained, slowly varying component of streamflow, usually attributed to ground-water discharge to a stream.
#> 4                                                                                                                                                                        Mean-annual potential evapotranspiration (PET), estimated using the Hamon (1961) equation.
#> 5                                               Subsurface flow contact time index. The subsurface contact time index estimates the number of days that infiltrated water resides in the saturated subsurface zone of the basin before discharging into the stream.
#> 6                                                                                                                                                                                                 Average depth to water table relatice to the land surface(meters)
#> 7                                                                                                                                                                                                                   Percentage of Horton overland flow as a percent
#> 8                                                                                                                                   Runoff (mm/yr) delivered to the stream network for every month for the years 1945 - 2013 divided by 12 to derive annual runoff.
#> 9                                                                                                                                                                                                                    Percentage of Dunne overland flow as a percent
#> 10 Topographic wetness index, ln(a/S); where ln is the natural log, a is the upslope area per unit contour length and S is the slope at that point.  See http://ks.water.usgs.gov/Kansas/pubs/reports/wrir.99-4242.html and Wolock and McCabe, 1995 for more detail
#>                                                          dataset_url
#> 1  https://www.sciencebase.gov/catalog/item/57c9d89ce4b0f2f0cec192da
#> 2  https://www.sciencebase.gov/catalog/item/5785595ce4b0e02680bf2fd8
#> 3  https://www.sciencebase.gov/catalog/item/5669a8e3e4b08895842a1d4f
#> 4  https://www.sciencebase.gov/catalog/item/56f96ed1e4b0a6037df06a2d
#> 5  https://www.sciencebase.gov/catalog/item/56f96fc5e4b0a6037df06b12
#> 6  https://www.sciencebase.gov/catalog/item/56f97456e4b0a6037df06b50
#> 7  https://www.sciencebase.gov/catalog/item/56f974e2e4b0a6037df06b55
#> 8  https://www.sciencebase.gov/catalog/item/56f97577e4b0a6037df06b5a
#> 9  https://www.sciencebase.gov/catalog/item/56f97acbe4b0a6037df06b6a
#> 10 https://www.sciencebase.gov/catalog/item/56f97be4e4b0a6037df06b70
#>    theme_label
#> 1    Hydro_Mod
#> 2      Regions
#> 3   Hydrologic
#> 4      Climate
#> 5   Hydrologic
#> 6   Hydrologic
#> 7   Hydrologic
#> 8   Hydrologic
#> 9   Hydrologic
#> 10  Hydrologic
#>                                                            theme_url
#> 1  https://www.sciencebase.gov/catalog/item/570670f2e4b03f95a075aacd
#> 2  https://www.sciencebase.gov/catalog/item/5785585ee4b0e02680bf2fd6
#> 3  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 4  https://www.sciencebase.gov/catalog/item/566ef828e4b09cfe53ca76f7
#> 5  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 6  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 7  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 8  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 9  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 10 https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#>    characteristic_type
#> 1      localCatch_name
#> 2      localCatch_name
#> 3      localCatch_name
#> 4      localCatch_name
#> 5      localCatch_name
#> 6      localCatch_name
#> 7      localCatch_name
#> 8      localCatch_name
#> 9      localCatch_name
#> 10     localCatch_name