Get National Hydrography Dataset V2 Subsets (Multirealization)

Subsets NHDPlusV2 features by location (POINT), area (POLYGON), or set of COMIDs. Multi realizations are supported allowing you to query for flowlines, catchments, or outlets.

get_nhdplus(
  AOI = NULL,
  comid = NULL,
  nwis = NULL,
  realization = "flowline",
  streamorder = NULL,
  t_srs = NULL
)

Arguments

AOI
sf (MULTI)POINT or (MULTI)POLYGON. An 'area of interest' can be provided as either a location (sf POINT) or area (sf POLYGON) in any Spatial Reference System.

comid
numeric or character. Search for NHD features by COMID(s)

nwis
numeric or character. Search for NHD features by collocated NWIS identifiers

realization
character. What realization to return. Default is flowline and options include: outlet, flowline, catchment, and all

streamorder
numeric or character. Only return NHD flowlines with a streamorder greater then or equal to this value for input value and higher. Only usable with AOI and flowline realizations.

t_srs
character (PROJ string or EPSG code) or numeric (EPSG code). A user specified - target -Spatial Reference System (SRS/CRS) for returned objects. Will default to the CRS of the input AOI if provided, and to 4326 for ID requests.

Value

a single, or list, of simple feature objects

Details

The returned object(s) will have the same Spatial Reference System (SRS) as the input AOI. If a individual or set of IDs are used to query, then the default geoserver CRS of EPSG:4326 is preserved. In all cases, a user-defined SRS can be passed to t_srs which will override all previous SRS's (either input or default). All buffer and distance operations are handled internally using in EPSG:5070 Albers Equal Area projection

Examples

 point <- sf::st_sfc(sf::st_point(c(-119.845, 34.4146)), crs = 4326)
 get_nhdplus(point)

#> Simple feature collection with 1 feature and 139 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -119.8823 ymin: 34.4044 xmax: -119.8256 ymax: 34.4179
#> geographic CRS: WGS 84
#> # A tibble: 1 x 140
#>   id     comid fdate               resolution gnis_id gnis_name lengthkm
#> * <chr>  <int> <dttm>              <chr>      <chr>   <chr>        <dbl>
#> 1 nhdf~ 9.48e8 2008-03-19 23:00:00 Medium     " "     " "           6.78
#> # ... with 133 more variables: reachcode <chr>, flowdir <chr>,
#> #   wbareacomi <int>, ftype <chr>, fcode <int>, shape_length <dbl>,
#> #   streamleve <int>, streamorde <int>, streamcalc <int>, fromnode <dbl>,
#> #   tonode <dbl>, hydroseq <dbl>, levelpathi <dbl>, pathlength <int>,
#> #   terminalpa <dbl>, arbolatesu <int>, divergence <int>, startflag <int>,
#> #   terminalfl <int>, dnlevel <int>, uplevelpat <dbl>, uphydroseq <dbl>,
#> #   dnlevelpat <dbl>, dnminorhyd <int>, dndraincou <int>, dnhydroseq <dbl>,
#> #   frommeas <int>, tomeas <dbl>, rtndiv <int>, vpuin <int>, vpuout <int>,
#> #   areasqkm <dbl>, totdasqkm <dbl>, divdasqkm <dbl>, tidal <int>, totma <int>,
#> #   wbareatype <chr>, pathtimema <int>, hwnodesqkm <chr>, maxelevraw <chr>,
#> #   minelevraw <chr>, maxelevsmo <chr>, minelevsmo <chr>, slope <chr>,
#> #   elevfixed <chr>, hwtype <chr>, slopelenkm <chr>, qa_ma <dbl>, va_ma <int>,
#> #   qc_ma <dbl>, vc_ma <int>, qe_ma <dbl>, ve_ma <int>, qa_01 <dbl>,
#> #   va_01 <int>, qc_01 <dbl>, vc_01 <int>, qe_01 <dbl>, ve_01 <int>,
#> #   qa_02 <dbl>, va_02 <int>, qc_02 <dbl>, vc_02 <int>, qe_02 <dbl>,
#> #   ve_02 <int>, qa_03 <dbl>, va_03 <int>, qc_03 <dbl>, vc_03 <int>,
#> #   qe_03 <dbl>, ve_03 <int>, qa_04 <dbl>, va_04 <int>, qc_04 <dbl>,
#> #   vc_04 <int>, qe_04 <dbl>, ve_04 <int>, qa_05 <dbl>, va_05 <int>,
#> #   qc_05 <dbl>, vc_05 <int>, qe_05 <dbl>, ve_05 <int>, qa_06 <dbl>,
#> #   va_06 <int>, qc_06 <dbl>, vc_06 <int>, qe_06 <dbl>, ve_06 <int>,
#> #   qa_07 <dbl>, va_07 <int>, qc_07 <dbl>, vc_07 <int>, qe_07 <dbl>,
#> #   ve_07 <int>, qa_08 <int>, va_08 <int>, qc_08 <int>, vc_08 <int>,
#> #   qe_08 <int>, ...

 get_nhdplus(point, realization = "catchment")

#> Simple feature collection with 1 feature and 8 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -119.8822 ymin: 34.4042 xmax: -119.8252 ymax: 34.4178
#> geographic CRS: WGS 84
#> # A tibble: 1 x 9
#>   id    gridcode featureid sourcefc areasqkm shape_length shape_area bbox 
#> * <chr>    <int>     <int> <chr>       <dbl>        <dbl>      <dbl> <lis>
#> 1 catc~  1482979 948060316 NHDFlow~     2.58        0.133   0.000253 <dbl~
#> # ... with 1 more variable: geometry <POLYGON [°]>

 get_nhdplus(point, realization = "all")

#> $catchment
#> Simple feature collection with 1 feature and 8 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -119.8822 ymin: 34.4042 xmax: -119.8252 ymax: 34.4178
#> geographic CRS: WGS 84
#> # A tibble: 1 x 9
#>   id    gridcode featureid sourcefc areasqkm shape_length shape_area bbox 
#> * <chr>    <int>     <int> <chr>       <dbl>        <dbl>      <dbl> <lis>
#> 1 catc~  1482979 948060316 NHDFlow~     2.58        0.133   0.000253 <dbl~
#> # ... with 1 more variable: geometry <POLYGON [°]>
#> 
#> $flowline
#> Simple feature collection with 1 feature and 139 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -119.8823 ymin: 34.4044 xmax: -119.8256 ymax: 34.4179
#> geographic CRS: WGS 84
#> # A tibble: 1 x 140
#>   id     comid fdate               resolution gnis_id gnis_name lengthkm
#> * <chr>  <int> <dttm>              <chr>      <chr>   <chr>        <dbl>
#> 1 nhdf~ 9.48e8 2008-03-19 23:00:00 Medium     " "     " "           6.78
#> # ... with 133 more variables: reachcode <chr>, flowdir <chr>,
#> #   wbareacomi <int>, ftype <chr>, fcode <int>, shape_length <dbl>,
#> #   streamleve <int>, streamorde <int>, streamcalc <int>, fromnode <dbl>,
#> #   tonode <dbl>, hydroseq <dbl>, levelpathi <dbl>, pathlength <int>,
#> #   terminalpa <dbl>, arbolatesu <int>, divergence <int>, startflag <int>,
#> #   terminalfl <int>, dnlevel <int>, uplevelpat <dbl>, uphydroseq <dbl>,
#> #   dnlevelpat <dbl>, dnminorhyd <int>, dndraincou <int>, dnhydroseq <dbl>,
#> #   frommeas <int>, tomeas <dbl>, rtndiv <int>, vpuin <int>, vpuout <int>,
#> #   areasqkm <dbl>, totdasqkm <dbl>, divdasqkm <dbl>, tidal <int>, totma <int>,
#> #   wbareatype <chr>, pathtimema <int>, hwnodesqkm <chr>, maxelevraw <chr>,
#> #   minelevraw <chr>, maxelevsmo <chr>, minelevsmo <chr>, slope <chr>,
#> #   elevfixed <chr>, hwtype <chr>, slopelenkm <chr>, qa_ma <dbl>, va_ma <int>,
#> #   qc_ma <dbl>, vc_ma <int>, qe_ma <dbl>, ve_ma <int>, qa_01 <dbl>,
#> #   va_01 <int>, qc_01 <dbl>, vc_01 <int>, qe_01 <dbl>, ve_01 <int>,
#> #   qa_02 <dbl>, va_02 <int>, qc_02 <dbl>, vc_02 <int>, qe_02 <dbl>,
#> #   ve_02 <int>, qa_03 <dbl>, va_03 <int>, qc_03 <dbl>, vc_03 <int>,
#> #   qe_03 <dbl>, ve_03 <int>, qa_04 <dbl>, va_04 <int>, qc_04 <dbl>,
#> #   vc_04 <int>, qe_04 <dbl>, ve_04 <int>, qa_05 <dbl>, va_05 <int>,
#> #   qc_05 <dbl>, vc_05 <int>, qe_05 <dbl>, ve_05 <int>, qa_06 <dbl>,
#> #   va_06 <int>, qc_06 <dbl>, vc_06 <int>, qe_06 <dbl>, ve_06 <int>,
#> #   qa_07 <dbl>, va_07 <int>, qc_07 <dbl>, vc_07 <int>, qe_07 <dbl>,
#> #   ve_07 <int>, qa_08 <int>, va_08 <int>, qc_08 <int>, vc_08 <int>,
#> #   qe_08 <int>, ...
#> 
#> $outlet
#> Simple feature collection with 1 feature and 139 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -119.8295 ymin: 34.4179 xmax: -119.8295 ymax: 34.4179
#> geographic CRS: WGS 84
#> # A tibble: 1 x 140
#>   id     comid fdate               resolution gnis_id gnis_name lengthkm
#> * <chr>  <int> <dttm>              <chr>      <chr>   <chr>        <dbl>
#> 1 nhdf~ 9.48e8 2008-03-19 23:00:00 Medium     " "     " "           6.78
#> # ... with 133 more variables: reachcode <chr>, flowdir <chr>,
#> #   wbareacomi <int>, ftype <chr>, fcode <int>, shape_length <dbl>,
#> #   streamleve <int>, streamorde <int>, streamcalc <int>, fromnode <dbl>,
#> #   tonode <dbl>, hydroseq <dbl>, levelpathi <dbl>, pathlength <int>,
#> #   terminalpa <dbl>, arbolatesu <int>, divergence <int>, startflag <int>,
#> #   terminalfl <int>, dnlevel <int>, uplevelpat <dbl>, uphydroseq <dbl>,
#> #   dnlevelpat <dbl>, dnminorhyd <int>, dndraincou <int>, dnhydroseq <dbl>,
#> #   frommeas <int>, tomeas <dbl>, rtndiv <int>, vpuin <int>, vpuout <int>,
#> #   areasqkm <dbl>, totdasqkm <dbl>, divdasqkm <dbl>, tidal <int>, totma <int>,
#> #   wbareatype <chr>, pathtimema <int>, hwnodesqkm <chr>, maxelevraw <chr>,
#> #   minelevraw <chr>, maxelevsmo <chr>, minelevsmo <chr>, slope <chr>,
#> #   elevfixed <chr>, hwtype <chr>, slopelenkm <chr>, qa_ma <dbl>, va_ma <int>,
#> #   qc_ma <dbl>, vc_ma <int>, qe_ma <dbl>, ve_ma <int>, qa_01 <dbl>,
#> #   va_01 <int>, qc_01 <dbl>, vc_01 <int>, qe_01 <dbl>, ve_01 <int>,
#> #   qa_02 <dbl>, va_02 <int>, qc_02 <dbl>, vc_02 <int>, qe_02 <dbl>,
#> #   ve_02 <int>, qa_03 <dbl>, va_03 <int>, qc_03 <dbl>, vc_03 <int>,
#> #   qe_03 <dbl>, ve_03 <int>, qa_04 <dbl>, va_04 <int>, qc_04 <dbl>,
#> #   vc_04 <int>, qe_04 <dbl>, ve_04 <int>, qa_05 <dbl>, va_05 <int>,
#> #   qc_05 <dbl>, vc_05 <int>, qe_05 <dbl>, ve_05 <int>, qa_06 <dbl>,
#> #   va_06 <int>, qc_06 <dbl>, vc_06 <int>, qe_06 <dbl>, ve_06 <int>,
#> #   qa_07 <dbl>, va_07 <int>, qc_07 <dbl>, vc_07 <int>, qe_07 <dbl>,
#> #   ve_07 <int>, qa_08 <int>, va_08 <int>, qc_08 <int>, vc_08 <int>,
#> #   qe_08 <int>, ...
#> 

 get_nhdplus(comid = 101)

#> Simple feature collection with 1 feature and 139 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -94.6484 ymin: 31.0838 xmax: -94.63 ymax: 31.0992
#> geographic CRS: WGS 84
#> # A tibble: 1 x 140
#>   id    comid fdate               resolution gnis_id gnis_name lengthkm
#> * <chr> <int> <dttm>              <chr>      <chr>   <chr>        <dbl>
#> 1 nhdf~   101 2010-11-30 23:00:00 Medium     " "     " "           3.25
#> # ... with 133 more variables: reachcode <chr>, flowdir <chr>,
#> #   wbareacomi <int>, ftype <chr>, fcode <int>, shape_length <dbl>,
#> #   streamleve <int>, streamorde <int>, streamcalc <int>, fromnode <dbl>,
#> #   tonode <dbl>, hydroseq <dbl>, levelpathi <dbl>, pathlength <dbl>,
#> #   terminalpa <dbl>, arbolatesu <dbl>, divergence <int>, startflag <int>,
#> #   terminalfl <int>, dnlevel <int>, uplevelpat <dbl>, uphydroseq <dbl>,
#> #   dnlevelpat <dbl>, dnminorhyd <int>, dndraincou <int>, dnhydroseq <dbl>,
#> #   frommeas <int>, tomeas <dbl>, rtndiv <int>, vpuin <int>, vpuout <int>,
#> #   areasqkm <dbl>, totdasqkm <dbl>, divdasqkm <dbl>, tidal <int>, totma <dbl>,
#> #   wbareatype <chr>, pathtimema <dbl>, hwnodesqkm <chr>, maxelevraw <int>,
#> #   minelevraw <int>, maxelevsmo <int>, minelevsmo <int>, slope <dbl>,
#> #   elevfixed <chr>, hwtype <chr>, slopelenkm <dbl>, qa_ma <dbl>, va_ma <dbl>,
#> #   qc_ma <dbl>, vc_ma <dbl>, qe_ma <dbl>, ve_ma <dbl>, qa_01 <dbl>,
#> #   va_01 <dbl>, qc_01 <dbl>, vc_01 <dbl>, qe_01 <dbl>, ve_01 <dbl>,
#> #   qa_02 <dbl>, va_02 <dbl>, qc_02 <dbl>, vc_02 <dbl>, qe_02 <dbl>,
#> #   ve_02 <dbl>, qa_03 <dbl>, va_03 <dbl>, qc_03 <dbl>, vc_03 <dbl>,
#> #   qe_03 <dbl>, ve_03 <dbl>, qa_04 <dbl>, va_04 <dbl>, qc_04 <dbl>,
#> #   vc_04 <dbl>, qe_04 <dbl>, ve_04 <dbl>, qa_05 <dbl>, va_05 <dbl>,
#> #   qc_05 <dbl>, vc_05 <dbl>, qe_05 <dbl>, ve_05 <dbl>, qa_06 <dbl>,
#> #   va_06 <dbl>, qc_06 <dbl>, vc_06 <dbl>, qe_06 <dbl>, ve_06 <dbl>,
#> #   qa_07 <dbl>, va_07 <dbl>, qc_07 <dbl>, vc_07 <dbl>, qe_07 <dbl>,
#> #   ve_07 <dbl>, qa_08 <dbl>, va_08 <dbl>, qc_08 <dbl>, vc_08 <dbl>,
#> #   qe_08 <dbl>, ...

 get_nhdplus(nwis  = c(11120000, 11120500))

#> Simple feature collection with 2 features and 139 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -119.8296 ymin: 34.4201 xmax: -119.8057 ymax: 34.4747
#> geographic CRS: WGS 84
#> # A tibble: 2 x 140
#>   id     comid fdate               resolution gnis_id gnis_name lengthkm
#> * <chr>  <int> <dttm>              <chr>      <chr>   <chr>        <dbl>
#> 1 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     269609  Atascade~     1.88
#> 2 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     273489  San Jose~     2.59
#> # ... with 133 more variables: reachcode <chr>, flowdir <chr>,
#> #   wbareacomi <int>, ftype <chr>, fcode <int>, shape_length <dbl>,
#> #   streamleve <int>, streamorde <int>, streamcalc <int>, fromnode <dbl>,
#> #   tonode <dbl>, hydroseq <dbl>, levelpathi <dbl>, pathlength <dbl>,
#> #   terminalpa <dbl>, arbolatesu <dbl>, divergence <int>, startflag <int>,
#> #   terminalfl <int>, dnlevel <int>, uplevelpat <dbl>, uphydroseq <dbl>,
#> #   dnlevelpat <dbl>, dnminorhyd <int>, dndraincou <int>, dnhydroseq <dbl>,
#> #   frommeas <int>, tomeas <int>, rtndiv <int>, vpuin <int>, vpuout <int>,
#> #   areasqkm <dbl>, totdasqkm <dbl>, divdasqkm <dbl>, tidal <int>, totma <dbl>,
#> #   wbareatype <chr>, pathtimema <dbl>, hwnodesqkm <chr>, maxelevraw <int>,
#> #   minelevraw <int>, maxelevsmo <int>, minelevsmo <int>, slope <dbl>,
#> #   elevfixed <chr>, hwtype <chr>, slopelenkm <dbl>, qa_ma <dbl>, va_ma <dbl>,
#> #   qc_ma <dbl>, vc_ma <dbl>, qe_ma <dbl>, ve_ma <dbl>, qa_01 <dbl>,
#> #   va_01 <dbl>, qc_01 <dbl>, vc_01 <dbl>, qe_01 <dbl>, ve_01 <dbl>,
#> #   qa_02 <dbl>, va_02 <dbl>, qc_02 <dbl>, vc_02 <dbl>, qe_02 <dbl>,
#> #   ve_02 <dbl>, qa_03 <dbl>, va_03 <dbl>, qc_03 <dbl>, vc_03 <dbl>,
#> #   qe_03 <dbl>, ve_03 <dbl>, qa_04 <dbl>, va_04 <dbl>, qc_04 <dbl>,
#> #   vc_04 <dbl>, qe_04 <dbl>, ve_04 <dbl>, qa_05 <dbl>, va_05 <dbl>,
#> #   qc_05 <dbl>, vc_05 <dbl>, qe_05 <dbl>, ve_05 <dbl>, qa_06 <dbl>,
#> #   va_06 <dbl>, qc_06 <dbl>, vc_06 <dbl>, qe_06 <dbl>, ve_06 <dbl>,
#> #   qa_07 <dbl>, va_07 <dbl>, qc_07 <dbl>, vc_07 <dbl>, qe_07 <dbl>,
#> #   ve_07 <dbl>, qa_08 <dbl>, va_08 <dbl>, qc_08 <dbl>, vc_08 <dbl>,
#> #   qe_08 <dbl>, ...

 area <- sf::st_as_sfc(sf::st_bbox(c(xmin = -119.8851, xmax =-119.8361,
 ymax = 34.42439, ymin = 34.40473), crs = 4326))
 get_nhdplus(area)

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> Simple feature collection with 19 features and 139 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -119.9143 ymin: 34.4044 xmax: -119.8256 ymax: 34.4548
#> geographic CRS: WGS 84
#> # A tibble: 19 x 140
#>    id     comid fdate               resolution gnis_id gnis_name lengthkm
#>  * <chr>  <int> <dttm>              <chr>      <chr>   <chr>        <dbl>
#>  1 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.725
#>  2 nhdf~ 9.48e8 2008-03-19 23:00:00 Medium     " "     " "          6.78 
#>  3 nhdf~ 9.48e8 2008-03-19 23:00:00 Medium     " "     " "          3.68 
#>  4 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.221
#>  5 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.369
#>  6 nhdf~ 1.76e7 2008-03-19 23:00:00 Medium     "27404~ "Tecolot~    0.867
#>  7 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.833
#>  8 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.241
#>  9 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.91 
#> 10 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.652
#> 11 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.715
#> 12 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.478
#> 13 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.058
#> 14 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.957
#> 15 nhdf~ 1.76e7 2008-06-04 23:00:00 Medium     " "     " "          1.23 
#> 16 nhdf~ 1.76e7 2008-06-04 23:00:00 Medium     " "     " "          1.05 
#> 17 nhdf~ 1.76e7 2008-06-04 23:00:00 Medium     " "     " "          4.42 
#> 18 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.672
#> 19 nhdf~ 9.48e8 2010-11-30 23:00:00 Medium     " "     " "          0.218
#> # ... with 133 more variables: reachcode <chr>, flowdir <chr>,
#> #   wbareacomi <int>, ftype <chr>, fcode <int>, shape_length <dbl>,
#> #   streamleve <int>, streamorde <int>, streamcalc <int>, fromnode <dbl>,
#> #   tonode <dbl>, hydroseq <dbl>, levelpathi <dbl>, pathlength <dbl>,
#> #   terminalpa <dbl>, arbolatesu <dbl>, divergence <int>, startflag <int>,
#> #   terminalfl <int>, dnlevel <int>, uplevelpat <dbl>, uphydroseq <dbl>,
#> #   dnlevelpat <dbl>, dnminorhyd <dbl>, dndraincou <int>, dnhydroseq <dbl>,
#> #   frommeas <dbl>, tomeas <dbl>, rtndiv <int>, vpuin <int>, vpuout <int>,
#> #   areasqkm <dbl>, totdasqkm <dbl>, divdasqkm <dbl>, tidal <int>, totma <dbl>,
#> #   wbareatype <chr>, pathtimema <dbl>, hwnodesqkm <dbl>, maxelevraw <int>,
#> #   minelevraw <int>, maxelevsmo <int>, minelevsmo <int>, slope <dbl>,
#> #   elevfixed <chr>, hwtype <chr>, slopelenkm <dbl>, qa_ma <dbl>, va_ma <dbl>,
#> #   qc_ma <dbl>, vc_ma <dbl>, qe_ma <dbl>, ve_ma <dbl>, qa_01 <dbl>,
#> #   va_01 <dbl>, qc_01 <dbl>, vc_01 <dbl>, qe_01 <dbl>, ve_01 <dbl>,
#> #   qa_02 <dbl>, va_02 <dbl>, qc_02 <dbl>, vc_02 <dbl>, qe_02 <dbl>,
#> #   ve_02 <dbl>, qa_03 <dbl>, va_03 <dbl>, qc_03 <dbl>, vc_03 <dbl>,
#> #   qe_03 <dbl>, ve_03 <dbl>, qa_04 <dbl>, va_04 <dbl>, qc_04 <dbl>,
#> #   vc_04 <dbl>, qe_04 <dbl>, ve_04 <dbl>, qa_05 <dbl>, va_05 <dbl>,
#> #   qc_05 <dbl>, vc_05 <dbl>, qe_05 <dbl>, ve_05 <dbl>, qa_06 <dbl>,
#> #   va_06 <dbl>, qc_06 <dbl>, vc_06 <dbl>, qe_06 <dbl>, ve_06 <dbl>,
#> #   qa_07 <dbl>, va_07 <dbl>, qc_07 <dbl>, vc_07 <dbl>, qe_07 <dbl>,
#> #   ve_07 <dbl>, qa_08 <dbl>, va_08 <dbl>, qc_08 <dbl>, vc_08 <dbl>,
#> #   qe_08 <dbl>, ...

 get_nhdplus(area, realization = "flowline", streamorder = 3)

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> Simple feature collection with 8 features and 139 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -119.8528 ymin: 34.4166 xmax: -119.8295 ymax: 34.4281
#> geographic CRS: WGS 84
#> # A tibble: 8 x 140
#>   id     comid fdate               resolution gnis_id gnis_name lengthkm
#> * <chr>  <int> <dttm>              <chr>      <chr>   <chr>        <dbl>
#> 1 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.725
#> 2 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.221
#> 3 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.369
#> 4 nhdf~ 1.76e7 2008-03-19 23:00:00 Medium     "27404~ "Tecolot~    0.867
#> 5 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.833
#> 6 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.241
#> 7 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.91 
#> 8 nhdf~ 1.76e7 1999-11-18 23:00:00 Medium     " "     " "          0.652
#> # ... with 133 more variables: reachcode <chr>, flowdir <chr>,
#> #   wbareacomi <int>, ftype <chr>, fcode <int>, shape_length <dbl>,
#> #   streamleve <int>, streamorde <int>, streamcalc <int>, fromnode <dbl>,
#> #   tonode <dbl>, hydroseq <dbl>, levelpathi <dbl>, pathlength <dbl>,
#> #   terminalpa <dbl>, arbolatesu <dbl>, divergence <int>, startflag <int>,
#> #   terminalfl <int>, dnlevel <int>, uplevelpat <dbl>, uphydroseq <dbl>,
#> #   dnlevelpat <dbl>, dnminorhyd <dbl>, dndraincou <int>, dnhydroseq <dbl>,
#> #   frommeas <int>, tomeas <int>, rtndiv <int>, vpuin <int>, vpuout <int>,
#> #   areasqkm <dbl>, totdasqkm <dbl>, divdasqkm <dbl>, tidal <int>, totma <int>,
#> #   wbareatype <chr>, pathtimema <int>, hwnodesqkm <chr>, maxelevraw <int>,
#> #   minelevraw <int>, maxelevsmo <int>, minelevsmo <int>, slope <dbl>,
#> #   elevfixed <chr>, hwtype <chr>, slopelenkm <dbl>, qa_ma <dbl>, va_ma <int>,
#> #   qc_ma <dbl>, vc_ma <int>, qe_ma <dbl>, ve_ma <int>, qa_01 <dbl>,
#> #   va_01 <int>, qc_01 <dbl>, vc_01 <int>, qe_01 <dbl>, ve_01 <int>,
#> #   qa_02 <dbl>, va_02 <int>, qc_02 <dbl>, vc_02 <int>, qe_02 <dbl>,
#> #   ve_02 <int>, qa_03 <dbl>, va_03 <int>, qc_03 <dbl>, vc_03 <int>,
#> #   qe_03 <dbl>, ve_03 <int>, qa_04 <dbl>, va_04 <int>, qc_04 <dbl>,
#> #   vc_04 <int>, qe_04 <dbl>, ve_04 <int>, qa_05 <dbl>, va_05 <int>,
#> #   qc_05 <dbl>, vc_05 <int>, qe_05 <dbl>, ve_05 <int>, qa_06 <dbl>,
#> #   va_06 <int>, qc_06 <dbl>, vc_06 <int>, qe_06 <dbl>, ve_06 <int>,
#> #   qa_07 <dbl>, va_07 <int>, qc_07 <dbl>, vc_07 <int>, qe_07 <dbl>,
#> #   ve_07 <int>, qa_08 <dbl>, va_08 <int>, qc_08 <dbl>, vc_08 <int>,
#> #   qe_08 <dbl>, ...