Get NHDPlus HiRes

Get NHDPlus HiRes

get_nhdplushr(
  hr_dir,
  out_gpkg = NULL,
  layers = c("NHDFlowline", "NHDPlusCatchment"),
  pattern = ".*GDB.gdb$",
  check_terminals = TRUE,
  overwrite = FALSE,
  keep_cols = NULL,
  ...
)

Arguments

hr_dir
character directory with geodatabases (gdb search is recursive)

out_gpkg
character path to write output geopackage

layers
character vector with desired layers to return. c("NHDFlowline", "NHDPlusCatchment") is default. Choose from: c("NHDFlowline", "NHDPlusCatchment", "NHDWaterbody", "NHDArea", "NHDLine", "NHDPlusSink", "NHDPlusWall", "NHDPoint", "NHDPlusBurnWaterbody", "NHDPlusBurnLineEvent", "HYDRO_NET_Junctions", "WBDHU2", "WBDHU4","WBDHU6", "WBDHU8" "WBDHU10", "WBDHU12", "WBDLine") Set to NULL to get all available.

pattern
character optional regex to select certain files in hr_dir

check_terminals
boolean if TRUE, run make_standalone on output.

overwrite
boolean should the output overwrite? If false and the output layer exists, it will be read and returned so this function will always return data even if called a second time for the same output. This is useful for workflows. Note that this will NOT delete the entire GeoPackage. It will overwite on a per layer basis.

keep_cols
character vector of column names to keep in the output. If NULL, all will be kept.

...
parameters passed along to get_hr_data for "NHDFlowline" layers.

Value

Response is a list of sf data.frames containing output that may also be written to a geopackage for later use.

Details

NHDFlowline is joined to value added attributes prior to being returned. Names are not modified from the NHDPlusHR geodatabase. Set layers to "NULL" to get all layers.

Examples

# \donttest{
# Note this will download a lot of data to a temp directory.
# Change 'tempdir()' to your directory of choice.
download_dir <- download_nhdplushr(tempdir(), c("0302", "0303"))

get_nhdplushr(download_dir, file.path(download_dir, "nhdplus_0302-03.gpkg"))

#> Error: vector memory exhausted (limit reached?)

get_nhdplushr(download_dir,
              file.path(download_dir, "nhdplus_0302-03.gpkg"),
              layers = NULL, overwrite = TRUE)

#> Warning: GDAL Message 1: organizePolygons() received a polygon with more than 100 parts. The processing may be really slow.  You can skip the processing by setting METHOD=SKIP, or only make it analyze counter-clock wise parts by setting METHOD=ONLY_CCW if you can assume that the outline of holes is counter-clock wise defined

#> Error: vector memory exhausted (limit reached?)

get_nhdplushr(download_dir,
              file.path(download_dir, "nhdplus_0302-03.gpkg"),
              layers = "NHDFlowline", overwrite = TRUE,
              min_size_sqkm = 10, simp = 10, proj = "+init=epsg:5070")

#> Warning: GDAL Message 1: +init=epsg:XXXX syntax is deprecated. It might return a CRS with a non-EPSG compliant axis order.

#> $NHDFlowline
#> Simple feature collection with 45242 features and 57 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: 1414457 ymin: 1355556 xmax: 1795513 ymax: 1640020
#> projected CRS:  NAD83 / Conus Albers
#> # A tibble: 45,242 x 58
#>    Permanent_Ident… FDate               Resolution GNIS_ID GNIS_Name LENGTHKM
#>    <chr>            <dttm>                   <int> <chr>   <chr>        <dbl>
#>  1 31374279         2012-03-11 01:59:53          2 009821… Buffalo …    0.029
#>  2 31373805         2012-03-11 01:59:53          2 009821… Buffalo …    0.092
#>  3 31368135         2012-03-11 03:01:26          2 009920… Perry Cr…    0.513
#>  4 129726996        2012-03-10 22:04:41          2 010239… Crooked …    0.74 
#>  5 129744832        2012-03-10 22:03:34          2 010241… Lynch Cr…    0.059
#>  6 129725633        2012-03-10 22:00:22          2 NA      NA           1.44 
#>  7 129723709        2012-03-10 22:03:44          2 009814… Billys C…    0.889
#>  8 31402583         2012-03-11 01:58:45          2 NA      NA           0.371
#>  9 129721510        2012-03-10 22:04:01          2 010238… Buffalo …    0.39 
#> 10 129720736        2012-03-10 22:01:17          2 010238… Buffalo …    0.116
#> # … with 45,232 more rows, and 52 more variables: REACHCODE <chr>,
#> #   FlowDir <int>, WBArea_Permanent_Identifier <chr>, FTYPE <int>, FCode <int>,
#> #   MainPath <int>, InNetwork <int>, VisibilityFilter <int>,
#> #   Shape_Length <dbl>, COMID <dbl>, VPUID <chr>, Enabled <int>,
#> #   Shape <LINESTRING [m]>, StreamLeve <int>, StreamOrde <int>,
#> #   StreamCalc <int>, FromNode <dbl>, ToNode <dbl>, Hydroseq <dbl>,
#> #   LevelPathI <dbl>, Pathlength <dbl>, TerminalPa <dbl>, ArbolateSu <dbl>,
#> #   Divergence <int>, StartFlag <int>, TerminalFl <dbl>, UpLevelPat <dbl>,
#> #   UpHydroseq <dbl>, DnLevel <dbl>, DnLevelPat <dbl>, DnHydroseq <dbl>,
#> #   DnMinorHyd <dbl>, DnDrainCou <int>, FromMeas <dbl>, ToMeas <dbl>,
#> #   RtnDiv <int>, Thinner <int>, VPUIn <int>, VPUOut <int>, AreaSqKM <dbl>,
#> #   TotDASqKM <dbl>, DivDASqKm <dbl>, MaxElevRaw <dbl>, MinElevRaw <dbl>,
#> #   MaxElevSmo <dbl>, MinElevSmo <dbl>, Slope <dbl>, SlopeLenKm <dbl>,
#> #   ElevFixed <int>, HWType <int>, HWNodeSqKm <dbl>, StatusFlag <chr>
#> 

# }