File path to value added attribute (vaa) Cache

nhdplusTools will download and cache an `fst` file with NHDPlusV2 attribute data sans geometry. This function returns the file path to the cached file. Will use the user data dir indicated by nhdplusTools_data_dir.

get_vaa_path(updated_network = FALSE)

Arguments

updated_network

logical default FALSE. If TRUE, returns path to updated network parameters. See get_vaa for more.

Value

character file path

Details

The VAA data is a aggregate table of information from the NHDPlusV2 elevslope.dbf(s), PlusFlowlineVAA.dbf(s); and NHDFlowlines. All data originates from the EPA NHDPlus Homepage here. To see the location of cached data on your machine use get_vaa_path. To view aggregate data and documentation, see here

Examples

get_vaa_path()
#> [1] "C:\\Users\\dblodgett\\AppData\\Roaming/R/data/R/nhdplusTools/nhdplusVAA.fst"

get_vaa_path(updated_network = TRUE)
#> [1] "C:\\Users\\dblodgett\\AppData\\Roaming/R/data/R/nhdplusTools/enhd_nhdplusatts.fst"