Match Level Paths

Attempts to match dendritic hydrologic unit networks to level paths.

match_levelpaths(fline_hu, start_comid, add_checks = FALSE)

Arguments

fline_hu
sf data.frame flowlines intersected with hydrologic units containing COMID, Hydroseq, LevelPathI, DnLevelPat, denTotalAreaSqKM, HUC12, TOHUC attributes.

start_comid
integer COMID to start search from.

add_checks
boolean if TRUE, checks for toHUC errors are added.

Details

Match level paths compares the set of hydrologic units found through spatial intersection to the path identified by navigating the TOHUC codes from headwater to outlet and attempts of reconcile anomolies found.

This function is preliminary and subject to revision. It has been tested thoroughly but complete methods description have not yet been published.

Examples

net_prep <- readRDS(system.file("extdata/match_levelpaths_2279159.rds", package = "nhdplusTools"))
match_levelpaths(net_prep, 2279159, add_checks = TRUE)

#> # A tibble: 256 x 9
#>    HUC12 TOHUC intersected_Lev… corrected_Level… head_HUC12 outlet_HUC12
#>    <chr> <chr>            <dbl>            <dbl> <chr>      <chr>       
#>  1 1020… 1020…        550003434        550003434 102001011… 102001011102
#>  2 1020… 1020…        550003434        550003434 102001011… 102001011102
#>  3 1020… 1020…        550143605        550143605 102001011… 102001011105
#>  4 1027… 1027…        550002171        550002171 102702010… 102702050705
#>  5 1027… 1027…        550002171        550002171 102702010… 102702050705
#>  6 1027… 1027…        550002171        550049901 102702050… 102702050701
#>  7 1027… 1027…        550002171        550002171 102702010… 102702050705
#>  8 1027… 1027…        550002171        550002171 102702010… 102702050705
#>  9 1027… 1027…        550002171        550002171 102702010… 102702050705
#> 10 1027… 1027…        550002171        550002171 102702010… 102702050705
#> # … with 246 more rows, and 3 more variables: trib_intersect <lgl>,
#> #   trib_no_intersect <lgl>, headwater_error <lgl>