Stage NHDPlus National Data (deprecated)

Breaks down the national geo database into a collection of quick to access R binary files.

stage_national_data(
  include = c("attribute", "flowline", "catchment"),
  output_path = NULL,
  nhdplus_data = NULL,
  simplified = TRUE
)

Arguments

include
character vector containing one or more of: "attributes", "flowline", "catchment".

output_path
character path to save the output to defaults to the directory of the nhdplus_data.

nhdplus_data
character path to the .gpkg or .gdb containing the national seamless dataset. Not required if nhdplus_path has been set.

simplified
boolean if TRUE (the default) the CatchmentSP layer will be included.

Value

list containing paths to the .rds files.

Details

"attributes" will save `NHDFlowline_Network` attributes as a separate data.frame without the geometry. The others will save the `NHDFlowline_Network` and `Catchment` or `CatchmentSP` (per the `simplified` parameter) as sf data.frames with superfluous Z information dropped.

The returned list of paths is also added to the nhdplusTools_env as "national_data".

Examples

stage_national_data,Stage NHDPlus National Data (deprecated),Stage NHDPlus National Data (deprecated) — stage_national_data,Source: R/subset_nhdplus.R,stage_national_data.Rd,character(0),stage_national_data,character(0),list(title = "Description", contents = "
Breaks down the national geo database into a collection\nof quick to access R binary files.
"),list(image = list(src = "https://usgs-r.github.io/nhdplusTools/logo.png"), description = "Breaks down the national geo database into a collection\nof quick to access R binary files."),stage_national_data(
  include = c("attribute", "flowline", "catchment"),
  output_path = NULL,
  nhdplus_data = NULL,
  simplified = TRUE
),list(list(name = "include", description = "
character vector containing one or more of:\n\"attributes\", \"flowline\", \"catchment\".
"), list(name = "output_path", description = "
character path to save the output to defaults\nto the directory of the nhdplus_data.
"), list(name = "nhdplus_data", description = "
character path to the .gpkg or .gdb\ncontaining the national seamless dataset. Not required if\nnhdplus_path has been set.
"), 
    list(name = "simplified", description = "
boolean if TRUE (the default) the CatchmentSP layer\nwill be included.
")),TRUE,

source(system.file("extdata/sample_data.R", package = "nhdplusTools"))

stage_national_data(nhdplus_data = sample_data, output_path = tempdir())

#> $attributes
#> [1] "C:\\Users\\DBLODG~1\\AppData\\Local\\Temp\\1\\RtmpgFVZSX/nhdplus_flowline_attributes.rds"
#> 
#> $flowline
#> [1] "C:\\Users\\DBLODG~1\\AppData\\Local\\Temp\\1\\RtmpgFVZSX/nhdplus_flowline.rds"
#> 
#> $catchment
#> [1] "C:\\Users\\DBLODG~1\\AppData\\Local\\Temp\\1\\RtmpgFVZSX/nhdplus_catchment.rds"
#> 

,list(list(title = "Value", contents = "
list containing paths to the .rds files.
", slug = "value"), list(title = "Details", contents = "
\"attributes\" will save `NHDFlowline_Network` attributes\nas a separate data.frame without the geometry. The others will save\nthe `NHDFlowline_Network` and `Catchment` or `CatchmentSP`\n(per the `simplified` parameter) as sf data.frames with\nsuperfluous Z information dropped.
\n
The returned list of paths is also added to the nhdplusTools_env\nas \"national_data\".
", 
    slug = "details")),2021,list(name = "nhdplusTools", version = "0.4.2", authors = "David Blodgett"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "nhdplusTools"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Get started\n
\n
\n  Reference\n
\n
\n  \n    Articles\n     \n    \n  \n  
\n    
\n      Working with NHDPlusHR\n    
\n    
\n      Plotting with nhdplusTools\n    
\n    
\n      Indexing and Referencing\n    
\n  
\n
\n
\n  Changelog\n
", 
    right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),TRUE