Discover Characteristics Metadata

Provides access to metadata for characteristics that are returned by `get_nldi_characteristics()`.

discover_nldi_characteristics(type = "all")

Arguments

type
character "all", "local", "total", or "divergence_routed".

Value

data.frame containing available characteristics

Examples

chars <- discover_nldi_characteristics()
names(chars)
#> [1] "local"             "total"             "divergence_routed"
head(chars$local, 10)
#>    characteristic_id
#> 1            CAT_HGC
#> 2           CAT_HGCD
#> 3            CAT_HGD
#> 4          CAT_HGVAR
#> 5      CAT_NLCD01_43
#> 6      CAT_NLCD01_52
#> 7      CAT_NLCD01_71
#> 8      CAT_NLCD01_81
#> 9      CAT_NLCD01_82
#> 10     CAT_NLCD01_90
#>                                                                                                                                                                                                                                                   characteristic_description
#> 1         Percentage of Hydrologic Group C soil. -9999 denotes NODATA, usually water. Hydrologic group C soils have slow soil infiltration rates. The soil profiles include layers impeding downward movement of water and, typically, have moderately fine or fine texture.
#> 2  Percentage of Hydrologic Group CD soil. -9999 denotes NODATA, usually water. Hydrologic group CD soils have group C characteristics (slow infiltration rates) when artificially drained and have group D characteristics (very slow infiltration rates) when not drained.
#> 3                                                     Percentage of Hydrologic Group D soil. -9999 denotes NODATA, usually water. Hydrologic group D soils have very slow infiltration rates. Soils are clayey, have a high water table, or have a shallow impervious layer.
#> 4                                                                                                                           Percentage of Hydrologic Group VAR soil. -9999 denotes NODATA, usually water. Hydrologic group VAR soils have variable drainage characteristics.
#> 5                                                                                                                                                                                                                                                                    unknown
#> 6                                                                                                                                                                                                                                                                    unknown
#> 7                                                                                                                                                                                                                                                                    unknown
#> 8                                                                                                                                                                                                                                                                    unknown
#> 9                                                                                                                                                                                                                                                                    unknown
#> 10                                                                                                                                                                                                                                                                   unknown
#>      units
#> 1  percent
#> 2  percent
#> 3  percent
#> 4  percent
#> 5  unknown
#> 6  unknown
#> 7  unknown
#> 8  unknown
#> 9  unknown
#> 10 unknown
#>                                                                                                                                                                                                                                                                 dataset_label
#> 1                                                               Percentage of soils found in an NHDPlus version 2 catchment that contains Hydrologic Group VAR soil. -9999 denotes NODATA,  usually water. Hydrologic group VAR soils have variable drainage characteristics.
#> 2                                                               Percentage of soils found in an NHDPlus version 2 catchment that contains Hydrologic Group VAR soil. -9999 denotes NODATA,  usually water. Hydrologic group VAR soils have variable drainage characteristics.
#> 3                                                               Percentage of soils found in an NHDPlus version 2 catchment that contains Hydrologic Group VAR soil. -9999 denotes NODATA,  usually water. Hydrologic group VAR soils have variable drainage characteristics.
#> 4                                                               Percentage of soils found in an NHDPlus version 2 catchment that contains Hydrologic Group VAR soil. -9999 denotes NODATA,  usually water. Hydrologic group VAR soils have variable drainage characteristics.
#> 5  2011 land-use and land-cover type Emergent Herbaceous Wetlands: Areas where perennial herbaceous vegetation accounts for greater than 80 percent of vegetative cover and the soil or substrate is periodically saturated with or covered with water. -9999 denotes NODATA.
#> 6  2011 land-use and land-cover type Emergent Herbaceous Wetlands: Areas where perennial herbaceous vegetation accounts for greater than 80 percent of vegetative cover and the soil or substrate is periodically saturated with or covered with water. -9999 denotes NODATA.
#> 7  2011 land-use and land-cover type Emergent Herbaceous Wetlands: Areas where perennial herbaceous vegetation accounts for greater than 80 percent of vegetative cover and the soil or substrate is periodically saturated with or covered with water. -9999 denotes NODATA.
#> 8  2011 land-use and land-cover type Emergent Herbaceous Wetlands: Areas where perennial herbaceous vegetation accounts for greater than 80 percent of vegetative cover and the soil or substrate is periodically saturated with or covered with water. -9999 denotes NODATA.
#> 9  2011 land-use and land-cover type Emergent Herbaceous Wetlands: Areas where perennial herbaceous vegetation accounts for greater than 80 percent of vegetative cover and the soil or substrate is periodically saturated with or covered with water. -9999 denotes NODATA.
#> 10 2011 land-use and land-cover type Emergent Herbaceous Wetlands: Areas where perennial herbaceous vegetation accounts for greater than 80 percent of vegetative cover and the soil or substrate is periodically saturated with or covered with water. -9999 denotes NODATA.
#>                                                          dataset_url
#> 1  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 2  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 3  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 4  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 5  https://www.sciencebase.gov/catalog/item/5761b67de4b04f417c2d30ae
#> 6  https://www.sciencebase.gov/catalog/item/5761b67de4b04f417c2d30ae
#> 7  https://www.sciencebase.gov/catalog/item/5761b67de4b04f417c2d30ae
#> 8  https://www.sciencebase.gov/catalog/item/5761b67de4b04f417c2d30ae
#> 9  https://www.sciencebase.gov/catalog/item/5761b67de4b04f417c2d30ae
#> 10 https://www.sciencebase.gov/catalog/item/5761b67de4b04f417c2d30ae
#>    theme_label
#> 1        Soils
#> 2        Soils
#> 3        Soils
#> 4        Soils
#> 5   Land Cover
#> 6   Land Cover
#> 7   Land Cover
#> 8   Land Cover
#> 9   Land Cover
#> 10  Land Cover
#>                                                            theme_url
#> 1  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 2  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 3  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 4  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 5                                                            unknown
#> 6                                                            unknown
#> 7                                                            unknown
#> 8                                                            unknown
#> 9                                                            unknown
#> 10                                                           unknown
#>    characteristic_type
#> 1      localCatch_name
#> 2      localCatch_name
#> 3      localCatch_name
#> 4      localCatch_name
#> 5      localCatch_name
#> 6      localCatch_name
#> 7      localCatch_name
#> 8      localCatch_name
#> 9      localCatch_name
#> 10     localCatch_name