Discover Characteristics Metadata

Provides access to metadata for characteristics that are returned by `get_nldi_characteristics()`.

discover_nldi_characteristics(type = "all")

Arguments

type
character "all", "local", "total", or "divergence_routed".

Examples

chars <- discover_nldi_characteristics()
names(chars)

#> [1] "local"             "total"             "divergence_routed"

head(chars$local, 10)

#>    characteristic_id
#> 1            CAT_BFI
#> 2        CAT_CONTACT
#> 3             CAT_ET
#> 4            CAT_EWT
#> 5            CAT_HGA
#> 6           CAT_HGAC
#> 7           CAT_HGAD
#> 8            CAT_HGB
#> 9           CAT_HGBC
#> 10          CAT_HGBD
#>                                                                                                                                                                                                                                                       characteristic_description
#> 1                 Base Flow Index (BFI), The BFI is a ratio of base flow to total streamflow, expressed as a percentage and ranging from 0 to 100. Base flow is the sustained, slowly varying component of streamflow, usually attributed to ground-water discharge to a stream.
#> 2                                                           Subsurface flow contact time index. The subsurface contact time index estimates the number of days that infiltrated water resides in the saturated subsurface zone of the basin before discharging into the stream. 
#> 3                                                                                                                                                                  Mean-annual actual evapotranspiration (ET), estimated using regression equation of Sanford and Selnick (2013)
#> 4                                                                                                                                                                                                              Average depth to water table relatice to the land surface(meters)
#> 5                                                          Percentage of Hydrologic Group A soil. -9999 denotes NODATA, usually water. Hydrologic group A soils have high infiltration rates. Soils are deep and well drained and, typically, have high sand and gravel content.
#> 6           Percentage of Hydrologic Group AC soil. -9999 denotes NODATA, usually water. Hydrologic group AC soils have group A characteristics (high infiltration rates) when artificially drained and have group C characteristics (slow infiltration rates) when not drained.
#> 7      Percentage of Hydrologic Group AD soil. -9999 denotes NODATA, usually water. Hydrologic group AD soils have group A characteristics (high infiltration rates) when artificially drained and have group D characteristics (very slow infiltration rates) when not drained.
#> 8                                                   Percentage of Hydrologic Group B soil. -9999 denotes NODATA, usually water. Hydrologic group B soils have moderate infiltration rates. Soils are moderately deep, moderately well drained, and moderately coarse in texture.
#> 9       Percentage of Hydrologic Group BC soil. -9999 denotes NODATA, usually water. Hydrologic group BC soils have group B characteristics (moderate infiltration rates) when artificially drained and have group C characteristics (slow infiltration rates) when not drained.
#> 10 Percentage of Hydrologic Group BD soil. -9999 denotes NODATA, usually water. Hydrologic group BD soils have group B characteristics (moderate infiltration rates) when artificially drained and have group D characteristics (very slow infiltration rates) when not drained.
#>      units dataset_label
#> 1  percent              
#> 2     days              
#> 3  mm/year              
#> 4   meters              
#> 5  percent              
#> 6  percent              
#> 7  percent              
#> 8  percent              
#> 9  percent              
#> 10 percent              
#>                                                          dataset_url
#> 1  https://www.sciencebase.gov/catalog/item/5669a8e3e4b08895842a1d4f
#> 2  https://www.sciencebase.gov/catalog/item/56f96fc5e4b0a6037df06b12
#> 3  https://www.sciencebase.gov/catalog/item/5705491be4b0d4e2b756cf8a
#> 4  https://www.sciencebase.gov/catalog/item/56f97456e4b0a6037df06b50
#> 5  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 6  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 7  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 8  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 9  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 10 https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#>    theme_label
#> 1   Hydrologic
#> 2   Hydrologic
#> 3      Climate
#> 4   Hydrologic
#> 5        Soils
#> 6        Soils
#> 7        Soils
#> 8        Soils
#> 9        Soils
#> 10       Soils
#>                                                            theme_url
#> 1  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 2  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 3  https://www.sciencebase.gov/catalog/item/566ef828e4b09cfe53ca76f7
#> 4  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 5  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 6  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 7  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 8  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 9  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 10 https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#>    characteristic_type
#> 1      localCatch_name
#> 2      localCatch_name
#> 3      localCatch_name
#> 4      localCatch_name
#> 5      localCatch_name
#> 6      localCatch_name
#> 7      localCatch_name
#> 8      localCatch_name
#> 9      localCatch_name
#> 10     localCatch_name