Discover Characteristics Metadata

Provides access to metadata for characteristics that are returned by `get_nldi_characteristics()`.

discover_nldi_characteristics(type = "all")

Arguments

type
character "all", "local", "total", or "divergence_routed".

Examples

chars <- discover_nldi_characteristics()
names(chars)

#> [1] "local"             "total"             "divergence_routed"

head(chars$local, 10)

#>    characteristic_id
#> 1            CAT_HGC
#> 2           CAT_HGCD
#> 3            CAT_HGD
#> 4          CAT_HGVAR
#> 5   CAT_ROCKTYPE_999
#> 6            CAT_BFI
#> 7        CAT_CONTACT
#> 8            CAT_EWT
#> 9             CAT_ET
#> 10          CAT_HGAD
#>                                                                                                                                                                                                                                                   characteristic_description
#> 1         Percentage of Hydrologic Group C soil. -9999 denotes NODATA, usually water. Hydrologic group C soils have slow soil infiltration rates. The soil profiles include layers impeding downward movement of water and, typically, have moderately fine or fine texture.
#> 2  Percentage of Hydrologic Group CD soil. -9999 denotes NODATA, usually water. Hydrologic group CD soils have group C characteristics (slow infiltration rates) when artificially drained and have group D characteristics (very slow infiltration rates) when not drained.
#> 3                                                     Percentage of Hydrologic Group D soil. -9999 denotes NODATA, usually water. Hydrologic group D soils have very slow infiltration rates. Soils are clayey, have a high water table, or have a shallow impervious layer.
#> 4                                                                                                                           Percentage of Hydrologic Group VAR soil. -9999 denotes NODATA, usually water. Hydrologic group VAR soils have variable drainage characteristics.
#> 5                                                                                                                                                                          Estimated percent of catchment that is underlain by the Principal Aquifer rock type, Other rocks.
#> 6             Base Flow Index (BFI), The BFI is a ratio of base flow to total streamflow, expressed as a percentage and ranging from 0 to 100. Base flow is the sustained, slowly varying component of streamflow, usually attributed to ground-water discharge to a stream.
#> 7                                                       Subsurface flow contact time index. The subsurface contact time index estimates the number of days that infiltrated water resides in the saturated subsurface zone of the basin before discharging into the stream. 
#> 8                                                                                                                                                                                                          Average depth to water table relatice to the land surface(meters)
#> 9                                                                                                                                                              Mean-annual actual evapotranspiration (ET), estimated using regression equation of Sanford and Selnick (2013)
#> 10 Percentage of Hydrologic Group AD soil. -9999 denotes NODATA, usually water. Hydrologic group AD soils have group A characteristics (high infiltration rates) when artificially drained and have group D characteristics (very slow infiltration rates) when not drained.
#>      units dataset_label
#> 1  percent              
#> 2  percent              
#> 3  percent              
#> 4  percent              
#> 5  percent              
#> 6  percent              
#> 7     days              
#> 8   meters              
#> 9  mm/year              
#> 10 percent              
#>                                                          dataset_url
#> 1  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 2  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 3  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 4  https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#> 5  https://www.sciencebase.gov/catalog/item/582b3855e4b0c253be05fc81
#> 6  https://www.sciencebase.gov/catalog/item/5669a8e3e4b08895842a1d4f
#> 7  https://www.sciencebase.gov/catalog/item/56f96fc5e4b0a6037df06b12
#> 8  https://www.sciencebase.gov/catalog/item/56f97456e4b0a6037df06b50
#> 9  https://www.sciencebase.gov/catalog/item/5705491be4b0d4e2b756cf8a
#> 10 https://www.sciencebase.gov/catalog/item/5728d93be4b0b13d3918a99f
#>    theme_label
#> 1        Soils
#> 2        Soils
#> 3        Soils
#> 4        Soils
#> 5      Geology
#> 6   Hydrologic
#> 7   Hydrologic
#> 8   Hydrologic
#> 9      Climate
#> 10       Soils
#>                                                            theme_url
#> 1  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 2  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 3  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 4  https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#> 5  https://www.sciencebase.gov/catalog/item/5703c816e4b0328dcb82295a
#> 6  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 7  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 8  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 9  https://www.sciencebase.gov/catalog/item/566ef828e4b09cfe53ca76f7
#> 10 https://www.sciencebase.gov/catalog/item/568d6554e4b0e7a44bc207f7
#>    characteristic_type
#> 1      localCatch_name
#> 2      localCatch_name
#> 3      localCatch_name
#> 4      localCatch_name
#> 5      localCatch_name
#> 6      localCatch_name
#> 7      localCatch_name
#> 8      localCatch_name
#> 9      localCatch_name
#> 10     localCatch_name