Discover Characteristics Metadata

Provides access to metadata for characteristics that are returned by `get_nldi_characteristics()`.

discover_nldi_characteristics(type = "all")

Arguments

type

character "all", "local", "total", or "divergence_routed".

Value

data.frame containing available characteristics

Examples

chars <- discover_nldi_characteristics()
names(chars)
#> [1] "local"             "total"             "divergence_routed"
head(chars$local, 10)
#>    characteristic_id
#> 1  CAT_WILDFIRE_2011
#> 2        CAT_CONTACT
#> 3            CAT_EWT
#> 4      CAT_FUNGICIDE
#> 5           CAT_IEOF
#> 6            CAT_PET
#> 7          CAT_RECHG
#> 8          CAT_SATOF
#> 9            CAT_TWI
#> 10    CAT_WB5100_ANN
#>                                                                                                                                                                                                                                          characteristic_description
#> 1                                                                                                                                                                                                                     percent of wild fires per catchments for 2011
#> 2                                               Subsurface flow contact time index. The subsurface contact time index estimates the number of days that infiltrated water resides in the saturated subsurface zone of the basin before discharging into the stream.
#> 3                                                                                                                                                                                                 Average depth to water table relatice to the land surface(meters)
#> 4                                                                                                                                                        Fungicide use on agricultural land in kilograms per square kilometer, 2009 per NHDPlus version 2 catchment
#> 5                                                                                                                                                                                                                   Percentage of Horton overland flow as a percent
#> 6                                                                                                                                                                       Mean-annual potential evapotranspiration (PET), estimated using the Hamon (1961) equation. 
#> 7                                                                                                                                                                                                 Mean annual natural ground-water recharge in millimeters per year
#> 8                                                                                                                                                                                                                    Percentage of Dunne overland flow as a percent
#> 9  Topographic wetness index, ln(a/S); where ln is the natural log, a is the upslope area per unit contour length and S is the slope at that point.  See http://ks.water.usgs.gov/Kansas/pubs/reports/wrir.99-4242.html and Wolock and McCabe, 1995 for more detail
#> 10                                                                                                                                                       Average annual runoff (mm) from McCabe and Wolock's Runoff Model 1951-2000 per NHDPlus version 2 catchment
#>                            units
#> 1                        percent
#> 2                           days
#> 3                         meters
#> 4  kilogram per square kilometer
#> 5                        percent
#> 6                        mm/year
#> 7                          mm/yr
#> 8                        percent
#> 9                          ln(m)
#> 10          millimeters per year
#>                                                                                                                                                                                  dataset_label
#> 1                                                                                                                    Estimated percent of catchment that experienced a wild fire in year XXXX.
#> 2                                                     Contact time, the length of time it takes for water to drain along subsurface flow paths to the stream, per NHDPlus version 2 catchment.
#> 3                                                                                                        Average depth to water table relative to land surface per NHDPlus version 2 catchment
#> 4                                                                                   Fungicide use on agricultural land in kilograms per square kilometer, 2009 per NHDPlus version 2 catchment
#> 5                                                                                            Mean infiltration-excess overland flow as a percent of streamflow per NHDPlus version 2 catchment
#> 6                                                                                                PRISM 30-year average Potential Evapotranspiration, 1971-2000 per NHDPlus version 2 catchment
#> 7                                  Estimated 30 year (1951-1980) mean annual natural groundwater recharge per NHDPlus version 2 catchment. -9999 denotes source data does not cover catchment.
#> 8  Mean saturation overland flow as a percent of streamflow per NHDPlus version 2 catchment. -9999 denotes NODATA or flowline has no catchment. -9999 denotes flowline reach has no catchment.
#> 9                                                                                                                             average topgraphic wetness index per NHDPlus version 2 catchment
#> 10                                                                                  Average annual runoff (mm) from McCabe and Wolock's Runoff Model 1951-2000 per NHDPlus version 2 catchment
#>                                                          dataset_url
#> 1  https://www.sciencebase.gov/catalog/item/57f66d66e4b0bc0bec06de3a
#> 2  https://www.sciencebase.gov/catalog/item/56f96fc5e4b0a6037df06b12
#> 3  https://www.sciencebase.gov/catalog/item/56f97456e4b0a6037df06b50
#> 4  https://www.sciencebase.gov/catalog/item/56fd610fe4b022712b81bf9f
#> 5  https://www.sciencebase.gov/catalog/item/56f974e2e4b0a6037df06b55
#> 6  https://www.sciencebase.gov/catalog/item/56f96ed1e4b0a6037df06a2d
#> 7  https://www.sciencebase.gov/catalog/item/56f97577e4b0a6037df06b5a
#> 8  https://www.sciencebase.gov/catalog/item/56f97acbe4b0a6037df06b6a
#> 9  https://www.sciencebase.gov/catalog/item/56f97be4e4b0a6037df06b70
#> 10 https://www.sciencebase.gov/catalog/item/56fd5bd0e4b0c07cbfa40473
#>    theme_label
#> 1   Land Cover
#> 2   Hydrologic
#> 3   Hydrologic
#> 4     Chemical
#> 5   Hydrologic
#> 6      Climate
#> 7   Hydrologic
#> 8   Hydrologic
#> 9   Hydrologic
#> 10  Hydrologic
#>                                                            theme_url
#> 1                                                            unknown
#> 2  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 3  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 4  https://www.sciencebase.gov/catalog/item/56fd600fe4b022712b81bf9a
#> 5  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 6  https://www.sciencebase.gov/catalog/item/566ef828e4b09cfe53ca76f7
#> 7  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 8  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 9  https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#> 10 https://www.sciencebase.gov/catalog/item/5669a8a4e4b08895842a1d4c
#>    characteristic_type
#> 1      localCatch_name
#> 2      localCatch_name
#> 3      localCatch_name
#> 4      localCatch_name
#> 5      localCatch_name
#> 6      localCatch_name
#> 7      localCatch_name
#> 8      localCatch_name
#> 9      localCatch_name
#> 10     localCatch_name