Get Flowline Index

given an sf point geometry column, return COMID, reachcode, and measure for each.

get_flowline_index(
  flines,
  points,
  search_radius = 0.1,
  precision = NA,
  max_matches = 1
)

Arguments

flines
sf data.frame of type LINESTRING or MULTILINESTRING including COMID, REACHCODE, ToMeas, and FromMeas. Can be "download_nhdplusv2" and remote nhdplusv2 data will be downloaded for the bounding box surround the submitted points. NOTE: The download option may not work for large areas, use with caution.

points
sf or sfc of type POINT

search_radius
numeric the distance for the nearest neighbor search to extend. See RANN nn2 documentation for more details.

precision
numeric the resolution of measure precision in the output in meters.

max_matches
numeric the maximum number of matches to return if multiple are found in search_radius

Value

data.frame with five columns, id, COMID, REACHCODE, REACH_meas, and offset. id is the row or list element in the point input.

Details

Note 1: Inputs are cast into LINESTRINGS. Because of this, the measure output of inputs that are true multipart lines may be in error.

Note 2: This algorithm finds the nearest node in the input flowlines to identify which flowline the point should belong to. As a second pass, it can calculate the measure to greater precision than the nearest flowline geometry node.

Note 3: Offset is returned in units consistent with the projection of the flowlines.

Note 4: See `dfMaxLength` input to sf::st_segmentize() for details of handling of precision parameter.

Examples

sample_flines <- sf::read_sf(system.file("extdata",
                                         "petapsco_flowlines.gpkg",
                                         package = "nhdplusTools"))
get_flowline_index(sample_flines,
                   sf::st_sfc(sf::st_point(c(-76.87479,
                                             39.48233)),
                              crs = 4326))

#> # A tibble: 1 x 5
#>      id    COMID REACHCODE      REACH_meas   offset
#>   <int>    <int> <chr>               <dbl>    <dbl>
#> 1     1 11689978 02060003001472       51.5 0.000301

get_flowline_index("download_nhdplusv2",
                   sf::st_sfc(sf::st_point(c(-76.87479,
                                             39.48233)),
                              crs = 4326))

#> Warning: crs of lines and points don't match. attempting st_transform of points

#> # A tibble: 1 x 5
#>      id    COMID REACHCODE      REACH_meas   offset
#>   <int>    <int> <chr>               <dbl>    <dbl>
#> 1     1 11689978 02060003001472       51.5 0.000311

get_flowline_index(sample_flines,
                   sf::st_sfc(sf::st_point(c(-76.87479,
                                             39.48233)),
                              crs = 4326), precision = 30)

#> # A tibble: 1 x 5
#>      id    COMID REACHCODE      REACH_meas    offset
#>   <int>    <int> <chr>               <dbl>     <dbl>
#> 1     1 11689978 02060003001472       52.3 0.0000458

get_flowline_index(sample_flines,
                   sf::st_sfc(list(sf::st_point(c(-76.86934, 39.49328)),
                                   sf::st_point(c(-76.91711, 39.40884)),
                                   sf::st_point(c(-76.88081, 39.36354))),
                              crs = 4326),
                   search_radius = 0.2,
                   max_matches = 10)

#> # A tibble: 30 x 5
#>       id    COMID REACHCODE      REACH_meas    offset
#>    <int>    <int> <chr>               <dbl>     <dbl>
#>  1     1 11688298 02060003000579        0   0.0000603
#>  2     2 11688808 02060003000519       53.6 0.000564 
#>  3     3 11688980 02060003000253       75.4 0.000310 
#>  4     1 11689926 02060003001467      100   0.0000603
#>  5     3 11688950 02060003000254       18.5 0.00113  
#>  6     1 11689928 02060003001468        0   0.00203  
#>  7     1 11689978 02060003001472      100   0.00203  
#>  8     1 11690530 02060003000585        0   0.00285  
#>  9     1 11690490 02060003000580        0   0.00285  
#> 10     3 11688948 02060003000516        0   0.00321  
#> # ... with 20 more rows