get geoconnex reference feature layers

Queries the geoconnex reference feature server for features of interest.

get_geoconnex_reference(
  AOI = NULL,
  ids = NULL,
  type = NULL,
  t_srs = NULL,
  buffer = 0.5,
  status = TRUE
)

Arguments

AOI

sf (MULTI)POINT or (MULTI)POLYGON. An 'area of interest' can be provided as either a location (sf POINT) or area (sf POLYGON) in any Spatial Reference System.

ids

character or numeric. A set of identifier(s) from the data type requested, for example if NHD, then a set of COMID(s).

type

character. Type of feature to return ('huc08','huc12', 'nhd', 'catchment', 'waterbodies', 'gagesII'). If NULL (default) a data.frame of available resources is returned

t_srs

character (PROJ string or EPSG code) or numeric (EPSG code). A user specified - target -Spatial Reference System (SRS/CRS) for returned objects. Will default to the CRS of the input AOI if provided, and to 4326 for ID requests.

buffer

numeric. The amount (in meters) to buffer a POINT AOI by for an extended search. Default = 0.5

status

boolean print status or not

Value

sf data.frame containing requested reference features

Examples

# \donttest{

dplyr::distinct(discover_geoconnex_reference()[c("id", "title")])
#> # A tibble: 18 × 2
#>    id            title                                                          
#>    <chr>         <chr>                                                          
#>  1 hu02          HU02                                                           
#>  2 hu04          HU04                                                           
#>  3 hu06          HU06                                                           
#>  4 hu08          HU08                                                           
#>  5 hu10          HU10                                                           
#>  6 nat_aq        USGS National Aquifers                                         
#>  7 principal_aq  USGS Principal Aquifers                                        
#>  8 sec_hydrg_reg USGS Secondary Hydrogeologic Regions                           
#>  9 gages         Reference Gages                                                
#> 10 mainstems     Reference Mainstems                                            
#> 11 dams          Reference Dams                                                 
#> 12 pws           Public Water Systems                                           
#> 13 states        States                                                         
#> 14 counties      Counties                                                       
#> 15 aiannh        American Indian/Alaska Native Areas/Hawaiian Home Lands (AIANN…
#> 16 cbsa          Core-based statistical areas (CBSA)                            
#> 17 ua10          Urban Areas                                                    
#> 18 places        Places                                                         

AOI <- sf::st_as_sfc(sf::st_bbox(c(xmin = -89.56684, ymin = 42.99816,
                                   xmax = -89.24681, ymax = 43.17192),
                                 crs = "+proj=longlat +datum=WGS84 +no_defs"))

get_geoconnex_reference(AOI, type = "hu04")
#> Starting download of first 1000 features.
#> Simple feature collection with 2 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -91.15213 ymin: 41.22784 xmax: -88.18307 ymax: 46.21648
#> Geodetic CRS:  WGS 84
#> # A tibble: 2 × 8
#>   id    name        fid gnis_id loaddate            gnis_url uri                
#> * <chr> <chr>     <int> <chr>   <dttm>              <chr>    <chr>              
#> 1 0707  Wisconsin   189 NA      2017-02-01 00:53:08 ""       https://geoconnex.…
#> 2 0709  Rock         65 NA      2012-06-11 07:54:59 ""       https://geoconnex.…
#> # ℹ 1 more variable: geometry <MULTIPOLYGON [°]>

get_geoconnex_reference(type = "hu04", ids = c("0707", "0709"))
#> Simple feature collection with 2 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -91.15213 ymin: 41.22784 xmax: -88.18307 ymax: 46.21648
#> Geodetic CRS:  WGS 84
#> # A tibble: 2 × 8
#>   id    gnis_id   fid name      loaddate            uri                 gnis_url
#> * <chr> <chr>   <int> <chr>     <dttm>              <chr>               <chr>   
#> 1 0707  NA        189 Wisconsin 2017-02-01 00:53:08 https://geoconnex.… ""      
#> 2 0709  NA         65 Rock      2012-06-11 07:54:59 https://geoconnex.… ""      
#> # ℹ 1 more variable: geometry <MULTIPOLYGON [°]>

AOI <- sf::st_sfc(sf::st_point(c(-89.56684, 42.99816)),
                  crs = "+proj=longlat +datum=WGS84 +no_defs")

get_geoconnex_reference(AOI, type = "hu04", buffer = 100000)
#> Starting download of first 1000 features.
#> Simple feature collection with 8 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -93.95968 ymin: 40.31667 xmax: -84.7235 ymax: 46.73351
#> Geodetic CRS:  WGS 84
#> # A tibble: 8 × 8
#>   id    uri                     loaddate            gnis_url   fid gnis_id name 
#> * <chr> <chr>                   <dttm>              <chr>    <int> <chr>   <chr>
#> 1 0403  https://geoconnex.us/r… 2019-04-02 12:17:54 ""         187 NA      Nort…
#> 2 0404  https://geoconnex.us/r… 2019-03-15 10:49:16 ""          66 NA      Sout…
#> 3 0704  https://geoconnex.us/r… 2017-04-18 15:32:25 ""          96 NA      Uppe…
#> 4 0706  https://geoconnex.us/r… 2017-02-01 00:53:13 ""          57 NA      Uppe…
#> 5 0707  https://geoconnex.us/r… 2017-02-01 00:53:08 ""         189 NA      Wisc…
#> 6 0708  https://geoconnex.us/r… 2012-06-11 07:54:59 ""          61 NA      Uppe…
#> 7 0709  https://geoconnex.us/r… 2012-06-11 07:54:59 ""          65 NA      Rock 
#> 8 0712  https://geoconnex.us/r… 2019-03-15 10:49:32 ""          62 NA      Uppe…
#> # ℹ 1 more variable: geometry <MULTIPOLYGON [°]>

# }