get geoconnex reference feature layers

Queries the geoconnex reference feature server for features of interest.

get_geoconnex_reference(
  AOI,
  type = NULL,
  t_srs = NULL,
  buffer = 0.5,
  status = TRUE
)

Arguments

AOI

bbox, sf polygon or point, or a URL that will return an sf object when passed to read_sf

type

character the feature type chosen from discover_geoconnex_reference

t_srs

character (PROJ string or EPSG code) or numeric (EPSG code). A user specified - target -Spatial Reference System (SRS/CRS) for returned objects. Will default to the CRS of the input AOI if provided, and to 4326 for ID requests.

buffer

numeric. The amount (in meters) to buffer a POINT AOI by for an extended search. Default = 0.5

status

boolean print status or not

Value

sf data.frame containing requested reference features

Examples

# \donttest{

dplyr::distinct(discover_geoconnex_reference()[c("id", "title")])
#> # A tibble: 18 × 2
#>    id            title                                                          
#>    <chr>         <chr>                                                          
#>  1 hu02          HU02                                                           
#>  2 hu04          HU04                                                           
#>  3 hu06          HU06                                                           
#>  4 hu08          HU08                                                           
#>  5 hu10          HU10                                                           
#>  6 nat_aq        USGS National Aquifers                                         
#>  7 principal_aq  USGS Principal Aquifers                                        
#>  8 sec_hydrg_reg USGS Secondary Hydrogeologic Regions                           
#>  9 gages         Reference Gages                                                
#> 10 mainstems     Reference Mainstems                                            
#> 11 dams          Reference Dams                                                 
#> 12 pws           Public Water Systems                                           
#> 13 states        States                                                         
#> 14 counties      Counties                                                       
#> 15 aiannh        American Indian/Alaska Native Areas/Hawaiian Home Lands (AIANN…
#> 16 cbsa          Core-based statistical areas (CBSA)                            
#> 17 ua10          Urban Areas                                                    
#> 18 places        Places                                                         

AOI <- sf::st_as_sfc(sf::st_bbox(c(xmin = -89.56684, ymin = 42.99816,
                                   xmax = -89.24681, ymax = 43.17192),
                                 crs = "+proj=longlat +datum=WGS84 +no_defs"))

get_geoconnex_reference(AOI, type = "hu04")
#> Starting download of first set of features.
#> Simple feature collection with 2 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -91.15213 ymin: 41.22784 xmax: -88.18307 ymax: 46.21648
#> Geodetic CRS:  WGS 84
#> # A tibble: 2 × 8
#>   id    uri                     gnis_id name  gnis_url   fid loaddate           
#> * <chr> <chr>                   <chr>   <chr> <chr>    <int> <dttm>             
#> 1 0707  https://geoconnex.us/r… NA      Wisc… ""         189 2017-02-01 00:53:08
#> 2 0709  https://geoconnex.us/r… NA      Rock  ""          65 2012-06-11 07:54:59
#> # ℹ 1 more variable: geometry <MULTIPOLYGON [°]>

get_geoconnex_reference("https://geoconnex.us/ref/mainstems/315626", type = "hu04", )
#> Starting download of first set of features.
#> Simple feature collection with 15 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -106.7484 ymin: 30.05369 xmax: -91.8491 ymax: 39.38306
#> Geodetic CRS:  WGS 84
#> # A tibble: 15 × 8
#>    id    uri                      fid name  loaddate            gnis_url gnis_id
#>  * <chr> <chr>                  <int> <chr> <dttm>              <chr>    <chr>  
#>  1 1102  https://geoconnex.us/…    29 Uppe… 2017-04-17 09:39:52 ""       NA     
#>  2 1103  https://geoconnex.us/…    70 Midd… 2012-06-11 07:54:59 ""       NA     
#>  3 1104  https://geoconnex.us/…   133 Uppe… 2012-06-11 07:54:59 ""       NA     
#>  4 1105  https://geoconnex.us/…   143 Lowe… 2012-06-11 07:54:59 ""       NA     
#>  5 1106  https://geoconnex.us/…   144 Arka… 2017-04-21 13:34:18 ""       NA     
#>  6 1107  https://geoconnex.us/…   145 Neos… 2017-04-21 13:34:23 ""       NA     
#>  7 1108  https://geoconnex.us/…   126 Uppe… 2017-09-20 20:59:45 ""       NA     
#>  8 1109  https://geoconnex.us/…   148 Lowe… 2017-04-24 11:26:44 ""       NA     
#>  9 1110  https://geoconnex.us/…   146 Nort… 2017-04-21 13:15:51 ""       NA     
#> 10 1111  https://geoconnex.us/…     7 Lowe… 2017-04-21 13:15:55 ""       NA     
#> 11 1112  https://geoconnex.us/…   147 Red … 2017-04-24 08:53:22 ""       NA     
#> 12 1113  https://geoconnex.us/…   179 Red-… 2017-04-24 08:53:22 ""       NA     
#> 13 1114  https://geoconnex.us/…   175 Red-… 2017-10-15 23:35:54 ""       NA     
#> 14 1305  https://geoconnex.us/…   178 Rio … 2017-11-05 19:43:51 ""       NA     
#> 15 1306  https://geoconnex.us/…   129 Uppe… 2017-09-20 20:59:52 ""       NA     
#> # ℹ 1 more variable: geometry <MULTIPOLYGON [°]>

AOI <- sf::st_sfc(sf::st_point(c(-89.56684, 42.99816)),
                  crs = "+proj=longlat +datum=WGS84 +no_defs")

get_geoconnex_reference(AOI, type = "hu04", buffer = 100000)
#> Starting download of first set of features.
#> Simple feature collection with 2 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -91.15213 ymin: 41.22784 xmax: -88.18307 ymax: 46.21648
#> Geodetic CRS:  WGS 84
#> # A tibble: 2 × 8
#>   id    name      gnis_id loaddate            uri                   fid gnis_url
#> * <chr> <chr>     <chr>   <dttm>              <chr>               <int> <chr>   
#> 1 0707  Wisconsin NA      2017-02-01 00:53:08 https://geoconnex.…   189 ""      
#> 2 0709  Rock      NA      2012-06-11 07:54:59 https://geoconnex.…    65 ""      
#> # ℹ 1 more variable: geometry <MULTIPOLYGON [°]>

# }