Get NHDPlus HiRes

Get NHDPlus HiRes

get_nhdplushr(hr_dir, out_gpkg = NULL, layers = c("NHDFlowline",
  "NHDPlusCatchment"), pattern = ".*GDB.gdb$", check_terminals = TRUE,
  overwrite = FALSE, ...)

Arguments

hr_dir
character directory with geodatabases (gdb search is recursive)

out_gpkg
character path to write output geopackage

layers
character vector with desired layers to return. c("NHDFlowline", "NHDPlusCatchment") is default. Choose from: c("NHDFlowline", "NHDPlusCatchment", "NHDWaterbody", "NHDArea", "NHDLine", "NHDPlusSink", "NHDPlusWall", "NHDPoint", "NHDPlusBurnWaterbody", "NHDPlusBurnLineEvent", "HYDRO_NET_Junctions", "WBDHU2", "WBDHU4","WBDHU6", "WBDHU8" "WBDHU10", "WBDHU12", "WBDLine") Set to NULL to get all available.

pattern
character optional regex to select certain files in hr_dir

check_terminals
boolean if TRUE, run make_standalone on output.

overwrite
boolean should the output overwrite? If false and the output layer exists, it will be read and returned so this function will always return data even if called a second time for the same output. This is useful for workflows. Note that this will NOT delete the entire GeoPackage. It will overwite on a per layer basis.

...
parameters passed along to get_hr_data for "NHDFlowline" layers.

Value

Response is a list of sf data.frames containing output that may also be written to a geopackage for later use.

Details

NHDFlowline is joined to value added attributes prior to being returned. Names are not modified from the NHDPlusHR geodatabase. Set layers to "NULL" to get all layers.

Examples

# \donttest{
# Note this will download a lot of data to a temp directory.
# Change 'tempdir()' to your directory of choice.
download_dir <- download_nhdplushr(tempdir(), c("0302", "0303"))

get_nhdplushr(download_dir, file.path(download_dir, "nhdplus_0302-03.gpkg"))

#> $NHDFlowline
#> Simple feature collection with 213073 features and 57 fields
#> geometry type:  MULTILINESTRING
#> dimension:      XY
#> bbox:           xmin: -80.06746 ymin: 33.84234 xmax: -75.48283 ymax: 36.44617
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 213,073 x 58
#>    Permanent_Ident… FDate               Resolution GNIS_ID GNIS_Name LENGTHKM
#>    <chr>            <dttm>                   <int> <chr>   <chr>        <dbl>
#>  1 31395075         2012-03-11 01:59:14          2 NA      NA           0.131
#>  2 31395981         2012-03-11 03:05:34          2 NA      NA           0.031
#>  3 31395993         2012-03-11 03:05:34          2 NA      NA           0.062
#>  4 31393833         2012-03-11 03:01:18          2 NA      NA           0.223
#>  5 31394775         2012-03-11 03:13:48          2 NA      NA           0.017
#>  6 31374279         2012-03-11 01:59:53          2 009821… Buffalo …    0.029
#>  7 31374951         2012-03-11 03:00:49          2 NA      NA           0.074
#>  8 31374695         2012-03-11 03:06:58          2 NA      NA           0.684
#>  9 31374725         2012-03-11 03:06:58          2 NA      NA           0.099
#> 10 31373805         2012-03-11 01:59:53          2 009821… Buffalo …    0.092
#> # … with 213,063 more rows, and 52 more variables: REACHCODE <chr>,
#> #   FlowDir <int>, WBArea_Permanent_Identifier <chr>, FTYPE <int>, FCode <int>,
#> #   MainPath <int>, InNetwork <int>, VisibilityFilter <int>,
#> #   Shape_Length <dbl>, COMID <dbl>, VPUID <chr>, Enabled <int>,
#> #   Shape <MULTILINESTRING [°]>, StreamLeve <int>, StreamOrde <int>,
#> #   StreamCalc <int>, FromNode <dbl>, ToNode <dbl>, Hydroseq <dbl>,
#> #   LevelPathI <dbl>, Pathlength <dbl>, TerminalPa <dbl>, ArbolateSu <dbl>,
#> #   Divergence <int>, StartFlag <int>, TerminalFl <int>, UpLevelPat <dbl>,
#> #   UpHydroSeq <dbl>, DnLevel <int>, DnLevelPat <dbl>, DnHydroseq <dbl>,
#> #   DnMinorHyd <dbl>, DnDrainCou <int>, FromMeas <dbl>, ToMeas <dbl>,
#> #   RtnDiv <int>, Thinner <int>, VPUIn <int>, VPUOut <int>, AreaSqKM <dbl>,
#> #   TotDASqKM <dbl>, DivDASqKm <dbl>, MaxElevRaw <dbl>, MinElevRaw <dbl>,
#> #   MaxElevSmo <dbl>, MinElevSmo <dbl>, Slope <dbl>, SlopeLenKm <dbl>,
#> #   ElevFixed <int>, HWType <int>, HWNodeSqKm <dbl>, StatusFlag <chr>
#> 
#> $NHDPlusCatchment
#> Simple feature collection with 185946 features and 7 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.07324 ymin: 33.84224 xmax: -75.48088 ymax: 36.45034
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 185,946 x 8
#>    FEATUREID SourceFC GridCode AreaSqKM VPUID SHAPE_Length SHAPE_Area
#>        <dbl> <chr>       <int>    <dbl> <chr>        <dbl>      <dbl>
#>  1   1.50e13 NHDPlus…     9810   0.574  0302       0.0458  0.0000576 
#>  2   1.50e13 NHDPlus…    11756   0.194  0302       0.0263  0.0000194 
#>  3   1.50e13 NHDPlus…    69563   0.201  0302       0.0253  0.0000202 
#>  4   1.50e13 NHDPlus…   100062   0.250  0302       0.0340  0.0000251 
#>  5   1.50e13 NHDPlus…    90919   0.176  0302       0.0215  0.0000177 
#>  6   1.50e13 NHDPlus…    13162   0.0281 0302       0.00827 0.00000282
#>  7   1.50e13 NHDPlus…    83340   0.258  0302       0.0314  0.0000259 
#>  8   1.50e13 NHDPlus…    12802   0.386  0302       0.0360  0.0000388 
#>  9   1.50e13 NHDPlus…     9775   0.791  0302       0.0526  0.0000794 
#> 10   1.50e13 NHDPlus…    12043   0.357  0302       0.0415  0.0000359 
#> # … with 185,936 more rows, and 1 more variable: SHAPE <MULTIPOLYGON [°]>
#> 

get_nhdplushr(download_dir,
              file.path(download_dir, "nhdplus_0302-03.gpkg"),
              layers = NULL, overwrite = TRUE)

#> $NHDPlusBurnLineEvent
#> Simple feature collection with 186937 features and 12 fields
#> geometry type:  MULTILINESTRING
#> dimension:      XY
#> bbox:           xmin: -80.06698 ymin: 33.84234 xmax: -75.48283 ymax: 36.44529
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 186,937 x 13
#>    FEATUREID GridCode REACHCODE FromMeas ToMeas BurnLenKm RPUID StatusFlag
#>        <dbl>    <int> <chr>        <dbl>  <dbl>     <dbl> <chr> <chr>     
#>  1   1.50e13      960 03020101…    79.4   84.3      0.334 0302  A         
#>  2   1.50e13      961 03020101…     0      2.51     0.063 0302  A         
#>  3   1.50e13      962 03020101…    62.7   78.3      0.385 0302  A         
#>  4   1.50e13      963 03020101…    66.0   80.9      0.47  0302  A         
#>  5   1.50e13     1188 03020101…     0     18.1      0.676 0302  A         
#>  6   1.50e13     1189 03020101…    36.7   54.2      0.183 0302  A         
#>  7   1.50e13     1190 03020101…    76.8  100        0.464 0302  A         
#>  8   1.50e13     1191 03020101…    51.4   54.6      0.051 0302  A         
#>  9   1.50e13     1192 03020101…     4.96  11.9      0.178 0302  A         
#> 10   1.50e13     1193 03020101…     0     19.6      0.191 0302  A         
#> # … with 186,927 more rows, and 5 more variables: Catchment <int>, Burn <int>,
#> #   VPUID <chr>, SHAPE_Length <dbl>, SHAPE <MULTILINESTRING [°]>
#> 
#> $NHDPlusBurnWaterbody
#> Simple feature collection with 51147 features and 8 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.06802 ymin: 33.8532 xmax: -75.48533 ymax: 36.4423
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 51,147 x 9
#>      COMID SourceFC OnOffNet PurpCode  Burn VPUID SHAPE_Length SHAPE_Area
#>      <dbl> <chr>       <int> <chr>    <int> <chr>        <dbl>      <dbl>
#>  1 1.50e13 NHDWate…        1 WB           1 0302      0.00320     6.13e-7
#>  2 1.50e13 NHDWate…        1 WB           1 0302      0.00151     1.51e-7
#>  3 1.50e13 NHDWate…        0 WB           1 0302      0.00213     2.61e-7
#>  4 1.50e13 NHDWate…        0 WB           1 0302      0.000734    2.65e-8
#>  5 1.50e13 NHDWate…        1 WB           1 0302      0.00170     1.49e-7
#>  6 1.50e13 NHDWate…        1 WB           1 0302      0.00514     1.27e-6
#>  7 1.50e13 NHDWate…        1 WB           1 0302      0.00246     2.82e-7
#>  8 1.50e13 NHDWate…        0 WB           1 0302      0.000671    3.18e-8
#>  9 1.50e13 NHDWate…        0 WB           1 0302      0.00180     1.93e-7
#> 10 1.50e13 NHDWate…        0 WB           1 0302      0.00358     9.48e-7
#> # … with 51,137 more rows, and 1 more variable: SHAPE <MULTIPOLYGON [°]>
#> 
#> $NHDPlusCatchment
#> Simple feature collection with 185946 features and 7 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.07324 ymin: 33.84224 xmax: -75.48088 ymax: 36.45034
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 185,946 x 8
#>    FEATUREID SourceFC GridCode AreaSqKM VPUID SHAPE_Length SHAPE_Area
#>        <dbl> <chr>       <int>    <dbl> <chr>        <dbl>      <dbl>
#>  1   1.50e13 NHDPlus…     9810   0.574  0302       0.0458  0.0000576 
#>  2   1.50e13 NHDPlus…    11756   0.194  0302       0.0263  0.0000194 
#>  3   1.50e13 NHDPlus…    69563   0.201  0302       0.0253  0.0000202 
#>  4   1.50e13 NHDPlus…   100062   0.250  0302       0.0340  0.0000251 
#>  5   1.50e13 NHDPlus…    90919   0.176  0302       0.0215  0.0000177 
#>  6   1.50e13 NHDPlus…    13162   0.0281 0302       0.00827 0.00000282
#>  7   1.50e13 NHDPlus…    83340   0.258  0302       0.0314  0.0000259 
#>  8   1.50e13 NHDPlus…    12802   0.386  0302       0.0360  0.0000388 
#>  9   1.50e13 NHDPlus…     9775   0.791  0302       0.0526  0.0000794 
#> 10   1.50e13 NHDPlus…    12043   0.357  0302       0.0415  0.0000359 
#> # … with 185,936 more rows, and 1 more variable: SHAPE <MULTIPOLYGON [°]>
#> 
#> $NHDPlusLandSea
#> Simple feature collection with 367 features and 5 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -79.28467 ymin: 33.72576 xmax: -75.42428 ymax: 36.49496
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 367 x 6
#>      COMID  Land VPUID SHAPE_Length  SHAPE_Area                            SHAPE
#>      <dbl> <int> <chr>        <dbl>       <dbl>               <MULTIPOLYGON [°]>
#>  1 1.50e13     1 0302       0.381       1.63e-3 (((-77.81398 34.18475, -77.8138…
#>  2 1.50e13     1 0302       0.0621      1.50e-4 (((-77.82168 34.19019, -77.8216…
#>  3 1.50e13     1 0302       0.00464     1.31e-6 (((-77.80739 34.19895, -77.8073…
#>  4 1.50e13     1 0302       0.00726     2.67e-6 (((-77.80622 34.2013, -77.80614…
#>  5 1.50e13     1 0302       0.00135     9.27e-8 (((-77.74916 34.30372, -77.7491…
#>  6 1.50e13     1 0302       0.0219      1.25e-5 (((-77.72203 34.31143, -77.7219…
#>  7 1.50e13     1 0302       0.229       1.52e-4 (((-77.72629 34.31415, -77.7261…
#>  8 1.50e13     1 0302       0.0608      7.12e-5 (((-77.71728 34.32186, -77.7171…
#>  9 1.50e13     1 0302       0.0360      9.85e-6 (((-77.34969 34.54426, -77.3493…
#> 10 1.50e13     1 0302       0.00968     2.29e-6 (((-77.33952 34.54479, -77.3395…
#> # … with 357 more rows
#> 
#> $NHDPlusSink
#> Simple feature collection with 369 features and 10 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -79.81687 ymin: 33.93375 xmax: -75.77991 ymax: 36.25015
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 369 x 11
#>      COMID GridCode PurpCode FeatureID SourceFC RPUID StatusFlag Catchment  Burn
#>      <dbl>    <int> <chr>        <dbl> <chr>    <chr> <chr>          <int> <int>
#>  1 1.50e13    78587 SE         1.50e13 NHDFlow… 0302  N                  1     1
#>  2 1.50e13    73682 SE         1.50e13 NHDFlow… 0302  N                  1     1
#>  3 1.50e13    62479 SE         1.50e13 NHDFlow… 0302  N                  1     1
#>  4 1.50e13    50401 SE         1.50e13 NHDFlow… 0302  N                  1     1
#>  5 1.50e13    87344 SE         1.50e13 NHDFlow… 0302  N                  1     1
#>  6 1.50e13    43850 SE         1.50e13 NHDFlow… 0302  N                  1     1
#>  7 1.50e13    18095 SE         1.50e13 NHDFlow… 0302  N                  1     1
#>  8 1.50e13    90089 SE         1.50e13 NHDFlow… 0302  N                  1     1
#>  9 1.50e13    45221 SE         1.50e13 NHDFlow… 0302  N                  1     1
#> 10 1.50e13     4438 SE         1.50e13 NHDFlow… 0302  N                  1     1
#> # … with 359 more rows, and 2 more variables: VPUID <chr>, SHAPE <POINT [°]>
#> 
#> $NHDPlusWall
#> Simple feature collection with 2320 features and 5 fields
#> geometry type:  MULTILINESTRING
#> dimension:      XY
#> bbox:           xmin: -80.07334 ymin: 33.76852 xmax: -75.47986 ymax: 36.45033
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 2,320 x 6
#>      COMID WallSource VPUID SHAPE_Length  Burn                             SHAPE
#>      <dbl> <chr>      <chr>        <dbl> <int>             <MULTILINESTRING [°]>
#>  1 1.50e13 WBDLine    0302        0.0348     1 ((-77.24908 35.3333, -77.24842 3…
#>  2 1.50e13 WBDLine    0302        0.0607     1 ((-76.61913 35.23761, -76.61766 …
#>  3 1.50e13 WBDLine    0302        0.0497     1 ((-77.85314 35.47356, -77.85306 …
#>  4 1.50e13 WBDLine    0302        0.128      1 ((-77.54543 35.50729, -77.54508 …
#>  5 1.50e13 WBDLine    0302        0.108      1 ((-77.61271 35.4832, -77.61223 3…
#>  6 1.50e13 WBDLine    0302        0.0921     1 ((-78.84642 35.75455, -78.84617 …
#>  7 1.50e13 WBDLine    0302        0.0368     1 ((-76.83527 35.06029, -76.83515 …
#>  8 1.50e13 WBDLine    0302        0.0648     0 ((-76.51809 35.13499, -76.51724 …
#>  9 1.50e13 WBDLine    0302        0.0214     1 ((-77.00828 35.14243, -77.00822 …
#> 10 1.50e13 WBDLine    0302        0.111      1 ((-77.44302 35.11485, -77.44265 …
#> # … with 2,310 more rows
#> 
#> $NHDArea
#> Simple feature collection with 1129 features and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XYZ
#> bbox:           xmin: -79.9673 ymin: 33.75017 xmax: -75.37459 ymax: 37.00015
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 1,129 x 15
#>    Permanent_Ident… FDate               Resolution GNIS_ID GNIS_Name AreaSqKM
#>    <chr>            <dttm>                   <int> <chr>   <chr>        <dbl>
#>  1 139013696        2005-08-20 04:44:15          2 NA      NA           0.042
#>  2 139010028        2005-08-20 04:39:26          2 NA      NA           0.011
#>  3 139024288        2005-08-20 04:58:58          2 NA      NA           0.009
#>  4 139024269        2005-08-20 04:58:53          2 NA      NA           0.001
#>  5 136049893        2005-06-07 09:39:34          2 NA      NA           0.013
#>  6 132126143        2005-02-10 15:47:21          2 NA      NA           0.041
#>  7 141615988        2005-10-25 17:09:29          2 NA      NA           0.008
#>  8 157525776        2007-06-22 08:31:36          2 NA      NA           0.003
#>  9 157526605        2007-06-22 08:32:24          2 NA      NA           0.367
#> 10 129758127        2004-12-02 16:38:02          2 NA      NA           2.96 
#> # … with 1,119 more rows, and 9 more variables: Elevation <dbl>, FTYPE <int>,
#> #   FCode <int>, VisibilityFilter <int>, Shape_Length <dbl>, Shape_Area <dbl>,
#> #   COMID <dbl>, VPUID <chr>, Shape <MULTIPOLYGON [°]>
#> 
#> $NHDFlowline
#> Simple feature collection with 213073 features and 57 fields
#> geometry type:  MULTILINESTRING
#> dimension:      XY
#> bbox:           xmin: -80.06746 ymin: 33.84234 xmax: -75.48283 ymax: 36.44617
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 213,073 x 58
#>    Permanent_Ident… FDate               Resolution GNIS_ID GNIS_Name LENGTHKM
#>    <chr>            <dttm>                   <int> <chr>   <chr>        <dbl>
#>  1 31395075         2012-03-11 01:59:14          2 NA      NA           0.131
#>  2 31395981         2012-03-11 03:05:34          2 NA      NA           0.031
#>  3 31395993         2012-03-11 03:05:34          2 NA      NA           0.062
#>  4 31393833         2012-03-11 03:01:18          2 NA      NA           0.223
#>  5 31394775         2012-03-11 03:13:48          2 NA      NA           0.017
#>  6 31374279         2012-03-11 01:59:53          2 009821… Buffalo …    0.029
#>  7 31374951         2012-03-11 03:00:49          2 NA      NA           0.074
#>  8 31374695         2012-03-11 03:06:58          2 NA      NA           0.684
#>  9 31374725         2012-03-11 03:06:58          2 NA      NA           0.099
#> 10 31373805         2012-03-11 01:59:53          2 009821… Buffalo …    0.092
#> # … with 213,063 more rows, and 52 more variables: REACHCODE <chr>,
#> #   FlowDir <int>, WBArea_Permanent_Identifier <chr>, FTYPE <int>, FCode <int>,
#> #   MainPath <int>, InNetwork <int>, VisibilityFilter <int>,
#> #   Shape_Length <dbl>, COMID <dbl>, VPUID <chr>, Enabled <int>,
#> #   Shape <MULTILINESTRING [°]>, StreamLeve <int>, StreamOrde <int>,
#> #   StreamCalc <int>, FromNode <dbl>, ToNode <dbl>, Hydroseq <dbl>,
#> #   LevelPathI <dbl>, Pathlength <dbl>, TerminalPa <dbl>, ArbolateSu <dbl>,
#> #   Divergence <int>, StartFlag <int>, TerminalFl <int>, UpLevelPat <dbl>,
#> #   UpHydroSeq <dbl>, DnLevel <int>, DnLevelPat <dbl>, DnHydroseq <dbl>,
#> #   DnMinorHyd <dbl>, DnDrainCou <int>, FromMeas <dbl>, ToMeas <dbl>,
#> #   RtnDiv <int>, Thinner <int>, VPUIn <int>, VPUOut <int>, AreaSqKM <dbl>,
#> #   TotDASqKM <dbl>, DivDASqKm <dbl>, MaxElevRaw <dbl>, MinElevRaw <dbl>,
#> #   MaxElevSmo <dbl>, MinElevSmo <dbl>, Slope <dbl>, SlopeLenKm <dbl>,
#> #   ElevFixed <int>, HWType <int>, HWNodeSqKm <dbl>, StatusFlag <chr>
#> 
#> $NHDLine
#> Simple feature collection with 3455 features and 12 fields
#> geometry type:  MULTILINESTRING
#> dimension:      XYZ
#> bbox:           xmin: -80.06293 ymin: 33.89045 xmax: -75.50004 ymax: 36.43778
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 3,455 x 13
#>    Permanent_Ident… FDate               Resolution GNIS_ID GNIS_Name LENGTHKM
#>    <chr>            <dttm>                   <int> <chr>   <chr>        <dbl>
#>  1 139025426        2005-08-20 05:00:37          2 NA      NA           0.078
#>  2 139012370        2005-08-20 04:42:34          2 NA      NA           6.82 
#>  3 139017111        2005-08-20 04:48:47          2 NA      NA           0.148
#>  4 139016936        2005-08-20 04:48:34          2 NA      NA           0.848
#>  5 139018701        2005-08-20 04:50:51          2 NA      NA           1.27 
#>  6 139019008        2005-08-20 04:51:25          2 NA      NA           0.969
#>  7 139023859        2005-08-20 04:58:26          2 NA      NA           0.104
#>  8 139023367        2005-08-20 04:57:37          2 NA      NA           0.495
#>  9 139023626        2005-08-20 04:58:04          2 NA      NA           1.51 
#> 10 139023588        2005-08-20 04:58:02          2 NA      NA           1.92 
#> # … with 3,445 more rows, and 7 more variables: FTYPE <int>, FCode <int>,
#> #   VisibilityFilter <int>, Shape_Length <dbl>, COMID <dbl>, VPUID <chr>,
#> #   Shape <MULTILINESTRING [°]>
#> 
#> $NHDPoint
#> Simple feature collection with 205 features and 10 fields
#> geometry type:  POINT
#> dimension:      XYZ
#> bbox:           xmin: -79.95278 ymin: 33.89692 xmax: -76.74611 ymax: 36.19512
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 205 x 11
#>    Permanent_Ident… FDate               Resolution GNIS_ID GNIS_Name REACHCODE
#>    <chr>            <dttm>                   <int> <chr>   <chr>     <chr>    
#>  1 {561E554C-9ACA-… 2015-03-06 10:56:16          2 NA      NA        NA       
#>  2 {E47148F4-D87A-… 2015-03-06 10:56:38          2 NA      NA        NA       
#>  3 {E4AFC45B-59F8-… 2015-03-06 10:54:47          2 NA      NA        NA       
#>  4 {3383D3AE-8A45-… 2015-03-06 10:53:19          2 NA      NA        NA       
#>  5 {4E943A20-7DFB-… 2015-03-06 10:55:30          2 NA      NA        NA       
#>  6 {0D8CF1D6-1633-… 2015-03-06 10:55:30          2 NA      NA        NA       
#>  7 {5E0C8165-0C5D-… 2015-03-06 10:53:19          2 NA      NA        NA       
#>  8 {D212825B-BD81-… 2015-03-06 10:56:38          2 NA      NA        NA       
#>  9 {75E55D94-E988-… 2015-03-06 10:56:16          2 NA      NA        NA       
#> 10 {A8BD10FB-F9C7-… 2015-03-06 10:56:16          2 NA      NA        NA       
#> # … with 195 more rows, and 5 more variables: FTYPE <int>, FCode <int>,
#> #   COMID <dbl>, VPUID <chr>, Shape <POINT [°]>
#> 
#> $NHDWaterbody
#> Simple feature collection with 63998 features and 15 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XYZ
#> bbox:           xmin: -80.06802 ymin: 33.85928 xmax: -75.48223 ymax: 36.4423
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 63,998 x 16
#>    Permanent_Ident… FDate               Resolution GNIS_ID GNIS_Name AreaSqKM
#>    <chr>            <dttm>                   <int> <chr>   <chr>        <dbl>
#>  1 31414233         2002-01-10 00:00:00          2 NA      NA           0.001
#>  2 31413699         2002-01-10 00:00:00          2 NA      NA           0.006
#>  3 31425323         2003-09-19 00:00:00          2 009934… Robertso…    0.618
#>  4 31412701         2002-01-10 00:00:00          2 NA      NA           0.007
#>  5 31411617         2002-01-10 00:00:00          2 NA      NA           0.007
#>  6 129759890        2004-12-03 05:01:04          2 NA      NA           0.005
#>  7 129759009        2004-12-03 04:46:36          2 NA      NA           0.002
#>  8 129758964        2004-12-03 04:45:47          2 NA      NA           0.003
#>  9 129758091        2004-12-03 04:29:04          2 NA      NA           0.003
#> 10 129756601        2004-12-03 04:02:43          2 NA      NA           0.004
#> # … with 63,988 more rows, and 10 more variables: Elevation <dbl>,
#> #   REACHCODE <chr>, FTYPE <int>, FCode <int>, VisibilityFilter <int>,
#> #   Shape_Length <dbl>, Shape_Area <dbl>, COMID <dbl>, VPUID <chr>,
#> #   Shape <MULTIPOLYGON [°]>
#> 
#> $WBDHU10
#> Simple feature collection with 127 features and 16 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.07334 ymin: 33.76852 xmax: -75.47986 ymax: 36.45033
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 127 x 17
#>    TNMID MetaSourceID SourceDataDesc SourceOriginator SourceFeatureID
#>    <chr> <chr>        <chr>          <chr>            <chr>          
#>  1 {F9D… NA           NA             NA               NA             
#>  2 {462… NA           NA             NA               NA             
#>  3 {554… {6CBC8E85-1… NA             NA               NA             
#>  4 {BD3… NA           NA             NA               NA             
#>  5 {9D5… NA           NA             NA               NA             
#>  6 {757… NA           NA             NA               NA             
#>  7 {19F… NA           NA             NA               NA             
#>  8 {5B2… NA           NA             NA               NA             
#>  9 {12C… NA           NA             NA               NA             
#> 10 {A3E… NA           NA             NA               NA             
#> # … with 117 more rows, and 12 more variables: LoadDate <dttm>, AreaSqKM <dbl>,
#> #   AreaAcres <dbl>, GNIS_ID <int>, Name <chr>, States <chr>, HUC10 <chr>,
#> #   HUType <chr>, HUMod <chr>, Shape_Length <dbl>, Shape_Area <dbl>,
#> #   Shape <MULTIPOLYGON [°]>
#> 
#> $WBDHU12
#> Simple feature collection with 728 features and 21 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.07334 ymin: 33.76852 xmax: -75.47986 ymax: 36.45033
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 728 x 22
#>    TNMID MetaSourceID SourceDataDesc SourceOriginator SourceFeatureID
#>    <chr> <chr>        <chr>          <chr>            <chr>          
#>  1 {C24… NA           NA             NA               NA             
#>  2 {4C5… NA           NA             NA               NA             
#>  3 {31D… NA           NA             NA               NA             
#>  4 {819… NA           NA             NA               NA             
#>  5 {E7B… NA           NA             NA               NA             
#>  6 {DCE… NA           NA             NA               NA             
#>  7 {A7B… NA           NA             NA               NA             
#>  8 {8D6… {F85C7E2B-B… NA             NA               NA             
#>  9 {108… NA           NA             NA               NA             
#> 10 {040… NA           NA             NA               NA             
#> # … with 718 more rows, and 17 more variables: LoadDate <dttm>,
#> #   NonContributingAreaAcres <dbl>, NonContributingAreaSqKm <dbl>,
#> #   AreaSqKM <dbl>, AreaAcres <dbl>, GNIS_ID <int>, Name <chr>, States <chr>,
#> #   HUC12 <chr>, HUType <chr>, HUMod <chr>, ToHUC <chr>, Shape_Length <dbl>,
#> #   Shape_Area <dbl>, COMID <dbl>, VPUID <chr>, Shape <MULTIPOLYGON [°]>
#> 
#> $WBDHU2
#> Simple feature collection with 2 features and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -90.6235 ymin: 24.39533 xmax: -75.3981 ymax: 37.52103
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 2 x 15
#>   TNMID MetaSourceID SourceDataDesc SourceOriginator SourceFeatureID
#>   <chr> <chr>        <chr>          <chr>            <chr>          
#> 1 {9C7… NA           NA             NA               NA             
#> 2 {9C7… NA           NA             NA               NA             
#> # … with 10 more variables: LoadDate <dttm>, AreaSqKM <dbl>, AreaAcres <dbl>,
#> #   GNIS_ID <int>, Name <chr>, States <chr>, HUC2 <chr>, Shape_Length <dbl>,
#> #   Shape_Area <dbl>, Shape <MULTIPOLYGON [°]>
#> 
#> $WBDHU4
#> Simple feature collection with 2 features and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.07334 ymin: 33.76852 xmax: -75.47986 ymax: 36.45033
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 2 x 15
#>   TNMID MetaSourceID SourceDataDesc SourceOriginator SourceFeatureID
#>   <chr> <chr>        <chr>          <chr>            <chr>          
#> 1 {E3A… NA           NA             NA               NA             
#> 2 {3F7… NA           NA             NA               NA             
#> # … with 10 more variables: LoadDate <dttm>, AreaSqKM <dbl>, AreaAcres <dbl>,
#> #   GNIS_ID <int>, Name <chr>, States <chr>, HUC4 <chr>, Shape_Length <dbl>,
#> #   Shape_Area <dbl>, Shape <MULTIPOLYGON [°]>
#> 
#> $WBDHU6
#> Simple feature collection with 4 features and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.07334 ymin: 33.76852 xmax: -75.47986 ymax: 36.45033
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 4 x 15
#>   TNMID MetaSourceID SourceDataDesc SourceOriginator SourceFeatureID
#>   <chr> <chr>        <chr>          <chr>            <chr>          
#> 1 {CE8… NA           NA             NA               NA             
#> 2 {B32… NA           NA             NA               NA             
#> 3 {989… NA           NA             NA               NA             
#> 4 {417… NA           NA             NA               NA             
#> # … with 10 more variables: LoadDate <dttm>, AreaSqKM <dbl>, AreaAcres <dbl>,
#> #   GNIS_ID <int>, Name <chr>, States <chr>, HUC6 <chr>, Shape_Length <dbl>,
#> #   Shape_Area <dbl>, Shape <MULTIPOLYGON [°]>
#> 
#> $WBDHU8
#> Simple feature collection with 17 features and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.07334 ymin: 33.76852 xmax: -75.47986 ymax: 36.45033
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 17 x 15
#>    TNMID MetaSourceID SourceDataDesc SourceOriginator SourceFeatureID
#>    <chr> <chr>        <chr>          <chr>            <chr>          
#>  1 {DA4… NA           NA             NA               NA             
#>  2 {300… NA           NA             NA               NA             
#>  3 {BE9… NA           NA             NA               NA             
#>  4 {552… NA           NA             NA               NA             
#>  5 {9B3… NA           NA             NA               NA             
#>  6 {908… NA           NA             NA               NA             
#>  7 {626… NA           NA             NA               NA             
#>  8 {F52… NA           NA             NA               NA             
#>  9 {FBD… NA           NA             NA               NA             
#> 10 {284… NA           NA             NA               NA             
#> 11 {DDB… NA           NA             NA               NA             
#> 12 {5C7… NA           NA             NA               NA             
#> 13 {7B3… NA           NA             NA               NA             
#> 14 {E67… NA           NA             NA               NA             
#> 15 {004… NA           NA             NA               NA             
#> 16 {EE4… NA           NA             NA               NA             
#> 17 {9FE… NA           NA             NA               NA             
#> # … with 10 more variables: LoadDate <dttm>, AreaSqKM <dbl>, AreaAcres <dbl>,
#> #   GNIS_ID <int>, Name <chr>, States <chr>, HUC8 <chr>, Shape_Length <dbl>,
#> #   Shape_Area <dbl>, Shape <MULTIPOLYGON [°]>
#> 
#> $WBDLine
#> Simple feature collection with 2320 features and 9 fields
#> geometry type:  MULTILINESTRING
#> dimension:      XY
#> bbox:           xmin: -80.07334 ymin: 33.76852 xmax: -75.47986 ymax: 36.45033
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 2,320 x 10
#>    TNMID HUDigit HUMod LineSource MetaSourceID LoadDate            Shape_Length
#>    <chr>   <int> <chr> <chr>      <chr>        <dttm>                     <dbl>
#>  1 {796…      10 NA    TOPO24,DR… NA           2012-10-15 12:56:53       0.0983
#>  2 {3E1…      10 NA    TOPO24,DR… NA           2012-10-15 12:56:53       0.255 
#>  3 {402…      10 NA    TOPO24,DR… NA           2012-10-15 12:56:42       0.228 
#>  4 {AB0…      10 NA    TOPO24,DR… NA           2012-10-15 12:56:42       0.109 
#>  5 {111…      12 NA    TOPO24,DR… NA           2012-10-15 12:56:42       0.0689
#>  6 {46A…      12 NA    TOPO24,DR… NA           2012-10-15 12:56:42       0.0272
#>  7 {990…       6 NA    TOPO24,DR… NA           2012-10-15 12:56:42       0.0286
#>  8 {C4E…      12 NA    TOPO24,DR… NA           2012-10-15 12:56:42       0.0623
#>  9 {F91…      10 NA    TOPO24,DR… NA           2012-10-15 12:56:53       0.0268
#> 10 {5B0…      10 NA    TOPO24,DR… NA           2012-10-15 12:56:53       0.0309
#> # … with 2,310 more rows, and 3 more variables: COMID <dbl>, VPUID <chr>,
#> #   Shape <MULTILINESTRING [°]>
#> 
#> $HYDRO_NET_Junctions
#> Simple feature collection with 205950 features and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -80.06746 ymin: 33.8536 xmax: -75.48296 ymax: 36.44617
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 205,950 x 2
#>    Enabled                Shape
#>      <int>          <POINT [°]>
#>  1       1 (-79.22356 36.06807)
#>  2       1 (-79.22117 36.07732)
#>  3       1 (-79.21987 36.07227)
#>  4       1 (-79.21977 36.04575)
#>  5       1 (-79.21945 36.04558)
#>  6       1 (-79.21937 36.08515)
#>  7       1 (-79.21504 36.06958)
#>  8       1 (-79.21484 36.05786)
#>  9       1 (-79.21476 36.06924)
#> 10       1 (-79.21457 36.05783)
#> # … with 205,940 more rows
#> 

get_nhdplushr(download_dir,
              file.path(download_dir, "nhdplus_0302-03.gpkg"),
              layers = "NHDFlowline", overwrite = TRUE,
              min_size_sqkm = 10, simp = 10, proj = "+init=epsg:5070")

#> $NHDFlowline
#> Simple feature collection with 45242 features and 57 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: 1414457 ymin: 1355556 xmax: 1795513 ymax: 1640020
#> epsg (SRID):    5070
#> proj4string:    +proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 45,242 x 58
#>    Permanent_Ident… FDate               Resolution GNIS_ID GNIS_Name LENGTHKM
#>    <chr>            <dttm>                   <int> <chr>   <chr>        <dbl>
#>  1 31374279         2012-03-11 01:59:53          2 009821… Buffalo …    0.029
#>  2 31373805         2012-03-11 01:59:53          2 009821… Buffalo …    0.092
#>  3 31368135         2012-03-11 03:01:26          2 009920… Perry Cr…    0.513
#>  4 129726996        2012-03-10 22:04:41          2 010239… Crooked …    0.74 
#>  5 129744832        2012-03-10 22:03:34          2 010241… Lynch Cr…    0.059
#>  6 129725633        2012-03-10 22:00:22          2 NA      NA           1.44 
#>  7 129723709        2012-03-10 22:03:44          2 009814… Billys C…    0.889
#>  8 31402583         2012-03-11 01:58:45          2 NA      NA           0.371
#>  9 129721510        2012-03-10 22:04:01          2 010238… Buffalo …    0.39 
#> 10 129720736        2012-03-10 22:01:17          2 010238… Buffalo …    0.116
#> # … with 45,232 more rows, and 52 more variables: REACHCODE <chr>,
#> #   FlowDir <int>, WBArea_Permanent_Identifier <chr>, FTYPE <int>, FCode <int>,
#> #   MainPath <int>, InNetwork <int>, VisibilityFilter <int>,
#> #   Shape_Length <dbl>, COMID <dbl>, VPUID <chr>, Enabled <int>,
#> #   Shape <LINESTRING [m]>, StreamLeve <int>, StreamOrde <int>,
#> #   StreamCalc <int>, FromNode <dbl>, ToNode <dbl>, Hydroseq <dbl>,
#> #   LevelPathI <dbl>, Pathlength <dbl>, TerminalPa <dbl>, ArbolateSu <dbl>,
#> #   Divergence <int>, StartFlag <int>, TerminalFl <int>, UpLevelPat <dbl>,
#> #   UpHydroSeq <dbl>, DnLevel <int>, DnLevelPat <dbl>, DnHydroseq <dbl>,
#> #   DnMinorHyd <dbl>, DnDrainCou <int>, FromMeas <dbl>, ToMeas <dbl>,
#> #   RtnDiv <int>, Thinner <int>, VPUIn <int>, VPUOut <int>, AreaSqKM <dbl>,
#> #   TotDASqKM <dbl>, DivDASqKm <dbl>, MaxElevRaw <dbl>, MinElevRaw <dbl>,
#> #   MaxElevSmo <dbl>, MinElevSmo <dbl>, Slope <dbl>, SlopeLenKm <dbl>,
#> #   ElevFixed <int>, HWType <int>, HWNodeSqKm <dbl>, StatusFlag <chr>
#> 

# }