Get Path Members

Given a network and set of IDs, finds paths between all identified flowpath outlets. This algorithm finds members between outlets regardless of flow direction.

get_path_members(outlets, network, cores = 1, status = FALSE)

Arguments

outlets

vector of IDs from data.frame

network

data.frame with ID, toID, and lengthkm attributes.

cores

integer number of cores to use for parallel computation.

status

logical print status and progress bars?

Value

list of lists containing flowpath identifiers along path that connect outlets.

Examples

source(system.file("extdata", "walker_data.R", package = "nhdplusTools"))
fline <- walker_flowline

outlets <- c(5329303, 5329357, 5329317, 5329365, 5329435, 5329817)

# Add toCOMID
fline <- nhdplusTools::get_tocomid(fline, add = TRUE)

fl <- dplyr::select(fline, ID = comid, toID = tocomid, lengthkm)

get_path_members(outlets, fl)
#>    id_1 id_2                                                      path
#> 1     1    8                                       7, 6, 5, 4, 3, 2, 1
#> 2     1    4                                                   3, 2, 1
#> 3     1    9                                    8, 7, 6, 5, 4, 3, 2, 1
#> 4     1   16 15, 14, 13, 12, 11, 60, 62, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1
#> 5     1   19                                     18, 17, 5, 4, 3, 2, 1
#> 6     8    4                                                7, 6, 5, 4
#> 7     8    9                                                         8
#> 8     8   16                      15, 14, 13, 12, 11, 60, 62, 10, 9, 8
#> 9     8   19                                              7, 6, 18, 17
#> 10    4    9                                             8, 7, 6, 5, 4
#> 11    4   16          15, 14, 13, 12, 11, 60, 62, 10, 9, 8, 7, 6, 5, 4
#> 12    4   19                                              18, 17, 5, 4
#> 13    9   16                         15, 14, 13, 12, 11, 60, 62, 10, 9
#> 14    9   19                                           8, 7, 6, 18, 17
#> 15   16   19        15, 14, 13, 12, 11, 60, 62, 10, 9, 8, 7, 6, 18, 17