NHDPlusV2 Attribute Subset

Return requested NHDPlusv2 Attributes

get_vaa(atts = NULL, path = get_vaa_path(), download = TRUE)

Arguments

atts
character The variable names you would like, always includes comid

path
character path where the file should be saved. Default is a persistent system data as retrieved by nhdplusTools_data_dir. Also see: get_vaa_path

download
logical if TRUE, the default, will download VAA table if not found at path.

Value

data.frame

Details

The VAA data is a aggregate table of information from the NHDPlusV2 elevslope.dbf(s), PlusFlowlineVAA.dbf(s); and NHDFlowlines. All data originates from the EPA NHDPlus Homepage here. To see the location of cached data on your machine use get_vaa_path. To view aggregate data and documentation, see here

Examples

# \donttest{
get_vaa("slope")
#> didn't find data, downloading.
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |=========                                                             |  14%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |================                                                      |  24%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |=======================                                               |  34%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |==============================                                        |  44%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |============================================                          |  64%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |===================================================                   |  74%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |==========================================================            |  84%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |=================================================================     |  94%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
#>       comid          slope
#> 1   8318793     0.06083397
#> 2   8318787     0.09534391
#> 3   8318775     0.11273792
#> 4   8318785     0.15142342
#> 5   8318789     0.16246142
#> 6   8318801     0.10504926
#> 7   8318819     0.15627586
#> 8   8318813     0.14838480
#> 9   8318825     0.17851902
#> 10  8318841     0.19833333
#> 11  8321213     0.30318281
#> 12  8319971     0.17701588
#> 13  8319977     0.15134239
#> 14  8319983     0.05639584
#> 15  8319985     0.22391050
#> 16  8319989     0.22171945
#> 17  8319987     0.26929955
#> 18  8321221     0.25149134
#> 19  8319995     0.21252046
#> 20  8320027     0.02283072
#> 21  8320029     0.18152041
#> 22  8320035     0.15245065
#> 23  8320037     0.10530418
#> 24  8320049     0.01863346
#> 25  8321227     0.09343534
#> 26  8320067     0.18970149
#> 27  8320051     0.03570175
#> 28  8320045     0.17278583
#> 29  8320047     0.02871794
#> 30  8320071     0.08506246
#> 31  8320075     0.01661354
#> 32  8320081     0.09785286
#> 33  8320077     0.09563380
#> 34  8320091     0.00654941
#> 35  8320093     0.06578509
#> 36  8320121     0.06613976
#> 37  8320117     0.03421798
#> 38  8320129     0.01167101
#> 39  8320131     0.09507233
#> 40  8320157     0.04404825
#> 41  8321247     0.06464353
#> 42  8320159     0.04644230
#> 43  8320171     0.01326405
#> 44  8320177     0.01551326
#> 45  8321253     0.08487241
#> 46  8321251     0.09899641
#> 47  8320187     0.11967595
#> 48  8320195     0.08029826
#> 49  8320175     0.13833488
#> 50  8320165     0.04617313
#> 51  8320153     0.02977542
#> 52  8320137     0.08623712
#> 53  8320113     0.03007255
#> 54  8320097     0.07578203
#> 55  8320115     0.11378725
#> 56  8321243     0.08227124
#> 57  8321239     0.07812816
#> 58  8321235     0.20535179
#> 59  8320061     0.13411111
#> 60  8320059     0.10721804
#> 61  8321241     0.14880311
#> 62  8320083     0.08166377
#> 63  8320063     0.11007383
#> 64  8321231     0.16020148
#> 65  8320039     0.11612556
#> 66  8321225     0.24164087
#> 67  8320021     0.00001000
#> 68  8320011     0.20397430
#> 69  8321223     0.22219711
#> 70  8320017     0.15138947
#> 71  8321281     0.00813725
#> 72  8319997     0.02553872
#> 73  8320015     0.41857142
#> 74  8320009     0.16588073
#> 75  8320005     0.14481554
#> 76  8319991     0.02752808
#> 77  8319999     0.11584666
#> 78  8319993     0.00001000
#> 79  8319973     0.23622572
#> 80  8319967     0.23350910
#> 81  8319961     0.16125262
#> 82  8319957     0.18179506
#> 83  8319953     0.20880579
#> 84  8319949     0.23240081
#> 85  8321217     0.20355245
#> 86  8319943     0.19102425
#> 87  8321215     0.23862620
#> 88  8319937     0.23577581
#> 89  8319931     0.27705882
#> 90  8319933     0.22026553
#> 91  8318847     0.24131455
#> 92  8319927     0.05524271
#> 93  8318845     0.23947907
#> 94  8318831     0.17320279
#> 95  8318823     0.00001000
#> 96  8318821     0.25119738
#> 97  8318829     0.19621969
#> 98  8318827     0.22418478
#> 99  8318815     0.26289540
#> 100 8318805     0.29542769
#> 101 8318803     0.24739652
#> 102 8318807     0.22145433
#> 103 8318811     0.21544303
#> 104 8318777     0.05770716
#> 105 8318761     0.12000000
#> 106 8318759     0.23509090
#> 107 8318765     0.20424190
#> 108 8318763     0.16822938
#> 109 8318757     0.29032932
#> 110 8318731     0.17643985
#> 111 8318721     0.35101086
#> 112 8318717     0.24808504
#> 113 8318699     0.26899950
#> 114 8318679     0.30396946
#> 115 8318663     0.25616666
#> 116 8318647     0.25100595
#> 117 8318615     0.22120736
#> 118 8318611     0.26263636
#> 119 8318549     0.25105691
#> 120 8318565     0.19246428
#> 121 8318637     0.10202711
#> 122 8318651     0.16245378
#> 123 8318671     0.06377579
#> 124 8318683     0.14034525
#> 125 8318703     0.17897260
#> 126 8318751     0.27488594
#> 127 8318745     0.17518259
#> 128 8318723     0.20423864
#> 129 8318711     0.17488664
#> 130 8318685     0.20989715
#> 131 8318673     0.17709493
#> 132 8318667     0.26097544
#> 133 8318635     0.17557298
#> 134 8318591     0.18856655
#> 135 8318621     0.18194473
#> 136 8318583     0.24591463
#> 137 8318623     0.20157046
#> 138 8318587     0.24492943
#> 139 8318551     0.24719232
#> 140 8318543     0.21510492
#> 141 8318515     0.26665768
#> 142 8318517     0.21239255
#> 143 8318533     0.24063420
#> 144 8318525     0.19292817
#> 145 8318523     0.18834862
#> 146 8318571     0.16607043
#> 147 8318531     0.18519337
#> 148 8318567     0.27565196
#> 149 8318597     0.18999166
#> 150 8318541     0.22647406
#> 151 8318579     0.16452030
#> 152 8318561     0.16763896
#> 153 8318537     0.21979455
#> 154 8318519     0.20262769
#> 155 8318509     0.11969230
#> 156 8318507     0.17174890
#> 157 8318489     0.18540173
#> 158 8318481     0.19705542
#> 159 8318479     0.15183537
#> 160 8318469     0.14901062
#> 161 8318457     0.17986104
#> 162 8318453     0.24622478
#> 163 8318459     0.20149580
#> 164 8318447     0.17287345
#> 165 8318419     0.20900488
#> 166 8318417     0.26889245
#> 167 8318397     0.00671256
#> 168 8318387     0.00661290
#> 169 8318373     0.17038401
#> 170 8319345     0.14720328
#> 171 8318441     0.06377997
#> 172 8318429     0.22560439
#> 173 8318461     0.12258526
#> 174 8318439     0.13163910
#> 175 8318451     0.19980730
#> 176 8318463     0.27626911
#> 177 8318455     0.19623214
#> 178 8318363     0.21902129
#> 179 8318339     0.17626815
#> 180 8318315     0.09532915
#> 181 8318303     0.11242044
#> 182 8318265     0.18525849
#> 183 8318263     0.15567434
#> 184 8318321     0.13117056
#> 185 8318299     0.16531281
#> 186 8318259     0.16725624
#> 187 8318241     0.21280571
#> 188 8318233     0.17521776
#> 189 8318239     0.18268981
#> 190 8318247     0.06592686
#> 191 8318235     0.09654333
#> 192 8318209     0.08230380
#> 193 8318155     0.18245810
#> 194 8318163     0.14550351
#> 195 8318187     0.10515584
#> 196 8318151     0.06622287
#> 197 8318089     0.18605161
#> 198 8318067     0.20757281
#> 199 8318117     0.23455231
#> 200 8318139     0.18482514
#> 201 8318157     0.17295377
#> 202 8318173     0.00001000
#> 203 8318059     0.00044067
#> 204 8316847     0.00497515
#> 205 8316841     0.00359223
#> 206 8316833     0.09120678
#> 207 8317993     0.13712318
#> 208 8318001     0.10702258
#> 209 8316839     0.10149700
#> 210 8316835     0.01319985
#> 211 8316783     0.01076880
#> 212 8317923     0.15903312
#> 213 8317929     0.09793542
#> 214 8316773     0.13117192
#> 215 8316771     0.12152348
#> 216 8316765     0.01295918
#> 217 8316757     0.09446746
#> 218 8316761     0.11045407
#> 219 8316753     0.02096062
#> 220 8316777     0.02803252
#> 221 8316829     0.00001000
#> 222 8316817     0.01330336
#> 223 8316795     0.02692786
#> 224 8316797     0.02408631
#> 225 8316813     0.00453125
#> 226 8316803     0.00016666
#> 227 8316793     0.01661529
#> 228 8316779     0.03502164
#> 229 8317415     0.00001000
#> 230 8317403     0.00001000
#> 231 8317409     0.00109452
#> 232 8323713     0.00001000
#> 233 8323701     0.00001000
#> 234 8317397     0.00001000
#> 235 8323681     0.00001000
#> 236 8317385     0.00001000
#> 237 8317547     0.00001000
#> 238 8323689     0.00001000
#> 239 8317411     0.00001000
#> 240 8316869     0.00054676
#> 241 8317405     0.00084507
#> 242 8316827     0.00001000
#> 243 8317549     0.00001000
#> 244 8317381     0.00001000
#> 245 8323687     0.00001000
#> 246 8323685     0.00001000
#> 247 8323711     0.00001000
#> 248 8317551 -9998.00000000
#> 249 8316843     0.00001000
#> 250 8317545 -9998.00000000
#> 251 8317383     0.00001000
#> 252 8316851     0.00001104
#> 253 8317391     0.00057803
#> 254 8316865     0.00104590
#> 255 8317407     0.00001000
#> 256 8316873     0.07239559
#> 257 8316875     0.00203339
#> 258 8317395     0.00559322
#> 259 8316895     0.10462288
#> 260 8318355     0.12495589
#> 261 8318319     0.05431400
#> 262 8318279     0.16676411
#> 263 8318267     0.13028862
#> 264 8316891     0.17284588
#> 265 8318223     0.10440000
#> 266 8316879     0.03170596
#> 267 8316893     0.00014373
#> 268 8317437     0.00001000
#> 269 8316907     0.00066945
#> 270 8316899     0.00001000
#> 271 8316897     0.00001277
#> 272 8316903     0.05268866
#> 273 8316921     0.00083076
#> 274 8316927     0.00646812
#> 275 8317361     0.06109024
#> 276 8316961     0.00221788
#> 277 8317355     0.02072972
#> 278 8316975     0.00001000
#> 279 8317015     0.00188913
#> 280 8317019     0.00808721
#> 281 8317007     0.07347537
#> 282 8317009     0.05994123
#> 283 8317353     0.02425149
#> 284 8317027     0.00187997
#> 285 8317021     0.03714487
#> 286 8317023     0.02752427
#> 287 8317037     0.03582191
#> 288 8317025     0.02556616
#> 289 8317031     0.04136929
#> 290 8316979     0.02415868
#> 291 8316983     0.05576820
#> 292 8316973     0.00879569
#> 293 8317357     0.00028106
#> 294 8317359     0.04341903
#> 295 8316977     0.06900000
#> 296 8316965     0.01019472
#> 297 8316959     0.03993899
#> 298 8316931     0.03037569
#> 299 8316995     0.05375880
#> 300 8317013     0.06520454
#> 301 8317029     0.13998334
#> 302 8317017     0.13808497
#> 303 8318407     0.21706255
#> 304 8318411     0.22624365
#> 305 8316971     0.11520249
#> 306 8316951     0.02379867
#> 307 8316933     0.08393221
#> 308 8316943     0.13554512
#> 309 8316937     0.02089053
#> 310 8316923     0.08375663
#> 311 8316925     0.07428844
#> 312 8316917     0.00505456
#> 313 8317435     0.00012422
#> 314 8316919     0.00619329
#> 315 8317431     0.00037037
#> 316 8316905     0.00497085
#> 317 8317427     0.00001000
#> 318 8317055     0.06883307
#> 319 8317061     0.07491298
#> 320 8317109     0.06566137
#> 321 8317065     0.02263492
#> 322 8317113     0.06500217
#> 323 8317101     0.05054779
#> 324 8317107     0.02146710
#> 325 8317123     0.07224127
#> 326 8317127     0.01984421
#> 327 8317139     0.05674228
#> 328 8317143     0.09135764
#> 329 8317173     0.03607496
#> 330 8317149     0.10958143
#> 331 8317163     0.07931756
#> 332 8317175     0.12950185
#> 333 8317141     0.09573283
#> 334 8317115     0.08041105
#> 335 8317119     0.04114114
#> 336 8317129     0.14667189
#> 337 8317135     0.08669142
#> 338 8317091     0.01686138
#> 339 8317089     0.01522058
#> 340 8318505     0.19094984
#> 341 8317097     0.04900173
#> 342 8317085     0.11303896
#> 343 8317083     0.09340928
#> 344 8317071     0.03523792
#> 345 8317077     0.03327027
#> 346 8317067     0.01992303
#> 347 8317039     0.04632575
#> 348 8317069     0.04293243
#> 349 8317057     0.03141729
#> 350 8317053     0.03846043
#> 351 8316967     0.00118787
#> 352 8317345     0.00001000
#> 353 8316969     0.00073946
#> 354 8316949     0.00006191
#> 355 8316941     0.00007177
#> 356 8316947     0.01202265
#> 357 8317351     0.01662127
#> 358 8317347     0.01883320
#> 359 8317349     0.02139447
#> 360 8316935     0.00954786
#> 361 8316939     0.00394970
#> 362 8317045     0.00001000
#> 363 8317535     0.00199829
#> 364 8317095     0.00055234
#> 365 8317155     0.01816800
#> 366 8317145     0.00630944
#> 367 8317167     0.02934166
#> 368 8317147     0.00923664
#> 369 8317165     0.05262387
#> 370 8317171     0.06221993
#> 371 8317177     0.13989755
#> 372 8317179     0.10991569
#> 373 8317185     0.03500884
#> 374 8317151     0.19142262
#> 375 8317111     0.00001000
#> 376 8317367     0.00001000
#> 377 8315731     0.07660523
#> 378 8315733     0.10049842
#> 379 8315721     0.03648606
#> 380 8315727     0.09978275
#> 381 8320107     0.14281679
#> 382 8320089     0.44804597
#> 383 8320123     0.09437927
#> 384 8320143     0.03071055
#> 385 8320127     0.05369242
#> 386 8320149     0.08403304
#> 387 8320151     0.03285960
#> 388 8320167     0.02513606
#> 389 8320173     0.16860896
#> 390 8320181     0.02650323
#> 391 8320189     0.01783430
#> 392 8320191     0.03123844
#> 393 8320197     0.04718559
#> 394 8320201     0.12490909
#> 395 8320203     0.14563950
#> 396 8315661     0.04184364
#> 397 8320185     0.17475711
#> 398 8320183     0.20333333
#> 399 8320161     0.11513745
#> 400 8315729     0.41600000
#> 401 8317537             NA
#> 402 8323709     0.00001000
#> 403 8317445     0.00001000
#> 404 8317541     0.00001000
#> 405 8317543     0.00001000
#> 406 8317503             NA
#> 407 8317489             NA
#> 408 8317491             NA
#> 409 8317501             NA
#> 410 8317531             NA
#> 411 8317529             NA
#> 412 8317493             NA
#> 413 8317527             NA
#> 414 8317515             NA
#> 415 8317499             NA
#> 416 8317449             NA
#> 417 8317495             NA
#> 418 8317525             NA
#> 419 8317417             NA
#> 420 8317461             NA
#> 421 8317479             NA
#> 422 8317487             NA
#> 423 8317475             NA
#> 424 8317375             NA
#> 425 8317459             NA
#> 426 8317477             NA
#> 427 8317481             NA
#> 428 8317485             NA
#> 429 8317457             NA
#> 430 8317523             NA
#> 431 8317373             NA
#> 432 8318839     0.00001000
#> 433 8319929     0.00001000
#> 434 8318843     0.00001000
#> 435 8315863             NA
#> 436 8315865             NA
#> 437 8315905             NA
#> 438 8317949     0.13365180
#> 439 8317559     0.00371428
#> 440 8316799     0.04582235
#> 441 8318061     0.00001000
#> 442 8317335     0.08781936
#> 443 8317401     0.00001000
#> 444 8323705     0.00001000
#> 445 8317399     0.00001000
#> 446 8323707     0.00001000
#> 447 8323703     0.00001000
#> 448 8317387     0.00001000
#> 449 8317393     0.00001000
#> 450 8323695     0.00001000
#> 451 8317389     0.00001000
#> 452 8323699     0.00001000
#> 453 8316855     0.00001000
#> 454 8323691     0.00001000
#> 455 8316901     0.00010000
#> 456 8317539     0.00001000
#> 457 8317133     0.00188337
#> 458 8317563     0.00001000
#> 459 8316825     0.00108108
#> 460 8317379     0.00084033
#> 461 8317555     0.00001000
#> 462 8317557     0.00110465
#> 463 8321257     0.00406876
#> 464 8323697     0.00013785
#> 465 8321285     0.00001000
#> 466 8321283     0.01805555
#> 467 8321275     0.00001000
#> 468 8321273     0.00001253
#> 469 8321271     0.00921428
#> 470 8321269     0.02117647
#> 471 8321267     0.02113300
#> 472 8321265     0.00001000
#> 473 8321263     0.00346613
#> 474 8321261     0.00001000
#> 475 8321259     0.00008159
#> 476 8315853     0.00041841
#> 477 8315851     0.00001000
#> 478 8315849     0.00050308
#> 479 8315855     0.00001000
#> 480 8315859     0.00066048
#> 481 8315857     0.00001000
#> 482 8303370     0.00001000
#> 483 8303376     0.00664140
#> 484 8303378     0.00001000
#> 485 8303384     0.00001000
#> 486 8303392     0.00904692
#> 487 8303396     0.00001000
#> 488 8303406     0.00835820
#> 489 8303430     0.00597294
#> 490 8303440     0.00001000
#> 491 8303446     0.00152882
#> 492 8304602     0.00001000
#> 493 8303480     0.00001000
#> 494 8303484     0.00602992
#> 495 8303518     0.00164301
#> 496 8303528     0.00954198
#> 497 8303552     0.00076379
#> 498 8303562     0.01009615
#> 499 8303566     0.00846846
#> 500 8303574     0.00685503
#>  [ reached 'max' / getOption("max.print") -- omitted 2690839 rows ]
get_vaa(c("slope", "lengthkm"))
#>       comid          slope lengthkm
#> 1   8318793     0.06083397    1.295
#> 2   8318787     0.09534391    1.323
#> 3   8318775     0.11273792    2.883
#> 4   8318785     0.15142342    2.370
#> 5   8318789     0.16246142    1.511
#> 6   8318801     0.10504926    0.406
#> 7   8318819     0.15627586    2.035
#> 8   8318813     0.14838480    2.546
#> 9   8318825     0.17851902    2.331
#> 10  8318841     0.19833333    0.156
#> 11  8321213     0.30318281    1.058
#> 12  8319971     0.17701588    2.165
#> 13  8319977     0.15134239    1.991
#> 14  8319983     0.05639584    3.425
#> 15  8319985     0.22391050    1.693
#> 16  8319989     0.22171945    2.581
#> 17  8319987     0.26929955    1.492
#> 18  8321221     0.25149134    1.502
#> 19  8319995     0.21252046    3.570
#> 20  8320027     0.02283072    1.406
#> 21  8320029     0.18152041    2.452
#> 22  8320035     0.15245065    2.834
#> 23  8320037     0.10530418    1.727
#> 24  8320049     0.01863346    1.727
#> 25  8321227     0.09343534    2.026
#> 26  8320067     0.18970149    0.284
#> 27  8320051     0.03570175    0.342
#> 28  8320045     0.17278583    1.873
#> 29  8320047     0.02871794    0.267
#> 30  8320071     0.08506246    3.351
#> 31  8320075     0.01661354    1.506
#> 32  8320081     0.09785286    1.984
#> 33  8320077     0.09563380    0.221
#> 34  8320091     0.00654941    2.388
#> 35  8320093     0.06578509    2.671
#> 36  8320121     0.06613976    2.826
#> 37  8320117     0.03421798    1.775
#> 38  8320129     0.01167101    2.298
#> 39  8320131     0.09507233    2.363
#> 40  8320157     0.04404825    2.015
#> 41  8321247     0.06464353    3.320
#> 42  8320159     0.04644230    2.334
#> 43  8320171     0.01326405    2.935
#> 44  8320177     0.01551326    0.867
#> 45  8321253     0.08487241    2.228
#> 46  8321251     0.09899641    1.823
#> 47  8320187     0.11967595    1.909
#> 48  8320195     0.08029826    3.837
#> 49  8320175     0.13833488    2.301
#> 50  8320165     0.04617313    3.940
#> 51  8320153     0.02977542    1.752
#> 52  8320137     0.08623712    3.937
#> 53  8320113     0.03007255    1.654
#> 54  8320097     0.07578203    3.122
#> 55  8320115     0.11378725    2.096
#> 56  8321243     0.08227124    2.599
#> 57  8321239     0.07812816    1.929
#> 58  8321235     0.20535179    2.203
#> 59  8320061     0.13411111    0.240
#> 60  8320059     0.10721804    0.283
#> 61  8321241     0.14880311    4.386
#> 62  8320083     0.08166377    2.451
#> 63  8320063     0.11007383    2.046
#> 64  8321231     0.16020148    3.376
#> 65  8320039     0.11612556    2.379
#> 66  8321225     0.24164087    3.575
#> 67  8320021     0.00001000    0.445
#> 68  8320011     0.20397430    2.562
#> 69  8321223     0.22219711    2.849
#> 70  8320017     0.15138947    3.000
#> 71  8321281     0.00813725    0.102
#> 72  8319997     0.02553872    0.594
#> 73  8320015     0.41857142    0.269
#> 74  8320009     0.16588073    2.549
#> 75  8320005     0.14481554    1.669
#> 76  8319991     0.02752808    0.267
#> 77  8319999     0.11584666    1.649
#> 78  8319993     0.00001000    0.153
#> 79  8319973     0.23622572    1.797
#> 80  8319967     0.23350910    1.743
#> 81  8319961     0.16125262    2.529
#> 82  8319957     0.18179506    2.746
#> 83  8319953     0.20880579    2.704
#> 84  8319949     0.23240081    2.117
#> 85  8321217     0.20355245    3.876
#> 86  8319943     0.19102425    1.262
#> 87  8321215     0.23862620    2.837
#> 88  8319937     0.23577581    1.845
#> 89  8319931     0.27705882    0.204
#> 90  8319933     0.22026553    1.883
#> 91  8318847     0.24131455    0.789
#> 92  8319927     0.05524271    0.103
#> 93  8318845     0.23947907    2.396
#> 94  8318831     0.17320279    0.715
#> 95  8318823     0.00001000    0.138
#> 96  8318821     0.25119738    1.219
#> 97  8318829     0.19621969    2.790
#> 98  8318827     0.22418478    2.358
#> 99  8318815     0.26289540    1.719
#> 100 8318805     0.29542769    1.775
#> 101 8318803     0.24739652    1.817
#> 102 8318807     0.22145433    1.869
#> 103 8318811     0.21544303    2.046
#> 104 8318777     0.05770716    1.605
#> 105 8318761     0.12000000    0.059
#> 106 8318759     0.23509090    2.735
#> 107 8318765     0.20424190    0.737
#> 108 8318763     0.16822938    2.389
#> 109 8318757     0.29032932    1.729
#> 110 8318731     0.17643985    2.278
#> 111 8318721     0.35101086    1.070
#> 112 8318717     0.24808504    3.420
#> 113 8318699     0.26899950    2.170
#> 114 8318679     0.30396946    1.591
#> 115 8318663     0.25616666    1.891
#> 116 8318647     0.25100595    1.831
#> 117 8318615     0.22120736    1.466
#> 118 8318611     0.26263636    0.481
#> 119 8318549     0.25105691    0.273
#> 120 8318565     0.19246428    1.680
#> 121 8318637     0.10202711    1.699
#> 122 8318651     0.16245378    0.746
#> 123 8318671     0.06377579    3.925
#> 124 8318683     0.14034525    1.772
#> 125 8318703     0.17897260    1.902
#> 126 8318751     0.27488594    1.509
#> 127 8318745     0.17518259    3.437
#> 128 8318723     0.20423864    3.032
#> 129 8318711     0.17488664    1.738
#> 130 8318685     0.20989715    2.678
#> 131 8318673     0.17709493    4.175
#> 132 8318667     0.26097544    1.617
#> 133 8318635     0.17557298    3.912
#> 134 8318591     0.18856655    0.293
#> 135 8318621     0.18194473    2.068
#> 136 8318583     0.24591463    1.627
#> 137 8318623     0.20157046    3.130
#> 138 8318587     0.24492943    2.063
#> 139 8318551     0.24719232    2.806
#> 140 8318543     0.21510492    2.335
#> 141 8318515     0.26665768    1.634
#> 142 8318517     0.21239255    2.245
#> 143 8318533     0.24063420    0.883
#> 144 8318525     0.19292817    2.322
#> 145 8318523     0.18834862    2.766
#> 146 8318571     0.16607043    1.939
#> 147 8318531     0.18519337    0.694
#> 148 8318567     0.27565196    2.109
#> 149 8318597     0.18999166    2.400
#> 150 8318541     0.22647406    1.674
#> 151 8318579     0.16452030    2.933
#> 152 8318561     0.16763896    3.857
#> 153 8318537     0.21979455    2.388
#> 154 8318519     0.20262769    3.113
#> 155 8318509     0.11969230    0.475
#> 156 8318507     0.17174890    3.351
#> 157 8318489     0.18540173    2.427
#> 158 8318481     0.19705542    1.016
#> 159 8318479     0.15183537    7.185
#> 160 8318469     0.14901062    1.506
#> 161 8318457     0.17986104    2.015
#> 162 8318453     0.24622478    0.844
#> 163 8318459     0.20149580    1.700
#> 164 8318447     0.17287345    3.390
#> 165 8318419     0.20900488    1.638
#> 166 8318417     0.26889245    0.773
#> 167 8318397     0.00671256    0.581
#> 168 8318387     0.00661290    0.459
#> 169 8318373     0.17038401    2.076
#> 170 8319345     0.14720328    2.467
#> 171 8318441     0.06377997    1.569
#> 172 8318429     0.22560439    1.547
#> 173 8318461     0.12258526    3.639
#> 174 8318439     0.13163910    1.391
#> 175 8318451     0.19980730    1.655
#> 176 8318463     0.27626911    0.476
#> 177 8318455     0.19623214    1.680
#> 178 8318363     0.21902129    2.827
#> 179 8318339     0.17626815    4.006
#> 180 8318315     0.09532915    2.382
#> 181 8318303     0.11242044    4.305
#> 182 8318265     0.18525849    1.504
#> 183 8318263     0.15567434    1.216
#> 184 8318321     0.13117056    2.842
#> 185 8318299     0.16531281    3.172
#> 186 8318259     0.16725624    3.073
#> 187 8318241     0.21280571    1.899
#> 188 8318233     0.17521776    2.491
#> 189 8318239     0.18268981    2.309
#> 190 8318247     0.06592686    1.737
#> 191 8318235     0.09654333    2.123
#> 192 8318209     0.08230380    4.699
#> 193 8318155     0.18245810    0.716
#> 194 8318163     0.14550351    1.431
#> 195 8318187     0.10515584    3.584
#> 196 8318151     0.06622287    2.719
#> 197 8318089     0.18605161    0.775
#> 198 8318067     0.20757281    1.077
#> 199 8318117     0.23455231    2.005
#> 200 8318139     0.18482514    2.837
#> 201 8318157     0.17295377    2.206
#> 202 8318173     0.00001000    0.209
#> 203 8318059     0.00044067    0.295
#> 204 8316847     0.00497515    1.610
#> 205 8316841     0.00359223    0.103
#> 206 8316833     0.09120678    1.682
#> 207 8317993     0.13712318    1.457
#> 208 8318001     0.10702258    0.487
#> 209 8316839     0.10149700    0.835
#> 210 8316835     0.01319985    2.846
#> 211 8316783     0.01076880    1.795
#> 212 8317923     0.15903312    1.117
#> 213 8317929     0.09793542    1.326
#> 214 8316773     0.13117192    1.681
#> 215 8316771     0.12152348    1.639
#> 216 8316765     0.01295918    1.176
#> 217 8316757     0.09446746    3.238
#> 218 8316761     0.11045407    3.057
#> 219 8316753     0.02096062    4.364
#> 220 8316777     0.02803252    3.902
#> 221 8316829     0.00001000    0.387
#> 222 8316817     0.01330336    2.113
#> 223 8316795     0.02692786    1.756
#> 224 8316797     0.02408631    1.634
#> 225 8316813     0.00453125    0.214
#> 226 8316803     0.00016666    0.270
#> 227 8316793     0.01661529    1.902
#> 228 8316779     0.03502164    1.767
#> 229 8317415     0.00001000    1.139
#> 230 8317403     0.00001000    0.129
#> 231 8317409     0.00109452    0.553
#> 232 8323713     0.00001000    0.067
#> 233 8323701     0.00001000    0.516
#> 234 8317397     0.00001000    0.485
#> 235 8323681     0.00001000    0.076
#> 236 8317385     0.00001000    0.280
#> 237 8317547     0.00001000    0.167
#> 238 8323689     0.00001000    0.050
#> 239 8317411     0.00001000    0.220
#> 240 8316869     0.00054676    0.695
#> 241 8317405     0.00084507    0.071
#> 242 8316827     0.00001000    2.754
#> 243 8317549     0.00001000    0.082
#> 244 8317381     0.00001000    0.085
#> 245 8323687     0.00001000    0.082
#> 246 8323685     0.00001000    0.041
#> 247 8323711     0.00001000    0.367
#> 248 8317551 -9998.00000000    0.132
#> 249 8316843     0.00001000    0.191
#> 250 8317545 -9998.00000000    0.050
#> 251 8317383     0.00001000    0.047
#> 252 8316851     0.00001104    3.769
#> 253 8317391     0.00057803    0.173
#> 254 8316865     0.00104590    5.553
#> 255 8317407     0.00001000    0.175
#> 256 8316873     0.07239559    3.694
#> 257 8316875     0.00203339    5.690
#> 258 8317395     0.00559322    0.236
#> 259 8316895     0.10462288    2.816
#> 260 8318355     0.12495589    2.644
#> 261 8318319     0.05431400    2.928
#> 262 8318279     0.16676411    1.505
#> 263 8318267     0.13028862    0.740
#> 264 8316891     0.17284588    2.284
#> 265 8318223     0.10440000    1.076
#> 266 8316879     0.03170596    4.305
#> 267 8316893     0.00014373    3.907
#> 268 8317437     0.00001000    0.109
#> 269 8316907     0.00066945    2.540
#> 270 8316899     0.00001000    0.151
#> 271 8316897     0.00001277    1.716
#> 272 8316903     0.05268866    4.271
#> 273 8316921     0.00083076    1.625
#> 274 8316927     0.00646812    1.914
#> 275 8317361     0.06109024    2.598
#> 276 8316961     0.00221788    4.360
#> 277 8317355     0.02072972    0.740
#> 278 8316975     0.00001000    0.379
#> 279 8317015     0.00188913    2.742
#> 280 8317019     0.00808721    2.018
#> 281 8317007     0.07347537    1.958
#> 282 8317009     0.05994123    1.021
#> 283 8317353     0.02425149    2.489
#> 284 8317027     0.00187997    1.633
#> 285 8317021     0.03714487    1.565
#> 286 8317023     0.02752427    0.206
#> 287 8317037     0.03582191    1.609
#> 288 8317025     0.02556616    1.617
#> 289 8317031     0.04136929    2.320
#> 290 8316979     0.02415868    1.462
#> 291 8316983     0.05576820    2.033
#> 292 8316973     0.00879569    0.465
#> 293 8317357     0.00028106    0.676
#> 294 8317359     0.04341903    3.144
#> 295 8316977     0.06900000    2.030
#> 296 8316965     0.01019472    1.741
#> 297 8316959     0.03993899    2.281
#> 298 8316931     0.03037569    3.556
#> 299 8316995     0.05375880    2.422
#> 300 8317013     0.06520454    1.909
#> 301 8317029     0.13998334    2.551
#> 302 8317017     0.13808497    1.726
#> 303 8318407     0.21706255    1.103
#> 304 8318411     0.22624365    0.741
#> 305 8316971     0.11520249    4.001
#> 306 8316951     0.02379867    1.053
#> 307 8316933     0.08393221    5.342
#> 308 8316943     0.13554512    2.922
#> 309 8316937     0.02089053    2.156
#> 310 8316923     0.08375663    2.410
#> 311 8316925     0.07428844    1.198
#> 312 8316917     0.00505456    2.166
#> 313 8317435     0.00012422    0.161
#> 314 8316919     0.00619329    2.685
#> 315 8317431     0.00037037    0.135
#> 316 8316905     0.00497085    2.038
#> 317 8317427     0.00001000    0.128
#> 318 8317055     0.06883307    2.122
#> 319 8317061     0.07491298    1.644
#> 320 8317109     0.06566137    3.552
#> 321 8317065     0.02263492    0.945
#> 322 8317113     0.06500217    2.450
#> 323 8317101     0.05054779    1.939
#> 324 8317107     0.02146710    1.520
#> 325 8317123     0.07224127    1.842
#> 326 8317127     0.01984421    1.497
#> 327 8317139     0.05674228    3.324
#> 328 8317143     0.09135764    2.276
#> 329 8317173     0.03607496    2.070
#> 330 8317149     0.10958143    1.314
#> 331 8317163     0.07931756    3.829
#> 332 8317175     0.12950185    2.038
#> 333 8317141     0.09573283    2.847
#>  [ reached 'max' / getOption("max.print") -- omitted 2691006 rows ]

#cleanup if desired
unlink(dirname(get_vaa_path()), recursive = TRUE)
# }