Get Cross Section From Point (experimental)

Uses a cross section retrieval web services to retrieve a cross section given a point and specified width. Orientation is determined based on direction of a the flowline found near point. This function uses a 10m National Elevation Dataset request on the back end.

get_xs_point(point, width, num_pts)

Arguments

point

sfc POINT including crs as created by: sf::st_sfc(sf::st_point(.. ,..), crs)crs.

width

Cross section width in meters.

num_pts

numeric number of points to retrieve along the cross section.

Value

sf data.frame containing points retrieved.

Examples

# \donttest{
point <- sf::st_sfc(sf::st_point(x = c(-105.97218, 36.17592)), crs = 4326)

(xs <- get_xs_point(point, 300, 100))
#> Request failed [504]. Retrying in 2 seconds...
#> Request failed [504]. Retrying in 1.4 seconds...
#> Error calling processing service. 
#>  Original error: 
#> Error in doTryCatch(return(expr), name, parentenv, handler): <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
#> <HTML><HEAD><META HTTP-EQUIV="Content-Type" CONTENT="text/html; charset=iso-8859-1">
#> <TITLE>ERROR: The request could not be satisfied</TITLE>
#> </HEAD><BODY>
#> <H1>504 ERROR</H1>
#> <H2>The request could not be satisfied.</H2>
#> <HR noshade size="1px">
#> CloudFront attempted to establish a connection with the origin, but either the attempt failed or the origin closed the connection.
#> We can't connect to the server for this app or website at this time. There might be too much traffic or a configuration error. Try again later, or contact the app or website owner.
#> <BR clear="all">
#> If you provide content to customers through CloudFront, you can find steps to troubleshoot and help prevent this error by reviewing the CloudFront documentation.
#> <BR clear="all">
#> <HR noshade size="1px">
#> <PRE>
#> Generated by cloudfront (CloudFront)
#> Request ID: MXuK8PTGGxRwOKKUHGGDwl8O-3ucEmTEsgLlfj33CPiZI8w8K0iIiA==
#> </PRE>
#> <ADDRESS>
#> </ADDRESS>
#> </BODY></HTML>
#> NULL

if(inherits(xs, "sf")) {

bbox <- sf::st_bbox(xs) + c(-0.005, -0.005, 0.005, 0.005)

nhdplusTools::plot_nhdplus(bbox = bbox, cache_data = FALSE)

plot(sf::st_transform(sf::st_geometry(xs), 3857), pch = ".", add = TRUE, col = "red")
plot(sf::st_transform(sf::st_sfc(point, crs = 4326), 3857), add = TRUE)

plot(xs$distance_m, xs$elevation_m)
}
# }