Prep NHDPlus Data

Function to prep NHDPlus data for use by nhdplusTools functions

prepare_nhdplus(
  flines,
  min_network_size = 0,
  min_path_length = 0,
  min_path_size = 0,
  purge_non_dendritic = TRUE,
  warn = TRUE,
  error = TRUE,
  skip_toCOMID = FALSE,
  align_names = TRUE
)

Arguments

flines

data.frame NHDPlus flowlines including: COMID, LENGTHKM, FTYPE (or FCODE), TerminalFl, FromNode, ToNode, TotDASqKM, StartFlag, StreamOrde, StreamCalc, TerminalPa, Pathlength, and Divergence variables.

min_network_size

numeric Minimum size (sqkm) of drainage network to include in output.

min_path_length

numeric Minimum length (km) of terminal level path of a network.

min_path_size

numeric Minimum size (sqkm) of outlet level path of a drainage basin. Drainage basins with an outlet drainage area smaller than this will be removed.

purge_non_dendritic

logical Should non dendritic paths be removed or not.

warn

logical controls whether warning an status messages are printed

error

logical controls whether to return potentially invalid data with a warning rather than an error

skip_toCOMID

logical if TRUE, toCOMID will not be added to output.

align_names

logical

Value

data.frame ready to be used with the refactor_flowlines function.

Examples

source(system.file("extdata", "sample_flines.R", package = "nhdplusTools"))

prepare_nhdplus(sample_flines,
                min_network_size = 10,
                min_path_length = 1,
                warn = FALSE)
#>        COMID  toCOMID LENGTHKM  Hydroseq LevelPathI TotDASqKM
#> 1   11689204 11689212    0.730 200027394  200024046  149.9148
#> 2   11689212 11689228    0.819 200026423  200024046  164.6676
#> 3   11689228 11690186    2.053 200025563  200024046  173.2878
#> 4   11690186 11694048    2.912 200024768  200024046  182.1492
#> 5   11694048 11690184    0.745 200024046  200024046  182.6982
#> 6   11689202 11689204    4.765 200028471  200024046  136.7730
#> 7   11689720 11689202    4.403 200029710  200024046  108.8460
#> 8   11689108 11689720    1.260 200031091  200024046   90.3771
#> 9   11689106 11689108    3.510 200032655  200024046   88.4268
#> 10  11689700 11689106    2.350 200034479  200024046   82.7694
#> 11  11689684 11689700    4.937 200036652  200024046   67.2804
#> 12  11688884 11689684    1.286 200039295  200024046   50.2722
#> 13  11688826 11688884    1.265 200042546  200024046   47.9880
#> 14  11688810 11688826    0.125 200046674  200024046   46.2582
#> 15  11689644 11688810    4.605 200059988  200024046   26.9019
#> 16  11688658 11689644    1.115 200072135  200024046   13.0149
#> 17  11688596 11688658    5.362 200095585  200024046   10.6101
#> 18  11690260        0    0.075 200009667  200009667 1601.1765
#> 19  11690262 11690260    0.457 200009721  200009667 1601.1693
#> 20  11690258 11690568    0.687 200009832  200009667 1574.3007
#> 21  11690568 11690262    0.159 200009774  200009667 1574.3907
#> 22  11690256 11690258    0.098 200009884  200009667 1573.6914
#> 23  11690570 11690256    1.981 200009944  200009667 1568.3292
#> 24  11690564 11690570    1.076 200010000  200009667 1549.5903
#> 25  11690246 11690538    1.057 200010132  200009667 1529.7570
#> 26  11690538 11690564    0.623 200010065  200009667 1531.3284
#> 27  11690566 11690602    0.046 200010257  200009667 1503.2439
#> 28  11690602 11690246    0.338 200010194  200009667 1503.4230
#> 29  11690604 11690566    0.275 200010324  200009667 1498.7727
#> 30  11690558 11690604    0.821 200010393  200009667 1492.9407
#> 31  11690556 11690558    0.306 200010464  200009667 1490.2362
#> 32  11690226 11690556    0.761 200010537  200009667 1489.7934
#> 33  11690560 11690226    0.303 200010609  200009667 1485.8496
#> 34  11690206 11690560    2.821 200010685  200009667 1451.9799
#> 35  11690194 11690198    0.317 200010837  200009667 1325.4678
#> 36  11690198 11690206    2.124 200010761  200009667 1344.0105
#> 37  11690190 11690194    2.280 200010914  200009667 1318.1364
#> 38  11690184 11690190    2.870 200010994  200009667 1142.9100
#> 39  11689258 11694052    0.590 200011252  200009667  944.9766
#> 40  11694052 11690196    1.200 200011157  200009667  948.1437
#> 41  11690196 11690184    0.837 200011074  200009667  952.8912
#> 42  11689276 11689258    1.106 200011445  200009667  938.2329
#> 43  11690202 11689276    0.170 200011542  200009667  936.5724
#> 44  11689286 11690202    0.476 200011645  200009667  936.2628
#> 45  11689288 11689286    0.497 200011747  200009667  934.8255
#> 46  11689280 11689288    2.361 200011856  200009667  930.5109
#> 47  11689294 11689280    1.413 200012576  200009667  918.3285
#> 48  11689314 11689294    0.359 200012703  200009667  916.1658
#> 49  11689316 11689314    0.463 200012842  200009667  914.6331
#> 50  11689312 11689316    0.418 200012980  200009667  914.0283
#> 51  11689322 11689312    0.711 200013127  200009667  895.9041
#> 52  11689328 11689322    0.297 200013281  200009667  893.5173
#> 53  11689332 11689328    0.158 200013432  200009667  868.3857
#> 54  11689334 11689332    1.472 200013595  200009667  817.3647
#> 55  11689302 11689334    1.211 200013758  200009667  814.1445
#> 56  11689274 11689302    0.040 200013927  200009667  807.8715
#> 57  11689272 11689274    2.305 200014093  200009667  798.1092
#> 58  11689750 11689272    0.962 200014256  200009667  792.9927
#> 59  11689744 11689750    0.519 200014422  200009667  790.1370
#> 60  11689254 11689744    1.406 200014591  200009667  785.3616
#> 61  11689730 11689254    0.808 200014772  200009667  779.1759
#> 62  11689246 11689730    2.395 200014965  200009667  775.9791
#> 63  11689224 11689246    0.433 200015168  200009667  770.0661
#> 64  11689214 11689224    0.701 200015381  200009667  760.3362
#> 65  11689208 11689214    2.987 200015600  200009667  753.0939
#> 66  11689192 11689208    0.354 200015832  200009667  746.0865
#> 67  11689718 11689192    2.922 200016057  200009667  741.6369
#> 68  11689150 11689718    3.386 200016305  200009667  719.3691
#> 69  11689152 11689150    1.663 200016553  200009667  689.2317
#> 70  11689146 11689152    1.092 200016805  200009667  685.9872
#> 71  11689126 11689146    0.734 200017080  200009667  683.7552
#> 72  11689128 11689126    0.645 200017374  200009667  680.3271
#> 73  11689116 11689128    2.207 200017682  200009667  674.8614
#> 74  11689080 11689116    0.292 200017995  200009667  669.5937
#> 75  11689072 11689080    2.400 200018319  200009667  665.8191
#> 76  11689022 11689072    1.903 200018662  200009667  441.0522
#> 77  11688980 11689022    1.295 200019022  200009667  436.9212
#> 78  11688950 11688980    2.290 200019387  200009667  427.8456
#> 79  11690532 11689928    0.987 200028466  200009667  146.9997
#> 80  11687548 11690532    1.278 200029716  200009667  145.6047
#> 81  11687482 11687548    2.598 200031094  200009667  136.3095
#> 82  11687424 11687482    0.483 200032658  200009667  128.9016
#> 83  11687402 11687424    2.173 200034492  200009667  110.4228
#> 84  11687358 11687402    6.800 200039306  200039306   54.2259
#> 85  11687234 11687358    0.693 200042565  200039306   43.6887
#> 86  11687224 11687234    3.508 200046693  200039306   27.4752
#> 87  11687180 11687224    0.938 200052180  200039306   18.4014
#> 88  11687150 11687180    2.762 200095793  200039306    7.2540
#> 89  11687356 11687402    0.468 200036663  200009667   53.3169
#> 90  11687334 11687356    3.382 200039307  200009667   50.9724
#> 91  11687294 11687334    3.284 200042563  200009667   45.8991
#> 92  11687252 11687294    0.687 200046692  200009667   33.5745
#> 93  11687244 11687252    0.529 200052179  200009667   30.8808
#> 94  11687232 11687244    1.703 200060053  200009667   27.7362
#> 95  11687210 11687232    0.628 200072193  200009667   16.4367
#> 96  11687202 11687210    4.850 200095778  200009667   11.1987
#> 97  11688350 11689972    0.961 200034486  200031089   37.7667
#> 98  11687542 11688310    0.884 200039303  200031089   32.4315
#> 99  11688310 11688350    1.852 200036657  200031089   33.6465
#> 100 11687540 11687542    0.589 200042558  200031089   29.1267
#> 101 11693988 11687540    0.126 200046644  200031089   28.6974
#> 102 11687538 11693988    3.231 200052172  200031089   28.3536
#> 103 11687458 11687538    6.156 200095713  200095713   10.4076
#> 104 11688556 11690050    0.450 200024052  200021629   75.9474
#> 105 11688538 11688556    0.748 200024774  200021629   72.7992
#> 106 11689606 11688538    1.941 200025562  200021629   70.5762
#> 107 11688438 11689606    0.381 200026425  200021629   65.6037
#> 108 11688432 11688438    1.359 200027397  200021629   62.0298
#> 109 11688414 11688432    0.127 200028472  200021629   58.9590
#> 110 11688408 11688414    1.578 200029712  200021629   49.5423
#> 111 11691290 11688408    0.979 200031083  200021629   41.5251
#> 112 11691288 11691290    0.966 200032647  200021629   39.2166
#> 113 11691284 11691288    1.313 200034475  200021629   36.9864
#> 114 11691276 11691284    0.679 200036645  200021629   22.7556
#> 115 11691272 11691276    0.128 200039283  200021629   14.5359
#> 116 11687878 11691272    0.458 200042554  200021629   14.5062
#> 117 11687866 11687878    0.262 200046685  200046685    9.9090
#> 118 11687864 11687866    1.294 200052168  200046685    8.6013
#> 119 11687858 11687864    1.047 200060033  200046685    6.4377
#> 120 11687854 11687858    0.924 200072159  200046685    3.6450
#> 121 11687850 11687854    1.334 200095664  200046685    2.1951
#> 122 11689016 11689072    1.780 200019782  200019782  222.3801
#> 123 11689006 11689016    0.276 200020206  200019782  173.2329
#> 124 11689020 11689006    2.239 200020653  200019782  167.7024
#> 125 11689010 11689020    1.714 200021129  200019782  159.9903
#> 126 11688982 11689010    2.329 200021631  200019782  154.3797
#> 127 11688972 11688982    0.608 200022162  200019782  149.5836
#> 128 11688974 11688972    1.711 200022742  200019782  140.6313
#> 129 11688958 11688974    0.718 200023372  200019782  137.2419
#> 130 11688956 11688958    0.225 200024051  200019782  135.9270
#> 131 11688960 11688956    0.721 200024772  200019782  134.5428
#> 132 11688966 11688960    0.662 200025564  200019782  131.0490
#> 133 11688984 11688966    0.314 200026424  200019782  109.7325
#> 134 11691408 11688986    2.359 200028465  200019782  101.7225
#> 135 11688986 11688984    0.157 200027395  200019782  101.7423
#> 136 11691406 11691408    1.014 200029706  200019782   95.5575
#> 137 11691402 11691406    1.371 200031082  200019782   92.5443
#> 138 11691396 11691402    1.800 200032646  200019782   85.5315
#> 139 11691388 11691396    0.183 200034474  200019782   79.8570
#> 140 11691382 11691388    4.324 200036644  200036644   49.9311
#> 141 11691358 11691382    0.044 200039281  200036644   29.7270
#> 142 11691356 11691358    2.659 200042525  200036644   27.2889
#> 143 11691336 11691356    1.009 200046645  200036644   20.3661
#> 144 11691330 11691336    0.571 200052114  200036644   15.5709
#> 145 11691326 11691330    1.278 200059945  200036644   13.4334
#> 146 11691318 11691326    1.777 200072004  200036644    7.4421
#> 147 11691304 11691318    1.887 200095295  200036644    1.6920
#> 148 11687226 11687234    7.629 200095771  200095771   14.8536
#> 149 11689148 11689718    0.414 200046669  200046669   18.2178
#> 150 11689124 11689148    0.554 200052146  200046669   14.1894
#> 151 11689110 11689124    1.600 200060005  200046669   12.5532
#> 152 11689076 11689110    0.723 200072108  200046669    9.5931
#> 153 11689066 11689076    3.320 200095475  200046669    6.8427
#> 154 11689122 11689150    1.229 200046670  200046670   25.0758
#> 155 11689090 11689122    3.152 200060007  200046670   17.8560
#> 156 11689008 11689090    2.599 200095488  200046670    4.7520
#> 157 11687638 11687232    5.573 200095676  200095676    9.3249
#> 158 11687400 11687424    5.093 200052175  200052175   18.2511
#> 159 11687292 11687400    4.767 200060050  200052175   10.2411
#> 160 11687206 11687292    0.711 200095777  200052175    1.3374
#> 161 11691384 11691388    0.963 200052112  200019782   29.8800
#> 162 11691400 11691384    2.820 200071998  200071998    9.2556
#> 163 11691418 11691400    3.243 200095268  200071998    3.9951
#> 164 11688406 11688414    2.986 200060027  200060027    9.4068
#> 165 11687524 11690534    2.131 200095697  200060027    4.0419
#> 166 11690534 11688406    0.273 200072009  200060027    4.0959
#>  [ reached 'max' / getOption("max.print") -- omitted 531 rows ]