Prep NHDPlus Data

Function to prep NHDPlus data for use by nhdplusTools functions

prepare_nhdplus(
  flines,
  min_network_size,
  min_path_length,
  min_path_size = 0,
  purge_non_dendritic = TRUE,
  warn = TRUE,
  error = TRUE,
  skip_toCOMID = FALSE
)

Arguments

flines
data.frame NHDPlus flowlines including: COMID, LENGTHKM, FTYPE (or FCODE), TerminalFl, FromNode, ToNode, TotDASqKM, StartFlag, StreamOrde, StreamCalc, TerminalPa, Pathlength, and Divergence variables.

min_network_size
numeric Minimum size (sqkm) of drainage network to include in output.

min_path_length
numeric Minimum length (km) of terminal level path of a network.

min_path_size
numeric Minimum size (sqkm) of outlet level path of a drainage basin. Drainage basins with an outlet drainage area smaller than this will be removed.

purge_non_dendritic
boolean Should non dendritic paths be removed or not.

warn
boolean controls whether warning an status messages are printed

error
boolean controls whether to return potentially invalid data with a warning rather than an error

skip_toCOMID
boolean if TRUE, toCOMID will not be added to output.

Value

data.frame ready to be used with the refactor_flowlines function.

Examples

source(system.file("extdata", "sample_flines.R", package = "nhdplusTools"))

prepare_nhdplus(sample_flines,
                min_network_size = 10,
                min_path_length = 1,
                warn = FALSE)
#> # A tibble: 697 x 6
#>       COMID LENGTHKM  Hydroseq LevelPathI TotDASqKM  toCOMID
#>       <int>    <dbl>     <dbl>      <dbl>     <dbl>    <int>
#>  1 11689204    0.73  200027394  200024046     150.  11689212
#>  2 11689212    0.819 200026423  200024046     165.  11689228
#>  3 11689228    2.05  200025563  200024046     173.  11690186
#>  4 11690186    2.91  200024768  200024046     182.  11694048
#>  5 11694048    0.745 200024046  200024046     183.  11690184
#>  6 11689202    4.76  200028471  200024046     137.  11689204
#>  7 11689720    4.40  200029710  200024046     109.  11689202
#>  8 11689108    1.26  200031091  200024046      90.4 11689720
#>  9 11689106    3.51  200032655  200024046      88.4 11689108
#> 10 11689700    2.35  200034479  200024046      82.8 11689106
#> # ... with 687 more rows