Prep NHDPlus Data

Function to prep NHDPlus data for use by nhdplusTools functions

prepare_nhdplus(
  flines,
  min_network_size = 0,
  min_path_length = 0,
  min_path_size = 0,
  purge_non_dendritic = TRUE,
  warn = TRUE,
  error = TRUE,
  skip_toCOMID = FALSE,
  align_names = TRUE
)

Arguments

flines

data.frame NHDPlus flowlines including: COMID, LENGTHKM, FTYPE (or FCODE), TerminalFl, FromNode, ToNode, TotDASqKM, StartFlag, StreamOrde, StreamCalc, TerminalPa, Pathlength, and Divergence variables.

min_network_size

numeric Minimum size (sqkm) of drainage network to include in output.

min_path_length

numeric Minimum length (km) of terminal level path of a network.

min_path_size

numeric Minimum size (sqkm) of outlet level path of a drainage basin. Drainage basins with an outlet drainage area smaller than this will be removed.

purge_non_dendritic

logical Should non dendritic paths be removed or not.

warn

logical controls whether warning an status messages are printed

error

logical controls whether to return potentially invalid data with a warning rather than an error

skip_toCOMID

logical if TRUE, toCOMID will not be added to output.

align_names

logical

Value

data.frame ready to be used with the refactor_flowlines function.

Examples

source(system.file("extdata", "sample_flines.R", package = "nhdplusTools"))

prepare_nhdplus(sample_flines,
                min_network_size = 10,
                min_path_length = 1,
                warn = FALSE)
#>         COMID   toCOMID LENGTHKM  Hydroseq LevelPathI TotDASqKM
#> 1    11689204  11689212    0.730 200027394  200024046  149.9148
#> 2    11689212  11689228    0.819 200026423  200024046  164.6676
#> 3    11689228  11690186    2.053 200025563  200024046  173.2878
#> 4    11690186  11694048    2.912 200024768  200024046  182.1492
#> 5    11694048  11690184    0.745 200024046  200024046  182.6982
#> 6    11689202  11689204    4.765 200028471  200024046  136.7730
#> 7    11689720  11689202    4.403 200029710  200024046  108.8460
#> 8    11689108  11689720    1.260 200031091  200024046   90.3771
#> 9    11689106  11689108    3.510 200032655  200024046   88.4268
#> 10   11689700  11689106    2.350 200034479  200024046   82.7694
#> 11   11689684  11689700    4.937 200036652  200024046   67.2804
#> 12   11688884  11689684    1.286 200039295  200024046   50.2722
#> 13   11688826  11688884    1.265 200042546  200024046   47.9880
#> 14   11688810  11688826    0.125 200046674  200024046   46.2582
#> 15   11689644  11688810    4.605 200059988  200024046   26.9019
#> 16   11688658  11689644    1.115 200072135  200024046   13.0149
#> 17   11688596  11688658    5.362 200095585  200024046   10.6101
#> 18   11690260         0    0.075 200009667  200009667 1601.1765
#> 19   11690262  11690260    0.457 200009721  200009667 1601.1693
#> 20   11690258  11690568    0.687 200009832  200009667 1574.3007
#> 21   11690568  11690262    0.159 200009774  200009667 1574.3907
#> 22   11690256  11690258    0.098 200009884  200009667 1573.6914
#> 23   11690570  11690256    1.981 200009944  200009667 1568.3292
#> 24   11690564  11690570    1.076 200010000  200009667 1549.5903
#> 25   11690246  11690538    1.057 200010132  200009667 1529.7570
#> 26   11690538  11690564    0.623 200010065  200009667 1531.3284
#> 27   11690566  11690602    0.046 200010257  200009667 1503.2439
#> 28   11690602  11690246    0.338 200010194  200009667 1503.4230
#> 29   11690604  11690566    0.275 200010324  200009667 1498.7727
#> 30   11690558  11690604    0.821 200010393  200009667 1492.9407
#> 31   11690556  11690558    0.306 200010464  200009667 1490.2362
#> 32   11690226  11690556    0.761 200010537  200009667 1489.7934
#> 33   11690560  11690226    0.303 200010609  200009667 1485.8496
#> 34   11690206  11690560    2.821 200010685  200009667 1451.9799
#> 35   11690194  11690198    0.317 200010837  200009667 1325.4678
#> 36   11690198  11690206    2.124 200010761  200009667 1344.0105
#> 37   11690190  11690194    2.280 200010914  200009667 1318.1364
#> 38   11690184  11690190    2.870 200010994  200009667 1142.9100
#> 39   11689258  11694052    0.590 200011252  200009667  944.9766
#> 40   11694052  11690196    1.200 200011157  200009667  948.1437
#> 41   11690196  11690184    0.837 200011074  200009667  952.8912
#> 42   11689276  11689258    1.106 200011445  200009667  938.2329
#> 43   11690202  11689276    0.170 200011542  200009667  936.5724
#> 44   11689286  11690202    0.476 200011645  200009667  936.2628
#> 45   11689288  11689286    0.497 200011747  200009667  934.8255
#> 46   11689280  11689288    2.361 200011856  200009667  930.5109
#> 47   11689294  11689280    1.413 200012576  200009667  918.3285
#> 48   11689314  11689294    0.359 200012703  200009667  916.1658
#> 49   11689316  11689314    0.463 200012842  200009667  914.6331
#> 50   11689312  11689316    0.418 200012980  200009667  914.0283
#> 51   11689322  11689312    0.711 200013127  200009667  895.9041
#> 52   11689328  11689322    0.297 200013281  200009667  893.5173
#> 53   11689332  11689328    0.158 200013432  200009667  868.3857
#> 54   11689334  11689332    1.472 200013595  200009667  817.3647
#> 55   11689302  11689334    1.211 200013758  200009667  814.1445
#> 56   11689274  11689302    0.040 200013927  200009667  807.8715
#> 57   11689272  11689274    2.305 200014093  200009667  798.1092
#> 58   11689750  11689272    0.962 200014256  200009667  792.9927
#> 59   11689744  11689750    0.519 200014422  200009667  790.1370
#> 60   11689254  11689744    1.406 200014591  200009667  785.3616
#> 61   11689730  11689254    0.808 200014772  200009667  779.1759
#> 62   11689246  11689730    2.395 200014965  200009667  775.9791
#> 63   11689224  11689246    0.433 200015168  200009667  770.0661
#> 64   11689214  11689224    0.701 200015381  200009667  760.3362
#> 65   11689208  11689214    2.987 200015600  200009667  753.0939
#> 66   11689192  11689208    0.354 200015832  200009667  746.0865
#> 67   11689718  11689192    2.922 200016057  200009667  741.6369
#> 68   11689150  11689718    3.386 200016305  200009667  719.3691
#> 69   11689152  11689150    1.663 200016553  200009667  689.2317
#> 70   11689146  11689152    1.092 200016805  200009667  685.9872
#> 71   11689126  11689146    0.734 200017080  200009667  683.7552
#> 72   11689128  11689126    0.645 200017374  200009667  680.3271
#> 73   11689116  11689128    2.207 200017682  200009667  674.8614
#> 74   11689080  11689116    0.292 200017995  200009667  669.5937
#> 75   11689072  11689080    2.400 200018319  200009667  665.8191
#> 76   11689022  11689072    1.903 200018662  200009667  441.0522
#> 77   11688980  11689022    1.295 200019022  200009667  436.9212
#> 78   11688950  11688980    2.290 200019387  200009667  427.8456
#> 79   11690532  11689928    0.987 200028466  200009667  146.9997
#> 80   11687548  11690532    1.278 200029716  200009667  145.6047
#> 81   11687482  11687548    2.598 200031094  200009667  136.3095
#> 82   11687424  11687482    0.483 200032658  200009667  128.9016
#> 83   11687402  11687424    2.173 200034492  200009667  110.4228
#> 84   11687358  11687402    6.800 200039306  200039306   54.2259
#> 85   11687234  11687358    0.693 200042565  200039306   43.6887
#> 86   11687224  11687234    3.508 200046693  200039306   27.4752
#> 87   11687180  11687224    0.938 200052180  200039306   18.4014
#> 88   11687150  11687180    2.762 200095793  200039306    7.2540
#> 89   11687356  11687402    0.468 200036663  200009667   53.3169
#> 90   11687334  11687356    3.382 200039307  200009667   50.9724
#> 91   11687294  11687334    3.284 200042563  200009667   45.8991
#> 92   11687252  11687294    0.687 200046692  200009667   33.5745
#> 93   11687244  11687252    0.529 200052179  200009667   30.8808
#> 94   11687232  11687244    1.703 200060053  200009667   27.7362
#> 95   11687210  11687232    0.628 200072193  200009667   16.4367
#> 96   11687202  11687210    4.850 200095778  200009667   11.1987
#> 97   11688350  11689972    0.961 200034486  200031089   37.7667
#> 98   11687542  11688310    0.884 200039303  200031089   32.4315
#> 99   11688310  11688350    1.852 200036657  200031089   33.6465
#> 100  11687540  11687542    0.589 200042558  200031089   29.1267
#> 101  11693988  11687540    0.126 200046644  200031089   28.6974
#> 102  11687538  11693988    3.231 200052172  200031089   28.3536
#> 103  11687458  11687538    6.156 200095713  200095713   10.4076
#> 104  11688556  11690050    0.450 200024052  200021629   75.9474
#> 105  11688538  11688556    0.748 200024774  200021629   72.7992
#> 106  11689606  11688538    1.941 200025562  200021629   70.5762
#> 107  11688438  11689606    0.381 200026425  200021629   65.6037
#> 108  11688432  11688438    1.359 200027397  200021629   62.0298
#> 109  11688414  11688432    0.127 200028472  200021629   58.9590
#> 110  11688408  11688414    1.578 200029712  200021629   49.5423
#> 111  11691290  11688408    0.979 200031083  200021629   41.5251
#> 112  11691288  11691290    0.966 200032647  200021629   39.2166
#> 113  11691284  11691288    1.313 200034475  200021629   36.9864
#> 114  11691276  11691284    0.679 200036645  200021629   22.7556
#> 115  11691272  11691276    0.128 200039283  200021629   14.5359
#> 116  11687878  11691272    0.458 200042554  200021629   14.5062
#> 117  11687866  11687878    0.262 200046685  200046685    9.9090
#> 118  11687864  11687866    1.294 200052168  200046685    8.6013
#> 119  11687858  11687864    1.047 200060033  200046685    6.4377
#> 120  11687854  11687858    0.924 200072159  200046685    3.6450
#> 121  11687850  11687854    1.334 200095664  200046685    2.1951
#> 122  11689016  11689072    1.780 200019782  200019782  222.3801
#> 123  11689006  11689016    0.276 200020206  200019782  173.2329
#> 124  11689020  11689006    2.239 200020653  200019782  167.7024
#> 125  11689010  11689020    1.714 200021129  200019782  159.9903
#> 126  11688982  11689010    2.329 200021631  200019782  154.3797
#> 127  11688972  11688982    0.608 200022162  200019782  149.5836
#> 128  11688974  11688972    1.711 200022742  200019782  140.6313
#> 129  11688958  11688974    0.718 200023372  200019782  137.2419
#> 130  11688956  11688958    0.225 200024051  200019782  135.9270
#> 131  11688960  11688956    0.721 200024772  200019782  134.5428
#> 132  11688966  11688960    0.662 200025564  200019782  131.0490
#> 133  11688984  11688966    0.314 200026424  200019782  109.7325
#> 134  11691408  11688986    2.359 200028465  200019782  101.7225
#> 135  11688986  11688984    0.157 200027395  200019782  101.7423
#> 136  11691406  11691408    1.014 200029706  200019782   95.5575
#> 137  11691402  11691406    1.371 200031082  200019782   92.5443
#> 138  11691396  11691402    1.800 200032646  200019782   85.5315
#> 139  11691388  11691396    0.183 200034474  200019782   79.8570
#> 140  11691382  11691388    4.324 200036644  200036644   49.9311
#> 141  11691358  11691382    0.044 200039281  200036644   29.7270
#> 142  11691356  11691358    2.659 200042525  200036644   27.2889
#> 143  11691336  11691356    1.009 200046645  200036644   20.3661
#> 144  11691330  11691336    0.571 200052114  200036644   15.5709
#> 145  11691326  11691330    1.278 200059945  200036644   13.4334
#> 146  11691318  11691326    1.777 200072004  200036644    7.4421
#> 147  11691304  11691318    1.887 200095295  200036644    1.6920
#> 148  11687226  11687234    7.629 200095771  200095771   14.8536
#> 149  11689148  11689718    0.414 200046669  200046669   18.2178
#> 150  11689124  11689148    0.554 200052146  200046669   14.1894
#> 151  11689110  11689124    1.600 200060005  200046669   12.5532
#> 152  11689076  11689110    0.723 200072108  200046669    9.5931
#> 153  11689066  11689076    3.320 200095475  200046669    6.8427
#> 154  11689122  11689150    1.229 200046670  200046670   25.0758
#> 155  11689090  11689122    3.152 200060007  200046670   17.8560
#> 156  11689008  11689090    2.599 200095488  200046670    4.7520
#> 157  11687638  11687232    5.573 200095676  200095676    9.3249
#> 158  11687400  11687424    5.093 200052175  200052175   18.2511
#> 159  11687292  11687400    4.767 200060050  200052175   10.2411
#> 160  11687206  11687292    0.711 200095777  200052175    1.3374
#> 161  11691384  11691388    0.963 200052112  200019782   29.8800
#> 162  11691400  11691384    2.820 200071998  200071998    9.2556
#> 163  11691418  11691400    3.243 200095268  200071998    3.9951
#> 164  11688406  11688414    2.986 200060027  200060027    9.4068
#> 165  11687524  11690534    2.131 200095697  200060027    4.0419
#> 166  11690534  11688406    0.273 200072009  200060027    4.0959
#> 167  11688632  11690080    0.451 200039297  200036648   22.3893
#> 168  11688688  11688632    0.976 200042548  200036648   19.0782
#> 169  11688686  11688688    2.082 200046676  200036648   13.4163
#> 170  11688622  11688686    0.951 200052157  200036648   10.0683
#> 171  11691312  11688606    0.556 200072005  200036648    6.3045
#> 172  11688606  11688622    1.149 200060019  200036648    7.3197
#> 173  11691308  11691312    2.659 200095293  200036648    3.2121
#> 174  11688804  11690102    0.171 200072128  200059970    7.0191
#> 175  11688860  11688804    3.295 200095527  200059970    5.2344
#> 176  11689696  11689016    6.385 200034480  200034480   47.5911
#> 177 932060007  11689696    3.565 200036363  200034480   32.6772
#> 178 932060008 932060007    0.426 200038932  200034480   27.5256
#> 179  11688856 932060008    2.138 200042544  200034480   25.0254
#> 180 932060003  11688856    0.498 200046059  200034480   16.6572
#> 181  11691340 932060004    2.727 200059944  200034480   15.6978
#> 182 932060004 932060003    0.404 200051315  200034480   16.3251
#> 183  11691322  11691340    2.134 200072003  200034480    8.8281
#> 184  11691314  11691322    2.753 200095291  200034480    2.9736
#> 185  11689310  11689316    0.075 200095432  200095432    0.0000
#> 186  11694074  11690298    0.345 200042520  200039287   20.8053
#> 187  11690298  11690262    3.935 200039287  200039287   26.6031
#> 188  11690312  11690310    0.683 200052120  200039287    7.8939
#> 189  11690310  11694074    3.292 200046649  200039287   16.3782
#> 190  11689776  11690314    2.473 200095334  200039287    4.1247
#> 191  11690314  11690312    0.527 200072016  200039287    4.6134
#> 192  11689496  11694080    0.799 200095389  200059954    1.6119
#> 193  11690318  11690310    0.127 200059954  200059954    2.0160
#> 194  11694080  11690318    0.557 200071988  200059954    1.9818
#> 195  11689494  11694078    1.925 200095390  200059955    2.5173
#> 196  11690316  11690312    0.048 200059955  200059955    2.6073
#> 197  11694078  11690316    0.239 200071989  200059955    2.6037
#> 198  11690598  11690302    2.696 200095308  200072019    4.1148
#> 199  11690302  11694074    0.226 200072019  200072019    4.1580
#> 200  11689452  11690282    0.119 200072086  200059957    0.8307
#> 201  11690282  11690256    2.212 200059957  200059957    5.2623
#> 202  11690348  11689452    0.335 200095319  200059957    0.7974
#> 203  11690306  11690280    0.577 200072018  200059958   11.6451
#> 204  11690280  11690564    3.153 200059958  200059958   16.9821
#> 205  11694076  11690306    3.113 200095263  200059958   10.4391
#> 206  11694070  11694068    1.624 200052106  200042528   19.6200
#> 207  11694068  11690284    0.594 200046642  200042528   20.9718
#> 208  11690600  11694070    0.100 200059948  200042528   17.6337
#> 209  11690284  11690246    1.663 200042528  200042528   22.6269
#> 210  11690308  11690600    3.812 200072017  200042528   14.1336
#> 211  11689486  11690308    3.129 200095394  200042528    3.4011
#> 212  11690286  11690600    0.280 200072021  200072021    3.4569
#> 213  11690596  11690286    1.953 200095309  200072021    3.2868
#> 214  11690522  11690566    2.193 200072011  200072011    4.3695
#> 215  11690608  11690522    1.236 200095307  200072011    2.1591
#> 216  11690238  11689372    0.096 200046651  200039288    2.9223
#> 217  11689374  11690238    0.192 200052139  200039288    2.8656
#> 218  11689372  11690244    0.022 200042540  200039288    2.9223
#> 219  11690244  11690604    2.561 200039288  200039288    5.7501
#> 220  11690240  11689374    0.225 200059961  200039288    2.7135
#> 221  11689386  11690240    0.326 200072090  200039288    1.0152
#> 222  11690346  11689386    0.238 200095320  200039288    0.8550
#> 223  11689362  11690232    0.316 200072092  200052121    0.3915
#> 224  11690232  11690242    0.008 200059962  200052121    0.4050
#> 225  11690242  11690558    1.665 200052121  200052121    1.5057
#> 226  11690342  11689362    0.379 200095322  200052121    0.2331
#> 227  11690228  11690226    1.774 200059963  200059963    2.3760
#> 228  11689352  11690228    0.056 200072093  200059963    0.8442
#> 229  11690340  11689352    0.389 200095323  200059963    0.8388
#> 230  11690208  11690206    1.956 200024767  200024767  100.9080
#> 231  11694056  11690208    2.252 200025557  200024767   97.7625
#> 232  11690252  11690250    1.040 200027391  200024767   72.1287
#> 233  11690250  11694056    2.477 200026417  200024767   78.4836
#> 234  11690288  11690264    2.290 200031085  200029707   31.0563
#> 235  11690292  11690288    0.787 200032648  200029707   25.7409
#> 236  11690300  11690292    0.692 200034476  200029707   22.5027
#> 237  11690264  11690252    0.622 200029707  200029707   35.1756
#> 238  11689448  11690300    0.180 200036654  200029707   21.6540
#> 239  11690304  11689448    0.445 200039286  200029707   21.5955
#> 240  11689460  11690304    0.859 200042537  200029707   19.8405
#> 241  11689462  11689460    0.164 200072084  200072084    4.4280
#> 242  11689466  11689462    1.859 200095400  200072084    1.0224
#> 243  11689356  11690234    0.067 200059995  200042521    1.7910
#> 244  11689392  11690236    1.751 200095418  200042521    1.3284
#> 245  11694060  11690250    0.470 200042521  200042521    2.8107
#> 246  11690236  11689356    0.186 200072023  200042521    1.4094
#> 247  11690234  11690248    0.002 200052122  200042521    1.7910
#> 248  11690248  11694060    0.735 200046650  200042521    2.2104
#> 249  11689408  11694062    1.758 200095416  200059960    2.6694
#> 250  11694062  11690266    1.026 200071992  200059960    3.8430
#> 251  11690266  11690264    0.128 200059960  200059960    3.8655
#> 252  11689454  11694072    1.775 200095403  200059956    1.4607
#> 253  11690290  11690288    0.174 200059956  200059956    2.7729
#> 254  11694072  11690290    0.335 200071990  200059956    2.7063
#> 255  11689476  11690294    3.118 200095396  200072020    2.6586
#> 256  11690294  11690292    0.198 200072020  200072020    2.7216
#> 257  11689464  11689462    1.762 200095401  200095401    3.3912
#> 258  11689472  11689460    1.035 200046663  200029707   13.8402
#> 259  11689478  11689472    1.256 200052133  200029707    8.8326
#> 260  11689482  11689478    0.081 200059990  200029707    5.6538
#> 261  11689492  11689484    2.409 200095391  200029707    3.4290
#> 262  11689484  11689482    0.402 200072083  200029707    3.5262
#> 263  11689480  11689478    2.353 200095395  200095395    1.2051
#> 264  11689490  11689482    2.568 200095392  200095392    2.1195
#> 265  11689488  11689472    2.652 200095393  200095393    2.0727
#> 266  11690268  11690272    0.361 200032649  200024767   23.3325
#> 267  11690272  11694064    0.399 200031086  200024767   26.7966
#> 268  11694064  11690270    1.629 200029704  200024767   35.2251
#> 269  11690270  11690252    0.383 200028467  200024767   35.4114
#> 270  11689396  11690268    0.116 200034482  200024767   22.9419
#> 271  11690274  11689396    0.334 200036646  200024767   22.8483
#> 272  11689404  11690274    1.304 200039292  200024767   22.5081
#> 273  11689406  11689404    2.544 200095417  200095417    2.7630
#> 274  11689432  11694066    1.873 200095409  200059959    1.3734
#> 275  11694066  11690278    0.286 200071991  200059959    2.0511
#> 276  11690278  11694064    0.191 200059959  200059959    2.0772
#> 277  11689428  11690276    2.099 200095411  200072022    3.2310
#> 278  11690276  11690272    0.232 200072022  200072022    3.3147
#> 279  11689418  11689404    1.131 200042538  200024767   17.7318
#> 280  11689424  11689418    0.582 200046664  200024767   12.9618
#> 281  11689426  11689424    0.578 200052134  200024767   11.3715
#> 282  11689444  11689426    2.309 200059991  200024767    8.1342
#> 283  11689474  11689458    1.592 200095397  200024767    1.5318
#> 284  11689458  11689444    2.362 200072085  200024767    4.3335
#> 285  11689416  11689418    0.871 200072088  200072088    3.9582
#> 286  11689412  11689416    2.197 200095414  200072088    2.8638
#> 287  11689410  11689416    1.731 200095415  200095415    0.7371
#> 288  11689422  11689424    2.376 200095412  200095412    1.3185
#> 289  11689434  11689426    2.388 200095408  200095408    2.4453
#> 290  11689446  11689444    1.920 200095405  200095405    1.5291
#> 291  11694058  11694056    0.363 200039277  200039277   13.4883
#> 292  11690212  11694058    0.037 200042529  200039277   13.3623
#> 293  11690222  11690220    0.009 200072024  200039277   10.5615
#> 294  11689382  11690222    8.000 200095419  200039277   10.5588
#> 295  11690220  11689336    0.254 200059964  200039277   10.7595
#> 296  11690214  11690212    0.571 200046653  200039277   11.1411
#> 297  11689336  11690214    0.088 200052140  200039277   10.7964
#> 298  11690210  11690212    0.166 200059965  200059965    2.2194
#> 299  11689330  11690210    1.616 200072095  200059965    2.1870
#> 300  11689342  11689330    2.453 200095428  200059965    1.3527
#> 301  11689752  11689258    0.209 200011349  200011349  942.4251
#> 302  11690200  11689752    0.597 200011443  200011349  942.3846
#> 303  11689264  11690200    0.363 200011543  200011349  939.5550
#> 304  11689754  11689264    3.481 200095338  200011349    3.0303
#> 305  11689266  11690230    0.844 200052143  200042541   17.4618
#> 306  11689308  11689312    0.543 200042541  200042541   18.0000
#> 307  11690230  11689308    0.373 200046652  200042541   17.6967
#> 308  11689740  11689266    5.438 200095342  200095342    6.5673
#> 309  11689256  11689266    2.046 200060000  200042541    9.2106
#> 310  11689738  11689256    0.555 200072074  200042541    4.5585
#> 311  11689242  11689738    2.452 200095437  200042541    3.0348
#> 312  11689240  11689738    1.757 200095438  200095438    1.1772
#> 313  11689736  11689256    2.183 200095343  200095343    2.4723
#> 314  11689326  11689322    3.499 200095431  200095431    1.9647
#> 315  11689300  11689334    2.097 200095433  200095433    1.0602
#> 316  11689748  11689756    4.137 200095339  200072073    4.5000
#> 317  11689756  11689302    1.667 200072073  200072073    5.4612
#> 318  11689734  11689272    3.866 200095344  200095344    2.8440
#> 319  11689732  11689744    3.675 200095345  200095345    4.3731
#> 320  11689244  11689730    2.268 200095436  200095436    2.8863
#> 321  11689222  11689246    2.736 200095445  200095445    2.0637
#> 322  11689200  11689208    2.463 200095448  200095448    3.7683
#> 323  11689188  11689192    0.312 200072102  200072102    4.3353
#> 324  11689722  11689188    3.019 200095349  200072102    2.8872
#> 325  11689182  11689188    2.078 200095453  200095453    1.3887
#> 326  11689136  11689148    3.129 200095462  200095462    3.9006
#> 327  11689068  11689076    2.259 200095474  200095474    1.6380
#> 328  11689078  11689110    2.259 200095472  200095472    1.6623
#> 329  11689112  11689124    1.880 200095465  200095465    1.2159
#> 330  11689014  11689090    1.072 200072115  200072115    8.6472
#> 331  11688996  11689014    2.432 200095491  200072115    2.7072
#> 332  11688998  11689014    3.345 200095490  200095490    4.8744
#> 333  11690332  11689030    0.198 200095327  200052147    0.8415
#> 334  11689092  11689122    3.291 200052147  200052147    5.3802
#> 335  11690154  11689092    0.125 200059968  200052147    1.0782
#> 336  11689030  11690154    0.094 200072114  200052147    0.9477
#> 337  11689708  11689116    2.363 200095353  200095353    1.9017
#> 338  11688990  11689022    2.133 200095493  200095493    2.9313
#> 339  11688948  11688980    4.916 200095499  200095499    8.5023
#> 340  11688868  11690124    2.010 200095524  200072032    1.3968
#> 341  11688822  11690112    1.884 200095533  200072033    2.0286
#> 342  11688808  11690110    1.760 200072126  200046656    7.2774
#> 343  11688778  11688808    2.109 200095541  200046656    2.8053
#> 344  11688742  11688808    1.885 200095551  200095551    1.9674
#> 345  11688714  11690082    1.999 200095559  200072036    1.9062
#> 346  11688634  11688632    4.033 200095574  200095574    3.0483
#> 347  11688604  11688622    1.882 200095584  200095584    1.7685
#> 348  11691310  11691312    2.080 200095292  200095292    1.5219
#> 349  11688646  11688686    1.897 200095571  200095571    1.0926
#> 350  11688718  11688688    4.568 200095557  200095557    4.6269
#> 351  11688536  11688556    2.301 200095596  200095596    2.0646
#> 352  11688522  11688538    2.257 200095601  200095601    1.3221
#> 353  11688428  11688438    2.113 200095626  200095626    3.4641
#> 354  11690488  11688406    0.451 200072015  200072015    0.9342
#> 355  11687522  11690488    0.664 200095698  200072015    0.7281
#> 356  11691282  11691288    1.535 200095301  200095301    1.7154
#> 357  11687862  11687866    2.082 200095660  200095660    1.2510
#> 358  11687860  11687864    1.988 200095661  200095661    1.5057
#> 359  11687856  11687858    1.716 200095662  200095662    1.4040
#> 360  11687852  11687854    1.219 200095663  200095663    0.6642
#> 361  11687868  11687878    4.254 200095659  200021629    3.9528
#> 362  11691274  11691276    0.475 200039282  200039282    7.7940
#> 363  11687876  11691274    0.647 200042555  200039282    7.3359
#> 364  11687874  11693994    0.704 200060032  200039282    6.2064
#> 365  11693994  11691270    0.080 200052109  200039282    6.2235
#> 366  11691270  11687876    0.195 200046646  200039282    6.4431
#> 367  11687870  11687874    1.488 200095658  200095658    1.1178
#> 368  11691286  11687872    2.516 200095300  200039282    3.5901
#> 369  11687872  11687874    0.573 200072158  200039282    4.5225
#> 370  11691278  11691284    1.051 200052115  200052115   11.0934
#> 371  11691296  11691278    2.814 200059946  200052115    9.2718
#> 372  11691294  11691296    0.173 200072006  200052115    3.5820
#> 373  11691298  11691294    1.836 200095298  200052115    2.7432
#> 374  11691280  11691278    1.759 200095302  200095302    1.0368
#> 375  11691302  11691296    1.679 200095296  200095296    1.6389
#> 376  11691292  11691290    1.624 200095299  200095299    1.2015
#> 377  11691300  11688408    4.464 200095297  200095297    6.3072
#> 378  11688466  11688432    1.661 200095619  200095619    2.3409
#> 379  11689596  11689606    2.308 200095384  200095384    1.8117
#> 380  11688558  11690048    1.981 200095592  200072038    1.3176
#> 381  11688576  11690058    0.949 200046677  200042532   21.8133
#> 382  11688548  11688576    2.879 200095594  200095594    3.3480
#> 383  11688546  11688576    4.188 200052161  200042532   17.3646
#> 384  11688386  11688546    2.596 200060029  200060029    6.6105
#> 385  11688290  11688386    0.865 200072156  200060029    2.8512
#> 386  11687530  11688290    2.435 200095696  200060029    2.3958
#> 387  11688292  11688386    1.572 200095655  200095655    1.7424
#> 388  11688398  11688546    5.402 200095635  200042532    5.7411
#> 389  11688416  11689974    1.842 200095631  200072047    1.2555
#> 390  11687514  11687542    1.956 200095699  200095699    1.5741
#> 391  11687454  11687538    1.790 200060044  200031089   14.2830
#> 392  11687390  11687454    1.706 200095728  200095728    2.2725
#> 393  11687392  11687454    2.128 200072183  200031089    9.8946
#> 394  11687342  11687392    2.957 200095740  200031089    5.4774
#> 395  11687340  11687392    2.174 200095741  200095741    1.4778
#> 396  11688332  11688350    2.883 200095645  200095645    2.7531
#> 397  11687422  11687482    3.688 200095721  200095721    5.2245
#> 398  11687296  11687400    2.279 200095753  200095753    3.5388
#> 399  11687736  11687204    0.242 200095667  200072194    0.6138
#> 400  11687204  11687292    0.351 200072194  200072194    0.7236
#> 401  11687184  11687224    1.632 200095785  200095785    3.0816
#> 402  11687726  11687152    0.320 200072163  200060054    9.2916
#> 403  11687120  11687726    3.705 200095796  200060054    6.4395
#> 404  11687152  11687180    1.017 200060054  200060054    9.9126
#> 405  11687236  11687358    2.611 200095768  200095768    3.2148
#> 406  11687242  11687252    2.176 200095765  200095765    2.3823
#> 407  11687230  11687244    2.928 200095769  200095769    2.9223
#> 408  11687250  11687210    2.652 200095764  200095764    4.2363
#> 409  11687254  11687294    1.950 200095763  200095763    2.4012
#> 410  11687318  11687334    2.225 200095748  200095748    2.5263
#> 411  11687336  11687356    2.471 200095742  200095742    2.1429
#> 412  11687484  11687548    4.850 200095708  200095708    6.4665
#> 413  11688298  11689926    0.262 200046684  200042534   23.5881
#> 414  11690490  11688298    0.284 200059953  200059953    8.3295
#> 415  11688338  11690490    2.182 200072154  200059953    5.8275
#> 416  11688420  11688338    1.629 200095629  200095629    1.7982
#> 417  11688380  11688338    1.870 200095639  200059953    2.1303
#> 418  11688296  11690490    2.495 200095653  200095653    2.3859
#> 419  11690530  11688298    0.680 200052116  200042534   15.1857
#> 420  11687550  11690530    2.967 200060041  200042534   14.6808
#> 421  11687486  11687550    2.328 200095707  200095707    2.6046
#> 422  11687448  11687550    1.614 200072179  200042534    7.9065
#> 423  11687404  11687448    1.365 200095726  200095726    2.3904
#> 424  11687434  11687448    1.717 200095719  200042534    2.1888
#> 425  11688450  11689976    4.023 200095623  200072046    4.1274
#> 426  11688514  11690014    2.878 200072143  200059977    7.0443
#> 427  11688508  11688514    1.774 200095604  200059977    1.5651
#> 428  11688542  11688514    1.720 200095595  200095595    1.8963
#> 429  11688612  11690046    0.409 200072137  200059974    6.6168
#> 430  11688648  11688612    2.846 200095570  200059974    3.1680
#> 431  11688610  11688612    2.757 200095582  200095582    2.8836
#> 432  11688726  11690096    2.817 200095556  200072034    3.1554
#> 433  11688802  11688804    2.236 200095535  200095535    1.7541
#> 434  11688894  11689696    0.640 200072122  200072122    5.8131
#> 435  11688878  11688894    2.615 200095519  200072122    3.4407
#> 436  11688876  11688894    1.757 200095520  200095520    2.1132
#> 437  11691334 932060001    2.199 200095285  200058698    2.2203
#> 438 932060001 932060002    0.898 200069950  200058698    3.7890
#> 439 932060002  11688856    0.295 200058698  200058698    3.9114
#> 440  11691316  11691322    2.579 200095290  200095290    3.2697
#> 441  11691324  11691340    2.896 200095288  200095288    3.1743
#> 442 932060005 932060008    1.019 200069949  200069949    1.7019
#> 443 932060006 932060005    0.866 200091170  200069949    0.8469
#> 444  11688964  11688982    1.743 200095496  200095496    1.1007
#> 445  11688954  11688974    1.903 200095498  200095498    0.8892
#> 446  11688942  11688956    1.871 200095501  200095501    1.3248
#> 447  11688962  11688960    2.254 200095497  200095497    2.9223
#> 448  11691398  11688968    2.111 200052111  200046673   20.7450
#> 449  11688968  11688966    0.498 200046673  200046673   20.8989
#> 450  11691374  11691398    0.981 200059942  200046673   16.4520
#> 451  11691370  11691374    2.575 200072001  200046673    8.3889
#> 452  11691342  11691370    2.926 200095283  200046673    2.7549
#> 453  11691344  11691370    2.461 200095282  200095282    1.5480
#> 454  11691372  11691374    1.928 200072000  200072000    7.1370
#> 455  11691362  11691372    2.581 200095277  200072000    1.5732
#> 456  11691364  11691372    2.619 200095276  200095276    2.7378
#> 457  11691376  11691398    2.250 200095274  200095274    2.1429
#> 458  11691390  11691402    3.291 200095271  200095271    3.6837
#> 459  11691386  11691396    3.056 200095272  200095272    2.8296
#> 460  11691368  11691358    3.982 200095275  200095275    2.4327
#> 461  11691328  11691330    1.637 200095287  200095287    1.7316
#> 462  11691306  11691318    1.737 200095294  200095294    2.1330
#> 463  11691320  11691326    2.763 200095289  200095289    3.5910
#> 464  11691332  11691336    3.287 200095286  200095286    3.7782
#> 465  11691338  11691356    3.005 200095284  200095284    2.9772
#> 466  11691360  11691382    0.237 200052113  200052113   15.4278
#> 467  11691366  11691360    1.281 200059943  200052113   13.6683
#> 468  11691348  11691366    2.441 200095280  200095280    2.4606
#> 469  11691352  11691360    2.128 200095279  200095279    1.7244
#> 470  11691350  11691366    1.824 200072002  200052113   10.3149
#> 471  11691346  11691350    3.090 200095281  200052113    3.8160
#> 472  11691354  11691350    2.066 200095278  200095278    3.6684
#> 473  11691392  11691380    1.360 200071999  200019782   10.8558
#> 474  11691404  11691392    2.590 200095269  200019782    5.0508
#> 475  11691378  11691380    2.286 200095273  200095273    3.0897
#> 476  11691394  11691392    2.361 200095270  200095270    3.2967
#> 477  11691426  11691400    2.172 200095266  200095266    1.9467
#> 478  11691422  11691406    1.777 200095267  200095267    1.8567
#> 479  11691450  11691408    1.935 200095265  200095265    3.0357
#> 480  11689026  11691414    0.329 200046671  200039294    6.0174
#> 481  11689058  11691424    1.319 200060008  200039294    5.6025
#> 482  11691424  11689026    0.130 200052110  200039294    5.6529
#> 483  11689012  11688984    0.680 200039294  200039294    7.9119
#> 484  11691460  11693996    0.908 200095264  200039294    1.8783
#> 485  11693996  11689058    0.081 200071996  200039294    1.9080
#> 486  11691414  11689012    0.208 200042524  200039294    6.2127
#> 487  11689024  11688958    1.826 200095487  200095487    1.0449
#> 488  11689018  11688972    1.185 200060010  200060010    8.5986
#> 489  11689032  11689018    0.541 200072113  200060010    5.7816
#> 490  11689088  11689032    2.571 200095470  200095470    3.3246
#> 491  11689074  11689032    2.773 200095473  200060010    2.1969
#> 492  11689094  11689018    2.863 200095469  200095469    1.7712
#> 493  11689056  11689010    3.104 200095478  200095478    2.6388
#> 494  11689714  11689020    4.604 200095351  200095351    6.1380
#> 495  11689702  11689006    2.130 200072075  200072075    5.4171
#> 496  11689104  11689702    2.020 200095466  200072075    1.6290
#> 497  11689120  11689702    1.962 200095463  200095463    1.3149
#> 498  11689082  11689080    0.370 200072107  200072107    3.6558
#> 499  11689096  11689082    1.984 200095468  200072107    2.2383
#> 500  11689084  11689082    1.767 200095471  200095471    1.3014
#> 501  11689142  11689128    3.866 200095460  200095460    4.4127
#> 502  11689170  11689126    1.886 200095457  200095457    1.7037
#> 503  11689176  11689146    2.178 200095455  200095455    1.6344
#> 504  11689180  11689152    2.034 200095454  200095454    1.8918
#> 505  11689726  11689214    5.008 200095347  200095347    6.9759
#> 506  11689218  11689224    0.257 200060002  200060002    9.5265
#> 507  11689238  11689218    2.052 200095439  200095439    3.9780
#> 508  11689220  11689218    0.263 200072101  200060002    5.4540
#> 509  11689216  11689220    4.133 200095446  200060002    4.0032
#> 510  11689234  11689220    2.283 200095441  200095441    1.4049
#> 511  11689728  11689254    3.574 200095346  200095346    4.3776
#> 512  11689746  11689750    2.674 200095340  200095340    2.1150
#> 513  11689762  11689274    3.015 200072072  200072072    9.7614
#> 514  11689760  11689762    4.251 200095337  200072072    5.7744
#> 515  11689764  11689762    1.773 200095336  200095336    1.0485
#> 516  11689378  11689332    3.948 200039293  200039293   51.0039
#> 517  11689390  11689378    0.815 200042539  200039293   33.1470
#> 518  11689394  11689390    1.157 200046665  200039293   21.9528
#> 519  11689398  11689394    1.191 200052137  200039293   19.8234
#> 520  11689400  11689398    0.069 200059993  200039293   18.5481
#> 521  11689772  11689400    6.344 200072070  200039293   11.6811
#> 522  11689360  11689772    4.502 200095423  200039293    5.1309
#> 523  11689380  11689378    0.576 200046666  200046666   13.0770
#> 524  11689384  11689380    1.662 200052138  200046666   10.0152
#> 525  11689366  11689384    2.208 200095421  200095421    1.2303
#> 526  11689376  11689380    1.622 200072091  200072091    2.6154
#> 527  11689348  11689376    1.014 200095426  200072091    1.2321
#> 528  11689368  11689384    0.409 200059994  200046666    7.6797
#> 529  11689770  11689368    0.761 200072071  200046666    5.1318
#> 530  11689370  11689770    2.162 200095420  200095420    2.3445
#> 531  11689766  11689368    3.200 200095335  200095335    2.4264
#> 532  11689364  11689770    2.779 200095422  200046666    2.2617
#> 533  11689358  11689772    2.126 200095424  200095424    0.9909
#> 534  11689774  11689400    1.804 200072069  200072069    6.8580
#> 535  11689430  11689774    1.811 200095410  200072069    2.0178
#> 536  11689438  11689774    1.772 200095406  200095406    2.0619
#> 537  11689456  11689390    5.326 200095402  200095402   10.5966
#> 538  11689768  11689328    2.236 200046662  200046662   25.0758
#> 539  11689420  11689768    3.090 200052135  200046662   17.4942
#> 540  11689440  11689420    1.531 200059992  200046662    9.6228
#> 541  11689442  11689440    0.054 200072087  200046662    5.0877
#> 542  11689468  11689442    3.160 200095399  200046662    3.8763
#> 543  11689450  11689442    1.866 200095404  200095404    1.2087
#> 544  11689470  11689440    2.517 200095398  200095398    2.9574
#> 545  11689436  11689420    1.698 200095407  200095407    3.4965
#> 546  11689388  11689768    1.433 200072089  200072089    5.6772
#> 547  11689414  11689388    2.101 200095413  200072089    2.7585
#> 548  11689318  11689316    0.187 200072096  200072096    0.4941
#> 549  11690344  11689318    0.488 200095321  200072096    0.4320
#> 550  11689346  11689314    1.979 200095427  200095427    1.4013
#> 551  11689268  11689280    0.464 200011971  200011971  921.4632
#> 552  11689282  11689268    0.343 200012079  200011971  920.8026
#> 553  11689298  11689282    0.275 200052141  200011971    4.2912
#> 554  11689354  11689344    1.944 200095425  200011971    2.9142
#> 555  11689344  11689320    0.791 200072094  200011971    3.5172
#> 556  11689320  11689298    0.455 200059997  200011971    3.6846
#> 557  11689742  11689280    3.395 200095341  200095341    4.7367
#> 558  11689292  11689288    0.381 200052142  200052142    3.9888
#> 559  11689338  11689292    2.220 200095430  200052142    1.3815
#> 560  11689304  11689292    0.233 200059998  200059998    2.3985
#> 561  11689340  11689304    1.542 200095429  200059998    2.3067
#> 562  11689306  11689304    0.042 200072097  200072097    0.0567
#> 563  11690338  11689306    0.119 200095324  200072097    0.0423
#> 564  11689290  11689286    0.054 200072098  200072098    1.3545
#> 565  11690336  11689290    0.304 200095325  200072098    0.4149
#> 566  11689102  11689720    1.377 200072105  200072105   10.5174
#> 567  11689060  11689102    3.601 200095477  200072105    4.1283
#> 568  11689062  11689102    1.942 200095476  200095476    5.4081
#> 569  11688594  11688658    2.212 200095586  200095586    1.7028
#> 570  11688660  11689644    4.659 200095568  200095568    6.1326
#> 571  11688818  11688810    2.077 200052150  200052150   19.2348
#> 572  11688816  11688818    2.436 200060012  200052150   14.9931
#> 573  11688764  11688816    0.290 200072129  200052150    9.0810
#> 574  11688748  11688764    4.506 200095549  200052150    5.9265
#> 575  11688750  11688764    2.961 200095548  200095548    3.0591
#> 576  11689668  11688816    2.423 200095364  200095364    2.6163
#> 577  11688862  11688818    2.044 200095526  200095526    1.9287
#> 578  11688886  11689684    4.144 200095517  200095517    5.2623
#> 579  11689692  11689700    8.241 200095357  200095357   10.4454
#> 580  11689704  11689106    2.047 200095355  200095355    1.5534
#> 581  11689118  11689108    1.999 200095464  200095464    1.1637
#> 582  11689186  11689202    4.780 200052144  200052144   20.5866
#> 583  11689156  11689186    0.520 200060004  200052144   11.9682
#> 584  11689144  11689156    1.225 200072103  200052144    5.3352
#> 585  11689712  11689144    2.562 200095352  200052144    2.0358
#> 586  11689716  11689144    1.802 200095350  200095350    2.6577
#> 587  11689198  11689156    4.226 200095449  200095449    6.3603
#> 588  11689194  11689186    2.385 200095450  200095450    1.5462
#> 589  11689206  11689204    4.458 200095447  200095447   12.4110
#> 590  11690182  11694046    4.995 200021628  200021126  169.3773
#> 591  11689158  11690540    1.304 200022741  200021126  142.2333
#> 592  11690540  11690182    2.735 200022160  200021126  150.4278
#> 593  11694046  11690190    0.830 200021126  200021126  170.1036
#> 594  11689132  11689158    3.325 200072104  200072104    8.7903
#> 595  11689040  11689132    2.043 200095484  200095484    2.5038
#> 596  11689042  11689132    2.936 200095483  200072104    2.6622
#> 597  11689134  11689158    3.517 200023371  200021126  127.4085
#> 598  11689050  11689134    3.107 200024050  200021126  120.2175
#> 599  11688952  11689050    1.782 200024773  200021126   99.0990
#> 600  11688910  11690090    0.073 200027396  200021126   93.8538
#> 601  11688864  11690062    2.412 200072124  200059973   15.4422
#> 602  11688754  11688864    4.047 200095546  200059973    8.0739
#> 603  11688890  11690070    2.298 200034485  200021126   68.1849
#> 604  11688854  11688890    0.345 200036655  200021126   59.1282
#> 605  11688842  11688854    2.344 200039296  200021126   53.9091
#> 606  11688760  11688842    1.305 200042547  200021126   44.6868
#> 607  11688768  11688760    2.114 200046675  200021126   35.6499
#> 608  11688780  11688768    0.923 200052151  200021126   13.5027
#> 609  11688794  11688780    0.500 200060014  200021126   11.1366
#> 610  11688858  11688794    1.893 200095528  200095528    2.4858
#> 611  11688752  11688842    4.369 200095547  200095547    5.7753
#> 612  11688758  11688760    0.450 200072130  200072130    7.2027
#> 613  11688734  11688758    3.556 200095553  200072130    5.9652
#> 614  11688736  11688758    1.833 200095552  200095552    1.0836
#> 615  11688770  11688768    4.307 200052152  200052152   18.3897
#> 616  11688642  11688770    0.561 200060017  200052152    8.8893
#> 617  11688640  11688642    1.103 200072136  200052152    4.2336
#> 618  11689616  11688640    2.370 200095379  200052152    2.2122
#> 619  11688636  11688770    2.745 200095573  200095573    3.1644
#> 620  11689624  11688640    1.692 200095375  200095375    1.0917
#> 621  11689628  11688642    2.287 200095374  200095374    4.2111
#> 622  11688782  11688780    1.968 200095540  200095540    1.3734
#> 623  11688806  11688794    0.884 200072127  200021126    7.6077
#> 624  11689666  11688806    2.946 200095365  200021126    4.3110
#> 625  11688800  11688806    2.158 200095536  200095536    2.4435
#> 626  11688852  11688854    3.832 200095529  200095529    5.1057
#> 627  11688908  11688890    4.415 200095512  200095512    5.6385
#> 628  11688756  11688864    3.651 200095545  200095545    3.7566
#> 629  11688872  11690064    2.208 200095522  200072037    1.9107
#> 630  11688888  11690072    1.065 200072123  200059972    7.4367
#> 631  11688870  11688888    1.845 200095523  200059972    4.0896
#> 632  11688904  11688888    2.030 200095513  200095513    2.5380
#> 633  11688934  11690094    1.532 200095505  200072035    1.8765
#> 634  11688970  11689050    4.740 200095495  200095495   14.4369
#> 635  11689052  11689134    1.871 200095480  200095480    2.1159
#> 636  11689236  11690188    0.856 200095440  200072026    2.6901
#> 637  11690188  11690194    2.206 200072026  200072026    7.0047
#> 638  11690192  11694050    1.572 200072025  200059938    9.7398
#> 639  11689230  11690192    1.109 200095443  200059938    7.5654
#> 640  11694050  11690198    1.931 200059938  200059938   13.7331
#> 641  11689260  11689278    1.213 200095434  200052117    3.4542
#> 642  11689262  11689278    0.049 200072099  200072099    0.5832
#> 643  11690334  11689262    0.624 200095326  200072099    0.5823
#> 644  11690614  11690616    5.885 200095305  200039278   21.4470
#> 645  11690616  11690204    0.182 200046647  200039278   27.7110
#> 646  11690204  11694054    1.582 200042530  200039278   31.1265
#> 647  11694054  11690560    1.545 200039278  200039278   33.6771
#> 648  11690618  11690606    3.044 200095304  200059950    5.4936
#> 649  11690620  11690606    2.622 200095303  200095303    3.8115
#> 650  11691380  11691384    4.631 200059941  200019782   19.5849
#> 651  11689278  11690526    1.482 200059999  200052117    5.6340
#> 652  11690526  11690616    0.568 200052117  200052117    6.2109
#> 653  11690536  11690570    2.546 200059950  200059950   16.1694
#> 654  11690606  11690536    1.028 200072008  200059950   10.6722
#> 655  11689926  11689978    0.213 200042534  200042534   23.6457
#> 656  11689928  11689978    0.517 200027393  200009667  147.4920
#> 657  11689972  11689988    1.363 200032651  200031089   39.5505
#> 658  11689974  11689988    0.313 200072047  200072047    1.6902
#> 659  11689976  11689986    0.326 200072046  200072046    4.3722
#> 660  11689978  11689986    3.277 200026421  200009667  175.6818
#> 661  11689986  11690018    0.544 200025560  200009667  180.1863
#> 662  11689988  11690018    0.690 200031089  200031089   41.8293
#> 663  11690014  11690044    1.239 200059977  200059977    7.7148
#> 664  11690018  11690044    3.736 200024770  200009667  224.4582
#> 665  11690044  11690056    2.594 200024049  200009667  234.5976
#> 666  11690046  11690056    1.561 200059974  200059974    8.0892
#> 667  11690048  11690060    0.123 200072038  200072038    1.8531
#> 668  11690050  11690060    0.247 200023370  200021629   76.0212
#> 669  11690056  11690084    1.415 200023369  200009667  243.6777
#> 670  11690058  11690084    1.690 200042532  200042532   23.1390
#> 671  11690060  11690068    1.404 200022740  200021629   79.8192
#> 672  11690068  11690086    0.352 200022161  200021629  103.4271
#> 673  11690080  11690068    1.075 200036648  200036648   23.4774
#> 674  11690082  11690086    0.494 200072036  200072036    2.6739
#> 675  11690084  11690092    1.627 200022739  200009667  267.9039
#> 676  11690086  11690092    5.327 200021629  200021629  115.0155
#> 677  11690092  11690104    1.268 200021127  200009667  385.9866
#> 678  11690096  11690104    1.278 200072034  200072034    4.2129
#> 679  11690102  11690126    2.033 200059970  200059970    9.0054
#> 680  11690104  11690126    1.741 200020652  200009667  392.9049
#> 681  11690110  11690122    0.355 200059969  200046656    7.4619
#> 682  11690112  11690122    0.106 200072033  200072033    2.0628
#> 683  11690122  11690128    1.272 200052125  200046656   11.0043
#> 684  11690124  11690128    0.609 200072032  200072032    2.5092
#> 685  11690126  11690130    3.020 200020204  200009667  407.0826
#> 686  11690128  11690130    1.475 200046656  200046656   15.5106
#> 687  11690130  11688950    0.813 200019781  200009667  423.9099
#> 688  11690090  11690098    0.411 200026419  200021126   94.5477
#> 689  11690094  11690098    0.089 200072035  200072035    1.8873
#> 690  11690098  11688952    0.311 200025559  200021126   96.5502
#> 691  11690062  11690066    0.425 200059973  200059973   15.6069
#> 692  11690064  11690074    0.202 200072037  200072037    1.9584
#> 693  11690066  11690074    0.135 200031087  200021126   83.8845
#> 694  11690070  11690066    0.216 200032650  200021126   68.2578
#> 695  11690072  11690088    0.115 200059972  200059972    7.4457
#> 696  11690074  11690088    0.090 200029708  200021126   85.8609
#> 697  11690088  11688910    0.871 200028469  200021126   93.8475