Prep NHDPlus Data

Function to prep NHDPlus data for use by nhdplusTools functions

prepare_nhdplus(flines, min_network_size, min_path_length,
  min_path_size = 0, purge_non_dendritic = TRUE, warn = TRUE,
  error = TRUE)

Arguments

flines
data.frame NHDPlus flowlines including: COMID, LENGTHKM, FTYPE, TerminalFl, FromNode, ToNode, TotDASqKM, StartFlag, StreamOrde, StreamCalc, TerminalPa, Pathlength, and Divergence variables.

min_network_size
numeric Minimum size (sqkm) of drainage network to include in output.

min_path_length
numeric Minimum length (km) of terminal level path of a network.

min_path_size
numeric Minimum size (sqkm) of outlet level path of a drainage basin. Drainage basins with an outlet drainage area smaller than this will be removed.

purge_non_dendritic
boolean Should non dendritic paths be removed or not.

warn
boolean controls whether warning an status messages are printed

error
boolean controls whether to return potentially invalid data with a warning rather than an error

Value

data.frame ready to be used with the refactor_flowlines function.

Examples

flines_in <- sf::read_sf(system.file("extdata/petapsco_flowlines.gpkg",
                                     package = "nhdplusTools"))
prepare_nhdplus(flines_in,
                min_network_size = 10,
                min_path_length = 1,
                warn = FALSE)

#> # A tibble: 697 x 6
#>       COMID LENGTHKM TotDASqKM  Hydroseq LevelPathI  toCOMID
#>       <int>    <dbl>     <dbl>     <dbl>      <dbl>    <int>
#>  1 11689204    0.73      150.  200027394  200024046 11689212
#>  2 11689212    0.819     165.  200026423  200024046 11689228
#>  3 11689228    2.05      173.  200025563  200024046 11690186
#>  4 11690186    2.91      182.  200024768  200024046 11694048
#>  5 11694048    0.745     183.  200024046  200024046 11690184
#>  6 11689202    4.76      137.  200028471  200024046 11689204
#>  7 11689720    4.40      109.  200029710  200024046 11689202
#>  8 11689108    1.26       90.4 200031091  200024046 11689720
#>  9 11689106    3.51       88.4 200032655  200024046 11689108
#> 10 11689700    2.35       82.8 200034479  200024046 11689106
#> # … with 687 more rows