Subset by Raster Processing Unit.

Given flowlines and an rpu_code, performs a network-safe subset such that the result can be used in downstream processing. Has been tested to work against the entire NHDPlusV2 domain and satisfies a number of edge cases.

subset_rpu(fline, rpu, run_make_standalone = TRUE)

Arguments

fline
sf data.frame NHD Flowlines with COMID, Pathlength, LENGTHKM, and Hydroseq. LevelPathI, RPUID, ToNode, FromNode, and ArbolateSu.

rpu
character e.g. "01a"

run_make_standalone
boolean should the run_make_standalone function be run on result?

Examples

source(system.file("extdata/sample_data.R", package = "nhdplusTools"))

nhdplus_path(sample_data)

staged_nhdplus <- stage_national_data(output_path = tempdir())

sample_flines <- readRDS(staged_nhdplus$flowline)

subset_rpu(sample_flines, rpu = "07b")
#> Simple feature collection with 267 features and 138 fields
#> Geometry type: MULTILINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -89.58537 ymin: 42.95163 xmax: -89.19935 ymax: 43.30179
#> Geodetic CRS:  GRS 1980(IUGG, 1980)
#> # A tibble: 267 x 139
#>      COMID FDATE               RESOLUTION GNIS_ID GNIS_NAME  LENGTHKM REACHCODE 
#>  *   <int> <dttm>              <chr>      <chr>   <chr>         <dbl> <chr>     
#>  1  1.33e7 1999-10-15 00:00:00 Medium     "15770~ "Yahara R~    0.091 070900020~
#>  2  1.33e7 2009-05-01 00:00:00 Medium     "15770~ "Yahara R~    2.85  070900020~
#>  3  1.33e7 2009-05-01 00:00:00 Medium     "15770~ "Yahara R~    0.041 070900020~
#>  4  1.33e7 1999-10-29 00:00:00 Medium     "15770~ "Yahara R~    0.096 070900020~
#>  5  1.33e7 1999-10-29 00:00:00 Medium     "15770~ "Yahara R~    1.72  070900020~
#>  6  1.33e7 1999-10-15 00:00:00 Medium     " "     " "           4.59  070900020~
#>  7  1.33e7 1999-10-15 00:00:00 Medium     "15606~ "Murphys ~    4.53  070900020~
#>  8  1.33e7 1999-10-15 00:00:00 Medium     "15606~ "Murphys ~    2.70  070900020~
#>  9  1.33e7 1999-10-29 00:00:00 Medium     "15751~ "Swan Cre~    0.111 070900020~
#> 10  1.33e7 1999-10-15 00:00:00 Medium     "15751~ "Swan Cre~    1.44  070900020~
#> # ... with 257 more rows, and 132 more variables: FLOWDIR <chr>,
#> #   WBAREACOMI <int>, FTYPE <chr>, FCODE <int>, Shape_Length <dbl>,
#> #   StreamLeve <int>, StreamOrde <int>, StreamCalc <int>, FromNode <dbl>,
#> #   ToNode <dbl>, Hydroseq <dbl>, LevelPathI <dbl>, Pathlength <dbl>,
#> #   TerminalPa <dbl>, ArbolateSu <dbl>, Divergence <int>, StartFlag <int>,
#> #   TerminalFl <dbl>, DnLevel <dbl>, UpLevelPat <dbl>, UpHydroseq <dbl>,
#> #   DnLevelPat <dbl>, DnMinorHyd <dbl>, DnDrainCou <int>, DnHydroseq <dbl>,
#> #   FromMeas <dbl>, ToMeas <dbl>, RtnDiv <int>, VPUIn <int>, VPUOut <int>,
#> #   AreaSqKM <dbl>, TotDASqKM <dbl>, DivDASqKM <dbl>, Tidal <int>, TOTMA <dbl>,
#> #   WBAreaType <chr>, HWNodeSqKM <dbl>, MAXELEVRAW <dbl>, MINELEVRAW <dbl>,
#> #   MAXELEVSMO <dbl>, MINELEVSMO <dbl>, SLOPE <dbl>, ELEVFIXED <chr>,
#> #   HWTYPE <chr>, SLOPELENKM <dbl>, QA_MA <dbl>, VA_MA <dbl>, QC_MA <dbl>,
#> #   VC_MA <dbl>, QE_MA <dbl>, VE_MA <dbl>, QA_01 <dbl>, VA_01 <dbl>,
#> #   QC_01 <dbl>, VC_01 <dbl>, QE_01 <dbl>, VE_01 <dbl>, QA_02 <dbl>,
#> #   VA_02 <dbl>, QC_02 <dbl>, VC_02 <dbl>, QE_02 <dbl>, VE_02 <dbl>,
#> #   QA_03 <dbl>, VA_03 <dbl>, QC_03 <dbl>, VC_03 <dbl>, QE_03 <dbl>,
#> #   VE_03 <dbl>, QA_04 <dbl>, VA_04 <dbl>, QC_04 <dbl>, VC_04 <dbl>,
#> #   QE_04 <dbl>, VE_04 <dbl>, QA_05 <dbl>, VA_05 <dbl>, QC_05 <dbl>,
#> #   VC_05 <dbl>, QE_05 <dbl>, VE_05 <dbl>, QA_06 <dbl>, VA_06 <dbl>,
#> #   QC_06 <dbl>, VC_06 <dbl>, QE_06 <dbl>, VE_06 <dbl>, QA_07 <dbl>,
#> #   VA_07 <dbl>, QC_07 <dbl>, VC_07 <dbl>, QE_07 <dbl>, VE_07 <dbl>,
#> #   QA_08 <dbl>, VA_08 <dbl>, QC_08 <dbl>, VC_08 <dbl>, QE_08 <dbl>,
#> #   VE_08 <dbl>, QA_09 <dbl>, ...